% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{as_draws.ernest_sampler}
\alias{as_draws.ernest_sampler}
\alias{as_draws_list.ernest_sampler}
\alias{as_draws_matrix.ernest_sampler}
\title{Transform \code{ernest_sampler} to \code{posterior::draws} objects}
\usage{
\method{as_draws}{ernest_sampler}(x, scale = c("original", "unit"), ...)

\method{as_draws_list}{ernest_sampler}(x, scale = c("original", "unit"), ...)

\method{as_draws_matrix}{ernest_sampler}(x, scale = c("original", "unit"), ...)
}
\arguments{
\item{x}{An \code{ernest_sampler} object, or another \R object for which
the methods are defined.}

\item{scale}{One of \code{original} or \code{unit}. If \code{unit}, the returned
\code{draws} object will contain points expressed in the unit hypercube, which
was originally used by ernest for generated sampling. If \code{original} (the
default), the returned \code{draws} object will contain points in the original
parameter space defined by the prior transformation.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\description{
Transform an \code{ernest_sampler} object to a format supported by the
posterior package.
}
