% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws.R
\name{as_draws.ErnestSampler}
\alias{as_draws.ErnestSampler}
\title{Transform an Ernest Run into a \code{draws} object}
\usage{
\method{as_draws}{ErnestSampler}(x, resample = FALSE, add_live = TRUE, unit_scale = FALSE, ...)
}
\arguments{
\item{x}{An \code{ErnestSampler} object.}

\item{resample}{If true, the returned \code{draws_matrix} will be reweighted
before being returned. If false (default), the matrix will instead be bound
to the appropriate importance weights.}

\item{add_live}{If true, the returned \code{draws_matrix} will include the
live points.}

\item{unit_scale}{If true, the returned \code{draws_matrix} will contain points
expressed in the unit hypercube, which was originally used for sampling.}

\item{...}{Passed to \code{posterior::resample_draws} if \code{reweight} is true.}
}
\value{
A \code{draws_matrix} object, which is a 2D array containing
\code{x$n_dim} "variables" and \code{x$iter} (plus \code{x$n_point} if \code{include_live} is
true) "draws".
}
\description{
Transform an \code{ErnestSampler} object into format supported by the
posterior package. This allows for the easy analysis of the posterior
distribution generated through nested sampling.
}
\seealso{
\code{\link[posterior:draws_matrix]{posterior::as_draws_matrix()}}, \code{\link[posterior:resample_draws]{posterior::resample_draws()}},
and \code{\link[posterior:weight_draws]{posterior::weight_draws()}} for more information on the returned object.
}
