% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws.R
\name{as_draws.ernest_run}
\alias{as_draws.ernest_run}
\alias{as_draws_matrix.ernest_run}
\alias{as_draws_rvars.ernest_run}
\title{Transform nested sampling runs to \code{draws} objects}
\usage{
\method{as_draws}{ernest_run}(x, ..., units = c("original", "unit_cube"), radial = FALSE)

\method{as_draws_matrix}{ernest_run}(x, ..., units = c("original", "unit_cube"), radial = FALSE)

\method{as_draws_rvars}{ernest_run}(x, ..., units = c("original", "unit_cube"), radial = FALSE)
}
\arguments{
\item{x}{An \link{ernest_run} object.}

\item{...}{These dots are for future extensions and must be empty.}

\item{units}{Case-sensitive string. The scale for the sampled points:
\itemize{
\item \code{"original"}: Points are on the scale of the prior space.
\item \code{"unit_cube"}: Points are on the (0, 1) unit hypercube scale.
}}

\item{radial}{Logical. If \code{TRUE}, returns an additional column \code{.radial}
containing the radial coordinate (i.e., the Euclidean norm) for each
sampled point.}
}
\value{
A \link[posterior:draws]{draws} object containing posterior samples
from the nested sampling run, with importance weights (in log units).

The returned object type depends on the function used:
\itemize{
\item For \code{as_draws} and \code{as_draws_matrix}, a \code{\link[posterior:draws_matrix]{posterior::draws_matrix()}}
object (class \code{c("draws_matrix", "draws", "matrix")}).
\item For \code{as_draws_rvars}, a \code{\link[posterior:draws_rvars]{posterior::draws_rvars()}} object (class
\code{c("draws_rvars", "draws", "list")}).
}
}
\description{
Convert an \link{ernest_run} to a format supported by the
\link[posterior:posterior-package]{posterior} package.
}
\examples{
# Load example run
library(posterior)
data(example_run)

# View importance weights
dm <- as_draws(example_run)
weights(dm) |> head()

# Summarise points after resampling
dm |>
  resample_draws() |>
  summarize_draws()

# View the radial coordinate in unit space over the run
dm_rad <- as_draws_rvars(
  example_run,
  units = "unit_cube",
  radial = TRUE
)
plot(
  x = example_run$log_volume,
  y = draws_of(dm_rad$.radial),
  xlab = "Log-volume",
  ylab = "Radial coordinate"
)
}
\seealso{
\itemize{
\item \code{\link[posterior:draws]{posterior::as_draws()}} for details on the \code{draws} object.
\item \code{\link[posterior:resample_draws]{posterior::resample_draws()}} uses the log weights from ernest's output
to produce a weighted posterior sample.
}
}
