% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws.R
\name{as_draws_rvars.ernest_run}
\alias{as_draws_rvars.ernest_run}
\title{Transform to draws objects}
\usage{
\method{as_draws_rvars}{ernest_run}(x, resample_draws = TRUE, ...)
}
\arguments{
\item{x}{The earnest run object}

\item{resample_draws}{Whether to run \code{posterior::resample_draws} on the
draws object, producing a sample from a resampled posterior.}

\item{...}{Additional arguments to pass to \code{posterior::resample_draws}}
}
\value{
An object of class \code{"draws_matrix"}.
}
\description{
Transform to draws objects
}
\details{
When \code{resample_draws} is false, the draws object will be bound to
the log importance weights from the nested sampling run.
}
