% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.ernest_run}
\alias{autoplot.ernest_run}
\title{Autoplot for Ernest Run Objects}
\usage{
\method{autoplot}{ernest_run}(object, ...)
}
\arguments{
\item{object}{An \code{ernest_run} object containing a nested sampling run.}

\item{...}{Additional arguments passed to the method.}
}
\value{
A ggplot object.
}
\description{
Autoplot for Ernest Run Objects
}
