% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate.ErnestSampler}
\alias{calculate.ErnestSampler}
\title{Calculate statistics from an \code{ErnestSampler} object.}
\usage{
\method{calculate}{ErnestSampler}(x, exponentiate = FALSE, progress = TRUE, ...)
}
\arguments{
\item{x}{An \code{ErnestSampler} object.}

\item{exponentiate}{Whether to report likelihoods and evidence estimates
in exponentiated units.}

\item{progress}{Whether to return a tibble describing
each update to the evidence integral estimate at each iteration (\code{TRUE},
the default), or to return a one-row \code{tibble} with the final estimates.}

\item{...}{Ignored.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{log_vol}: The estimated log volume of the prior.
\item \code{lik}: The likelihood of the live points at that iteration.
\item \code{evidence}: The evidence estimate.
\item \code{log_evidence_err}: The error in the evidence estimate.
}

Additionally, if \code{progress} is \code{TRUE}, add these additional columns:
\itemize{
\item \code{lik_norm}: The normalized likelihood at each iteration.
\item \code{weight}: The importance weight attributed to each discarded point
at each iteration.
}

If \code{exponentiate} is \code{FALSE}, the columns \code{lik}, \code{lik_norm}, \code{weight}, and
\code{evidence} will be in log units and labelled appropriately.
The columns \code{log_vol} and \code{log_evidence_err} are always returned in the
log scale.
}
\description{
Calculate statistics from an \code{ErnestSampler} object.
}
