% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_live}
\alias{check_live}
\title{Internal method for validating an existing nested sample.}
\usage{
check_live(live, n_points, n_var, call = caller_env())
}
\description{
Checks the following list of fatal conditions for the nested sampler:
\itemize{
\item The live points are a list.
\item The live points contain the following elements with corresponding types:
\item \code{units}: A matrix of values between \link{0, 1}.
\item \code{points}: A matrix of finite values.
\item \code{log_lik}: A vector of finite or -Inf values.
\item The number of live points matches the expected number of live points.
\item The number of dimensions in the live points matches the expected number (n_points)
of dimensions (lrps)
\item \code{units} and \code{points} are of the exact same dimensions.
\item \code{points} has the same number of rows as \code{log_lik} has elements.
}
}
