% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_prior-class.R
\name{compile.ernest_prior}
\alias{compile.ernest_prior}
\title{Compile an ernest_prior into an inverse transformation function for NS}
\usage{
\method{compile}{ernest_prior}(object, ...)
}
\arguments{
\item{...}{Ignored}

\item{x}{An \code{ernest_prior} object.}
}
\value{
A function that takes a vector of uniform random variables and returns
a vector of the same length with the corresponding values from the prior
distribution.
}
\description{
Compile an ernest_prior into an inverse transformation function for NS
}
