% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile.ErnestSampler}
\alias{compile.ErnestSampler}
\title{Configure an ErnestSampler object}
\usage{
\method{compile}{ErnestSampler}(object, ...)
}
\arguments{
\item{object}{An ErnestSampler or ErnestSampler object.}

\item{...}{Ignored.}
}
\value{
\code{object}, now attached to an \code{ErnestWorkspace}
}
\description{
Ernest conducts nested sampling in an enviroment bound to the nested sampler.
This function configures this environment so that it contains the necessary
variables.
}
