% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_integral}
\alias{compute_integral}
\title{Compute the nested sampling integral}
\usage{
compute_integral(log_lik, log_vol)
}
\arguments{
\item{log_lik}{A vector of log-likelihoods in descending order.}

\item{log_vol}{A vector of log-volumes in ascending order.}
}
\value{
A list with the following components:
log_z: Log. evidence
log_z_var: Variance of the log. evidence
h: Information
dh: Differential information
}
\description{
Compute the nested sampling integral
}
