% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{confint.ernest_run}
\alias{confint.ernest_run}
\title{Confidence intervals for model evidence}
\usage{
\method{confint}{ernest_run}(
  object,
  parm = c("log_evidence", "evidence"),
  level = 0.95,
  times = 100,
  ...
)
}
\arguments{
\item{object}{[\link{ernest_run}]\cr An object containing the results of a nested
sampling run.}

\item{parm}{\verb{[character(1)]}\cr A specification describing whether to return
the model evidence in log-units.}

\item{level}{\verb{[double(1)]}\cr The confidence level required.}

\item{times}{\verb{[[integer(1)]]}\cr The number of bootstrap resamples.}
}
\value{
A named vector with two elements, describing the lower and upper
confidence limits for each parameter.
}
\description{
Confidence intervals for model evidence
}
\seealso{
\code{\link[=generate_bootstraps]{generate_bootstraps()}} for generating resampled nested sampling
runs from an \code{ernest_run} object.
}
