% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_priors.R
\name{create_cauchy_prior}
\alias{create_cauchy_prior}
\title{Specify a prior with Cauchy-distributed marginals}
\usage{
create_cauchy_prior(
  n_dim,
  location = 0,
  scale = 1,
  varnames = NULL,
  lower = -Inf,
  upper = Inf
)
}
\arguments{
\item{n_dim}{A whole number, representing the number of dimensions within the
prior distribution. Must be larger than or equal to 1.}

\item{location, scale}{location and scale parameters.}

\item{varnames}{A character vector of names for the variables in the
prior distribution.}

\item{lower, upper}{Numeric vectors, representing the values to retain from
each distribution. Changing these from default will truncate the distribution.}
}
\value{
A \code{cauchy_prior}, which is a subclass of \code{ernest_prior} with
an efficient implementation of the unit hypercube transformation.
}
\description{
A specialization of \code{\link[=create_prior]{create_prior()}}, where the parameter space is
described by independently distributed Cauchy marginals that are possibly
truncated.
}
\examples{
prior <- create_cauchy_prior(3, location = c(0, 0, -2), scale = c(0.5, 1, 1))
prior$fn(c(0.25, 0.5, 0.75))
}
\references{
For the truncation routine:
Nadarajah, S., & Kotz, S. (2006).
R Programs for Truncated Distributions. Journal of Statistical Software,
Code Snippets, 16(2), 1–8. \url{https://doi.org/10.18637/jss.v016.c02}
}
\seealso{
Other special priors: 
\code{\link{create_normal_prior}()},
\code{\link{create_t_prior}()},
\code{\link{create_uniform_prior}()}
}
\concept{special priors}
