% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_priors.R
\name{create_normal_prior}
\alias{create_normal_prior}
\title{Specify a prior with normally distributed marginals}
\usage{
create_normal_prior(
  n_dim,
  mean = 0,
  sd = 1,
  varnames = NULL,
  lower = -Inf,
  upper = Inf
)
}
\arguments{
\item{n_dim}{A positive integer. The number of dimensions of the prior
distribution.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{varnames}{An optional character vector. Names for the variables in the
prior distribution.}

\item{lower, upper}{Double vectors. Bounds for a truncated distribution.
Recycled to length \code{n_dim}.}
}
\value{
A \code{normal_prior}, a subclass of \code{ernest_prior} with an efficient
implementation of the unit hypercube transformation.
}
\description{
A specialisation of \code{\link[=create_prior]{create_prior()}} where the parameter space is
described by independent normal variables, possibly truncated.
}
\examples{
prior <- create_normal_prior(3)
prior$fn(c(0.25, 0.5, 0.75))
}
\seealso{
\itemize{
\item \code{\link[=create_prior]{create_prior()}} for more on the \code{ernest_prior} object.
\item \code{\link[truncnorm:dtruncnorm]{truncnorm::qtruncnorm()}} for the truncated normal quantile function.
}

Other special_priors: 
\code{\link{create_uniform_prior}()}
}
\concept{special_priors}
