% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{create_normal_prior}
\alias{create_normal_prior}
\alias{create_special_prior}
\alias{create_t_prior}
\alias{create_cauchy_prior}
\alias{create_uniform_prior}
\title{Specify a common prior distribution}
\usage{
create_normal_prior(
  n_dim,
  mean = 0,
  sd = 1,
  varnames = NULL,
  lower = -Inf,
  upper = Inf
)

create_t_prior(
  n_dim,
  df,
  mu = 0,
  sigma = 1,
  ncp = NULL,
  varnames = NULL,
  lower = -Inf,
  upper = Inf
)

create_cauchy_prior(
  n_dim,
  location = 0,
  scale = 1,
  varnames = NULL,
  lower = -Inf,
  upper = Inf
)

create_uniform_prior(lower = 0, upper = 1, n_dim, varnames = NULL)
}
\arguments{
\item{n_dim}{The number of dimensions for the prior distribution.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{varnames}{Optional names for the variables in the prior distribution.}

\item{lower}{Optional lower bounds for the prior distribution.}

\item{upper}{Optional upper bounds for the prior distribution.}

\item{df}{degrees of freedom (\eqn{> 0}, maybe non-integer).  \code{df
      = Inf} is allowed.}

\item{mu}{The location parameter of the distribution.
If \code{ncp == 0} (or \code{NULL}), this is the median.}

\item{sigma}{The scale parameter of the distribution.}

\item{ncp}{non-centrality parameter \eqn{\delta}{delta};
    currently except for \code{rt()}, only for \code{abs(ncp) <= 37.62}.
    If omitted, use the central t distribution.}

\item{location, scale}{location and scale parameters.}
}
\value{
A special instance of \code{ernest_prior} with an efficient implementation
of the prior transformation function.
}
\description{
Create a specialized version of \code{\link[=create_prior]{create_prior()}} for prior spaces
that are marginally independent and share a common marginal distribution.
}
\details{
\code{create_t_prior} differs from \code{\link[stats:TDist]{stats::qt()}} due to the inclusion of
location and scale parameters. If \code{df} is \eqn{\nu}, then \code{create_t_prior}
returns a prior that is equivalent to \deqn{T(\nu) * \sigma + \mu}
}
\examples{
prior <- create_normal_prior(3)
prior$fn(c(0.25, 0.5, 0.75))

bound_prior <- create_normal_prior(3, lower = -1, upper = 1)
bound_prior$fn(c(0.25, 0.5, 0.75))

# A prior for a simple linear regression model
create_t_prior(
 3,
 df = 3,
 varnames = c("beta0", "beta1", "sigma"),
 lower = c(-Inf, -Inf, 0)
)
}
\references{
For the truncation routine:
Nadarajah, S., & Kotz, S. (2006).
R Programs for Truncated Distributions. Journal of Statistical Software,
Code Snippets, 16(2), 1–8. \url{https://doi.org/10.18637/jss.v016.c02}
}
\seealso{
\code{\link[=create_prior]{create_prior()}} for a richer explanation of the \code{ernest_prior} object.
}
