% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_priors.R
\name{create_t_prior}
\alias{create_t_prior}
\title{Specify a prior with Student's t-distributed marginals}
\usage{
create_t_prior(
  n_dim,
  df,
  mu = 0,
  sigma = 1,
  varnames = NULL,
  lower = -Inf,
  upper = Inf
)
}
\arguments{
\item{n_dim}{(positive integer) The dimensionality of the prior distribution.}

\item{df}{degrees of freedom (\eqn{> 0}, maybe non-integer).  \code{df
      = Inf} is allowed.}

\item{mu}{The location parameter of the distribution.
If \code{ncp == 0} (or \code{NULL}), this is the median.}

\item{sigma}{The scale parameter of the distribution.}

\item{varnames}{(optional character vector) A character vector of names for
the variables in the prior distribution.}

\item{lower, upper}{(optional numeric vector) The values to retain from a
truncated distribution.}
}
\value{
A \code{student_t_prior}, which is a subclass of \code{ernest_prior} with
an efficient implementation of the unit hypercube transformation.
}
\description{
A specialization of \code{\link[=create_prior]{create_prior()}}, where the parameter space is
described by independently distributed Student's t marginals that are
possibly truncated.
}
\details{
The parameters used to specify this prior are for the location-scale
t-distribution with location \eqn{\mu}, scale \eqn{\sigma}, and degrees of
freedom \eqn{\nu}. If the random variable \eqn{X} is Student-T distributed,
then \deqn{X \sim \mu + \sigma T} where \eqn{T} follows the Student-T
distribution with \eqn{\nu} degrees of freedom.
}
\examples{
prior <- create_t_prior(3, df = c(1,2,5), mu = c(0,1,2), sigma = c(1,2,3))
prior$fn(c(0.25, 0.5, 0.75))
}
\references{
For the truncation routine:
Nadarajah, S., & Kotz, S. (2006).
R Programs for Truncated Distributions. Journal of Statistical Software,
Code Snippets, 16(2), 1–8. \url{https://doi.org/10.18637/jss.v016.c02}
}
\seealso{
Other special priors: 
\code{\link{create_cauchy_prior}()},
\code{\link{create_normal_prior}()},
\code{\link{create_uniform_prior}()}
}
\concept{special priors}
