% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_likelihood-class.R
\name{ernest_likelihood}
\alias{ernest_likelihood}
\alias{ernest_likelihood.function}
\alias{ernest_likelihood.glm}
\title{Wrap a Log-Likelihood Function for Ernest}
\usage{
ernest_likelihood(object, ...)

\method{ernest_likelihood}{`function`}(object, ...)

\method{ernest_likelihood}{glm}(object, ...)
}
\arguments{
\item{object}{A function or a fitted model object from \code{glm()}.}

\item{...}{Additional arguments passed to the method. These are currently ignored.}
}
\value{
An object of class \code{ernest_likelihood}, which is a function that
computes the log-likelihood of the data given a set of parameters. If the object
is of class \code{glm}, it will also include the original call, \code{\link[=family]{family()}}, and \code{\link[=terms]{terms()}} object.
}
\description{
Wrap a Log-Likelihood Function for Ernest
}
