% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ernest_logging}
\alias{ernest_logging}
\title{Configure logging for ernest runs}
\usage{
ernest_logging(
  dir = tempdir(),
  threshold = "INFO",
  layout = c("json", "logfmt")
)
}
\arguments{
\item{dir}{Character or \code{FALSE.} Directory in which to create the log file.
This directory must already exist. If \code{FALSE}, logging is disabled.}

\item{threshold}{Character. Minimum message level to record, one of
"INFO", "DEBUG", "WARN", "ERROR", or "FATAL".}

\item{layout}{Character. Log output format, either "json" or "logfmt".}
}
\value{
A list with S3 class "ernest_logging" if \code{dir} is a valid directory, or
NULL if \code{dir} is FALSE. The list contains arguments used to create an
instance of \link[log4r:logger]{log4r::logger} when \link{generate} is called. The return value is
also stored in the R option 'ernest_logging'.
}
\description{
Set up log for ernest. #' "INFO"-level logging saves the results of each call
to \link{update_lrps}; "DEBUG"-level logging additionally saves the results of
each call to \link{propose}.
}
\keyword{internal}
