% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_prior-class.R
\name{ernest_prior}
\alias{ernest_prior}
\alias{ernest_prior.function}
\alias{ernest_prior.distribution}
\title{Specify a prior space for nested sampling}
\usage{
ernest_prior(x, ...)

\method{ernest_prior}{`function`}(x, variables, ...)

\method{ernest_prior}{distribution}(x, ...)
}
\arguments{
\item{x}{Either a function or a vector with S3 class \code{distribution}. See details.}

\item{...}{Arguments forwarded to \code{vctrs::vec_as_names}.}

\item{variables}{Either a character vector of names for each free variable in the
parameter space, or a single number indicating the number of variables.}
}
\value{
An \code{ernest_prior} object, which contains the following elements:
}
\description{
Specify a prior space for nested sampling
}
\details{
The \code{ernest_prior} function is used to specify a prior space for nested
sampling. It can take either a function or a vector with S3 class
\code{distribution}. If a function is provided, it should take in a
matrix slice from the unit space and return the slice in the original
parameter space. If a distribution vector is provided, it will be used to
create a function that applies the inverse probability transformation to
each slice.

Note that, at this time, \code{ernest_prior} only supports the following families:
\code{dist_beta}, \code{dist_binomial}, \code{dist_cauchy}, \code{dist_chisq}, \code{dist_exponential},
\code{dist_f}, \code{dist_gamma}, \code{dist_geometric}, \code{dist_hypergeometric},
\code{dist_lognormal}, \code{dist_negbin}, \code{dist_normal}, \code{dist_poisson},
\code{dist_student_t}, \code{dist_uniform}, and \code{dist_weibull}. You may also use
\code{dist_truncated} to specify a truncated distribution around one of the prior
families.
}
