% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-class.R
\docType{class}
\name{ernest_run-class}
\alias{ernest_run-class}
\alias{ernest_run}
\title{Results from nested sampling runs}
\description{
The \code{ernest_run} object contains the results of a nested sampling run. It has
methods for summary, plotting, and for simulating the uncertainty around the
run's estimates.
}
\details{
The \code{ernest_run} object is returned by running a nested sampling procedure
in the \code{ernest} package. It can be used for posterior analysis, evidence
estimation, and diagnostic plotting.
}
\section{Slots}{

\describe{
\item{\code{n_iter}}{Total number of iterations performed.}

\item{\code{n_points}}{Number of points drawn into the live set.}

\item{\code{n_calls}}{Number of calls to the log likelihood function.}

\item{\code{log_lik}}{A vector of log likelihood values associated with each point
generated during the run.}

\item{\code{log_volume}}{A vector of the estimated prior volumes associated with the
removal of each point from the live set.}

\item{\code{log_weight}}{A vector of unnormalized posterior weights for each point.}

\item{\code{log_evidence}}{A vector of log. evidence estimates, generated after the
removal of each point.}

\item{\code{log_evidence_var}}{A vector of uncertainty values associated with each
entry in \code{log_evidence}. 1.}

\item{\code{information}}{A vector of the estimated KL-divergence (or information)
between the prior and posterior distributions.}

\item{\code{id}}{The index of each point within the live set.}

\item{\code{points}}{The number of live points associated with each point's removal.}

\item{\code{calls}}{The number of likelihood calls made when generating a replacement
live point.}

\item{\code{birth}}{The iteration at which the point was created and added to the
live set.}

\item{\code{samples}}{A matrix of the sampled points, expressed in the units of the
prior space.}

\item{\code{samples_unit}}{Identical to \code{samples}, but expressed in the units of the
0-1 hypercube.}
}}

\seealso{
\link{ernest_run_example} for an example object, and \code{\link[=generate]{generate()}} for how to
create a new \code{ernest_run} object. \code{\link[=plot]{plot()}} and \code{\link[=calculate]{calculate()}} on how to use
\code{ernest_run} objects to examine uncertainty in the log volume estimates.
}
