% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErnestWorkspace_class.R
\name{ErnestWorkspace}
\alias{ErnestWorkspace}
\title{The Nested Sampling Workspace}
\value{
A new ErnestWorkspace object, which is an R environment.
}
\description{
Nested sampling is conducted within an environment (implemented as a
stripped-down \code{R6} object). This allows for the state of the sampling to be
continued later. Users aren't expected to access this item directly, rather
they may use the generics provided to interact with the objects in tidy and
readable formats.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{live_units}}{A matrix of live points, stored in the unit hypercube}

\item{\code{live_points}}{A matrix of live points, stored in the orig. param space}

\item{\code{live_lik}}{A vector of log-likelihoods for the live points}

\item{\code{args}}{A list of parameters for a given sampler}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_iter}}{The total number of sampling iterations.}

\item{\code{n_call}}{The total calls made to the likelihood function,
or \code{0L} if no calls have been made yet.}

\item{\code{log_vol}}{The current prior volume occupied by the dead
points, or \code{0.0} if no points have been killed yet.}

\item{\code{log_z}}{The current evidence estimate from the dead
points, or \code{-1.e300} if no points have been killed yet.}

\item{\code{worst_lik}}{The worst log-likelihood value most recently added
to the dead set, or -1.e300 if no points have been killed yet.}

\item{\code{worst_idx}}{Get the index of the current worst point in the live set.}

\item{\code{last_update}}{The number of times the sampler has been updated, or
\code{0L} if no updates have been made yet.}

\item{\code{scale}}{Returns the current scale of the nested sampler.}

\item{\code{integral_rcrd}}{Returns the log_vol and log_lik vectors as a \code{vctrs::rcrd}
object.}

\item{\code{n_points}}{The number of live points in the workspace.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ErnestWorkspace-new}{\code{ErnestWorkspace$new()}}
\item \href{#method-ErnestWorkspace-pop_point}{\code{ErnestWorkspace$pop_point()}}
\item \href{#method-ErnestWorkspace-push_point}{\code{ErnestWorkspace$push_point()}}
\item \href{#method-ErnestWorkspace-get_dead}{\code{ErnestWorkspace$get_dead()}}
\item \href{#method-ErnestWorkspace-clone}{\code{ErnestWorkspace$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErnestWorkspace-new"></a>}}
\if{latex}{\out{\hypertarget{method-ErnestWorkspace-new}{}}}
\subsection{Method \code{new()}}{
Initialize the workspace with the live points and their log-likelihoods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErnestWorkspace$new(log_lik, prior_transform, n_dim, n_points)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_lik}}{The log-likelihood function.}

\item{\code{prior_transform}}{The prior transform function.}

\item{\code{n_dim}}{The number of dimensions in the parameter space.}

\item{\code{n_points}}{The number of live points to use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErnestWorkspace-pop_point"></a>}}
\if{latex}{\out{\hypertarget{method-ErnestWorkspace-pop_point}{}}}
\subsection{Method \code{pop_point()}}{
Log a point and its log-likelihood to the dead set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErnestWorkspace$pop_point(iter, log_z, log_vol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iter}}{The current iteration number.}

\item{\code{log_z}}{The current evidence estimate.}

\item{\code{log_vol}}{The current prior volume.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErnestWorkspace-push_point"></a>}}
\if{latex}{\out{\hypertarget{method-ErnestWorkspace-push_point}{}}}
\subsection{Method \code{push_point()}}{
Add a new point to the live set, replacing the point existing at
index \code{worst_idx}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErnestWorkspace$push_point(iter, new, copy, cur_update)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iter}}{The current iteration number.}

\item{\code{new}}{The new point to add.}

\item{\code{copy}}{The index of the point to copy from the live set.}

\item{\code{cur_update}}{The number of updates made to the sampler.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErnestWorkspace-get_dead"></a>}}
\if{latex}{\out{\hypertarget{method-ErnestWorkspace-get_dead}{}}}
\subsection{Method \code{get_dead()}}{
Get the dead points returned as a matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErnestWorkspace$get_dead(unit_scale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit_scale}}{Whether to return the points in the unit hypercube.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErnestWorkspace-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ErnestWorkspace-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErnestWorkspace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
