% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_run.R
\docType{data}
\name{example_run}
\alias{example_run}
\title{Example Nested Sampling Run with Ernest}
\format{
An object of class \code{ernest_run} containing the results of a nested
sampling run.
}
\source{
This example problem comes from the crash course for the
\href{https://dynesty.readthedocs.io/en/v2.1.5/crashcourse.html}{dynesty}
Python-based nested sampling software.
}
\usage{
example_run
}
\description{
Load a precomputed example nested sampling run generated using the ernest
package. It demonstrates a typical output from a nested sampling run on a
simple 3-dimensional Gaussian likelihood, with a uniform prior over each
dimension. This dataset is intended for use in documentation, tutorials,
and gainining experience with \code{ernest_run}'s S3 methods.
}
\details{
The likelihood used to generate the points is \eqn{MVN(0, \Sigma)}, with
each variance in \eqn{\Sigma} set to 1 and each covariance set to 0.95.
The prior for each parameter is uniform on the interval \verb{[-10, 10\]}.

This run uses the following non-default settings:
\itemize{
\item \code{log_lik}: A 3D multivariate Gaussian with mean zero and covariance
matrix \code{diag(0.95, 3)}.
\item \code{prior}: Uniform over each dimension (x, y, z) in the range [-10, 10].
\item \code{seed}: 42
}

View the \verb{$spec} element of \code{example_run} to see the full R specification
of the likelihood and prior.

[-10, 10]: R:-10,\%2010\%5C
[-10, 10]: R:-10,\%2010
}
\keyword{datasets}
