% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate.ErnestSampler}
\alias{generate.ErnestSampler}
\title{Generate nested samples from an \code{ErnestSampler} object.}
\usage{
\method{generate}{ErnestSampler}(x, max_it = Inf, max_call = Inf, dlogz = 0.05, refresh = FALSE, ...)
}
\arguments{
\item{x}{An \code{ErnestSampler} object.}

\item{max_it}{The maximum number of iterations needed to run the sampler. Must
be a number larger than zero.}

\item{max_call}{Maximum number of calls to the likelihood function. Must
be a number larger than zero.}

\item{dlogz}{The threshold for the remaining prior volume to the total
evidence. Must represent a number larger or equal to zero.}

\item{refresh}{If \code{TRUE}, and if \code{object} already contains nested sampling
results, then the workspace attached to \code{wrk} will be overwritten. Else, the
run will be continued from the most recent iteration provided by ernest.}

\item{...}{Ignored.}
}
\value{
An \code{ErnestSampler} object containing the results of the nested
}
\description{
Generate nested samples from an \code{ErnestSampler} object.
}
