% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{generate_bootstraps}
\alias{generate_bootstraps}
\title{Bootstrap resampling over a nested sampling run.}
\usage{
generate_bootstraps(
  x,
  times = 100,
  units = c("original", "unit_cube"),
  apparent = FALSE,
  ...
)
}
\arguments{
\item{times}{\verb{[[integer(1)]]}\cr The number of bootstrap resamples.}

\item{units}{\verb{[character(1)]}\cr The scale of the sampled points reported in
the resample; see \link{as_draws.ernest}.}

\item{apparent}{\verb{[[logical(1)]]}\cr Whether to perform an additional resample
using a copy of the original run data stored in \code{x}.}

\item{...}{
  Arguments passed on to \code{\link[rlang:check_dots_empty]{rlang::check_dots_empty}}
  \describe{
    \item{\code{env}}{Environment in which to look for \code{...}.}
    \item{\code{error}}{An optional error handler passed to \code{\link[rlang:try_fetch]{try_fetch()}}. Use
this e.g. to demote an error into a warning.}
    \item{\code{call}}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
    \item{\code{action}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
  }}

\item{object}{[\link{ernest_run}]\cr An object containing the results of a nested
sampling run.}
}
\value{
A tibble. The results include a column for an \link{ernest_rcrd} object
containing the resampled run, and a column called \code{id} that has a character
string with the resample identifier.
}
\description{
Bootstrap resampling over a nested sampling run.
}
\details{
Higson et al. (2019) describes a bootstrap resampling procedure for nested
sampling. This involves splitting a run into \code{nlive} runs of one live point,
then merging runs together by sampling from these runs with replacement.
This provides an empirical estimate of the uncertainty due to the
stochastic nature of selecting likelihood shells from the parameter space.
}
\examples{
# Generate 100 bootstrap resamples
generate_bootstraps(example_run)
}
\references{
\itemize{
\item Higson, E., Handley, W., Hobson, M., & Lasenby, A. (2019).
Nestcheck: Diagnostic Tests for Nested Sampling Calculations. Monthly Notices
of the Royal Astronomical Society, 483(2), 2044–2056.
\doi{10.1093/mnras/sty3090}
\item Speagle, J. S. (2020). dynesty: A Dynamic Nested Sampling Package for
Estimating Bayesian Posteriors and Evidences. Monthly Notices of the Royal
Astronomical Society, 493, 3132–3158. \doi{10.1093/mnras/staa278}
}
}
\seealso{
\code{\link[=calculate.ernest_run]{calculate.ernest_run()}} for evidence estimation without
bootstrapping.
}
