% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence_getters.R
\name{get_integration_df}
\alias{get_integration_df}
\title{Get a \code{tibble} containing the integration results of a nested sampling run}
\usage{
get_integration_df(x, exponentiate = TRUE, error = TRUE)
}
\arguments{
\item{x}{An object of class \code{ernest_run}}

\item{exponentiate}{Whether or not to exponentiate the likelihood values}

\item{error}{Whether or not to include the estimate of evidence error in the returned table}
}
\value{
A \code{tibble} with columns
}
\description{
Get a \code{tibble} containing the integration results of a nested sampling run
}
