% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.ernest_sampler}
\alias{glance.ernest_sampler}
\title{Glance at an \code{ernest_sampler} object}
\usage{
\method{glance}{ernest_sampler}(x, ...)
}
\arguments{
\item{x}{An \code{ernest_sampler} object}

\item{...}{Must be left empty.}
}
\value{
A single-row \code{\link[tibble:tibble]{tibble::tibble()}} summarizing the run.
}
\description{
Construct a single row summary of a given ernest run object
}
