% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.ErnestSampler}
\alias{glance.ErnestSampler}
\title{Glance at an \code{ErnestSampler} object}
\usage{
\method{glance}{ErnestSampler}(x, ...)
}
\arguments{
\item{x}{An \code{ErnestSampler} object}

\item{...}{Ignored}
}
\value{
A single-row \code{\link[tibble:tibble]{tibble::tibble()}} summarizing the run.
}
\description{
Construct a single row summary of a given ernest run object
}
