% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn.ernest_run}
\alias{learn.ernest_run}
\title{Bootstrap Resampling for Nested Sampling Uncertainty}
\usage{
\method{learn}{ernest_run}(x, times = 100, include_weights = FALSE, ...)
}
\arguments{
\item{x}{[\link{ernest_run}]\cr An object containing the results of a nested
sampling run.}

\item{times}{\verb{[[integer(1)]]}\cr Number of bootstrap resamples to generate.
Must be larger than zero.}

\item{include_weights}{\verb{[[logical(1)]]}\cr If \code{TRUE}, include a list-column
of resampled weights for each resampled run in the output.}

\item{...}{Reserved for future extensions; must be empty.}
}
\value{
A special class of [\link{tibble}] with one row per resample, containing:
\itemize{
\item \code{niter}: Number of iterations in the resample.
\item \code{log_evidence}: Estimated log-evidence for the resample.
\item One column per parameter: weighted posterior mean for each parameter.
\item One column per parameter: If \code{second_moment = TRUE}, the weighted posterior
second moment (named \verb{pow2_<param>}).
\item If \code{include_weights = TRUE}, a list-column \code{weights} with the resampled
weights and iteration indices for each resampled run.
}

Also contains an attribute \code{original} as a single tibble row, with
statistics taken from the original run.
}
\description{
Computes bootstrap resamples of a nested sampling run to empirically
estimate uncertainty in evidence and posterior summaries.
}
\details{
Higson et al. (2019) describes a bootstrap resampling procedure for nested
sampling. This involves spliting a run into \code{nlive} runs of one live point,
then merging runs together by sampling from these runs with replacement.
This provides an empirical estimate of the uncertainty due to the
stochastic nature of selecting likelihood shells from the parameter space.
}
\examples{
data(example_run)
# Generate 100 bootstrap resamples
learn(example_run, times = 100)
# Include resampled weights for each draw
learn(example_run, times = 10, include_weights = TRUE)
}
\references{
\itemize{
\item Higson, E., Handley, W., Hobson, M., & Lasenby, A. (2019).
Nestcheck: Diagnostic Tests for Nested Sampling Calculations. Monthly Notices
of the Royal Astronomical Society, 483(2), 2044–2056.
\doi{10.1093/mnras/sty3090}
\item Speagle, J. S. (2020). dynesty: A Dynamic Nested Sampling Package for
Estimating Bayesian Posteriors and Evidences. Monthly Notices of the Royal
Astronomical Society, 493, 3132–3158. \doi{10.1093/mnras/staa278}
}
}
\seealso{
\code{\link[=calculate.ernest_run]{calculate.ernest_run()}} for evidence estimation without
bootstrapping.
}
