% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_balls.R
\name{mini_balls}
\alias{mini_balls}
\title{Generate samples from a p-norm ball}
\usage{
mini_balls(enlarge = 1.1, method, p)
}
\arguments{
\item{enlarge}{Double larger or equal to 1. Factor by which to inflate the
ball's volume before sampling.}

\item{method, p}{Pick one of \code{method} and \code{p}:
* \code{method} Sets the distance measure to be used. This must be one of
\code{"euclidean"}, \code{"maximum"}, or \code{"manhattan"}.
* \code{p} A positive number, indicating the p-norm to use.}
}
\value{
A list with class \code{c("mini_ball", "ernest_lrps")}. Use with
\code{\link[=ernest_sampler]{ernest_sampler()}} to specify nested sampling behaviour.
}
\description{
Propose new live points by selecting a random live point \code{c} and drawing
uniformly from a multidimensional p-norm ball with centre \code{c} and radius \code{r}.
\code{r} is set such that the ball encompassess at least one other live point.
}
\details{
The p-norm naturaly define a distance that makes the vector space a
metric space. For two points \code{x} and \code{y}, their distance in \eqn{L^p} space
is given by \deqn{||x-y||_p = \sigma{(x_i - y_i)^p}^{1/p}}.

The p-norm ball is the set of all vectors such that the distance between
itself and the ball's centre \code{c} is less than the radius \code{r}. The distance
\code{r} is updated throughout a run to ensure that at least one other live point
is contained within the ball. To avoid biasing proposals toward the center,
inflating the volume of the ball before sampling with \code{enlarge > 1} is
recommended.
}
\examples{
data(example_run)
euclid_balls <- mini_balls(method = "euclidean")
euclid_balls <- mini_balls(p = 2)

# Suprenum balls (or L-infinity norm)
suprenum_balls <- mini_balls(method = "maximum")
suprenum_balls <- mini_balls(p = Inf, enlarge = 1.1)

ernest_sampler(
  example_run$log_lik_fn,
  example_run$prior,
  sampler = euclid_balls
)
}
\references{
Buchner, J. (2014). A Statistical Test for Nested Sampling Algorithms.
Statistics and Computing, 26(1–2), 383–392.
\doi{10.1007/s11222-014-9512-y}

Buchner, J. (2019). Collaborative Nested Sampling: Big Data versus Complex
Physical Models. Publications of the Astronomical Society of the Pacific,
131(1004), 108005. \doi{10.1088/1538-3873/aae7fc}
}
\seealso{
Other ernest_lrps: 
\code{\link{multi_ellipsoid}()},
\code{\link{no_underrun}()},
\code{\link{rwmh_cube}()},
\code{\link{slice_rectangle}()},
\code{\link{unif_cube}()},
\code{\link{unif_ellipsoid}()}
}
\concept{ernest_lrps}
