% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ellipsoid.R
\name{multi_ellipsoid}
\alias{multi_ellipsoid}
\title{Generate new points from multiple spanning ellipsoids}
\usage{
multi_ellipsoid(enlarge = 1.25)
}
\arguments{
\item{enlarge}{\verb{[double(1)]}\cr Factor by which to inflate the bounding
ellipsoid's volume before sampling (see Details). Must be at least 1.}
}
\value{
\verb{[multi_ellipsoid]}, a named list that inherits from
[\link{ernest_lrps}].
}
\description{
Partitions the prior space into a set of ellipsoids whose union bounds
the live set. New points are created by randomly
selecting an ellipsoid (weighted by their respective volumes), then using it
to generate a random point as in \link{unif_ellipsoid}. Effective for multimodal
posteriors where a single ellipsoid would be inefficient.
}
\details{
Nested likelihood contours for multimodal distributions are poorly
represented by a single ellipsoid. This method fits multiple ellipsoids to
better capture disconnected or elongated regions.

Ellipsoids are generated using the following procedure:
\enumerate{
\item A single ellipsoid is fit to the live set, with volume \eqn{V}.
\item The live set is clustered into two groups using k-means clustering.
\item Ellipsoids are fit to each cluster.
\item The split ellipsoids are accepted if both ellipsoids are non-degenerate,
and if the combined volume of the split ellipsoids is significantly
smaller than the original ellipsoid (calculated using Bayes' Information
Criterion).
\item Steps 2–4 are repeated recursively on each new ellipsoid until no further
splits are accepted, updating \eqn{V} to the volume of the currently split
ellipsoid.
}
}
\examples{
data(example_run)
lrps <- multi_ellipsoid(enlarge = 1.25)

ernest_sampler(example_run$log_lik_fn, example_run$prior, sampler = lrps)
}
\references{
\itemize{
\item Feroz, F., Hobson, M. P., Bridges, M. (2009) MULTINEST: An Efficient and
Robust Bayesian Inference Tool for Cosmology and Particle Physics. Monthly
Notices of the Royal Astronomical Society. 398(4), 1601–1614.
\doi{10.1111/j.1365-2966.2009.14548.x}
\item Speagle, J. S. (2020). Dynesty: A Dynamic Nested Sampling Package for
Estimating Bayesian Posteriors and Evidences. Monthly Notices of the
Royal Astronomical Society, 493, 3132–3158.
\doi{10.1093/mnras/staa278}
}
}
\seealso{
Other ernest_lrps:
\code{\link[=rwmh_cube]{rwmh_cube()}},
\code{\link[=slice_rectangle]{slice_rectangle()}},
\code{\link[=unif_cube]{unif_cube()}},
\code{\link[=unif_ellipsoid]{unif_ellipsoid()}}
}
\concept{ernest_lrps}
