% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_sampling.R
\name{nested_sampling.control}
\alias{nested_sampling.control}
\title{Auxiliary for Controlling Nested Sampling}
\usage{
\method{nested_sampling}{control}(
  num_points = 500,
  first_update = 2,
  update_interval = 0.6,
  max_iter = NA,
  max_call = NA,
  dlogz = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{num_points}{Number of live points. Larger numbers result in a more
finely sampled posterior, in exchange for longer run times.}

\item{first_update}{Number of calls to log-likelihood before the bounding
distribution is reengaged. If this is given as a float, it is
converted to an integer through \code{round(update_interval * num_points)}.}

\item{update_interval}{Number of calls between updates of the bounding
behaviour of certain nested samplers. If this is given as a float, it is
converted to an integer through \code{round(update_interval * num_points)}}

\item{max_iter}{Maximum number of iterations of the nested sampling loop.
If left \code{NA}, the loop will run until another stopping criteria is met.}

\item{max_call}{Maximum number of likelihood evaluations. If left \code{NA}, the
loop will run until another stopping criteria is met.}

\item{dlogz}{Nested sampling will run until the remaining prior volume to
the total evidence falls below this threshold. If set to \code{0}, the
loop will run until another stopping criteria is met.}

\item{verbose}{Whether to print progress updates to the terminal.}

\item{...}{Extra parameters.}
}
\value{
A list with formatted and checked control parameters.
}
\description{
Auxiliary for Controlling Nested Sampling
}
