% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_sampling.R
\name{nested_sampling}
\alias{nested_sampling}
\title{Prepare a nested sampler to evaluate Bayesian evidence}
\usage{
nested_sampling(
  log_lik,
  prior,
  sampler = rwmh_cube(),
  n_points = 500,
  first_update = 2.5,
  update_interval = 1.5
)
}
\arguments{
\item{log_lik}{Either an \code{ernest_likelihood} or a function that takes in a vector
of parameters and returns the corresponding log. likelihood. This parameter is
sent to \code{\link[=create_likelihood]{create_likelihood()}}.}

\item{prior}{An \code{ernest_prior} object, created by \code{\link[=create_prior]{create_prior()}} or
its \link[=create_special_prior]{specializations}.}

\item{sampler}{An \link{ernest_sampling} object, declaring which likelihood-restricted
prior sampler to use.}

\item{n_points}{The number of live points to use in the nested sampling run.}

\item{first_update}{The number of calls to the likelihood function
before the first update to the behaviour of \code{sampler}. If left as a double,
\code{first_update} is set to \code{first_update * n_points}.}

\item{update_interval}{The number of calls to the likelihood function
between updates to \code{sampler} after \code{first_update} has been reached. If left
as a double, \code{update_interval} is set to \code{update_interval * n_points}.}
}
\value{
An \link{ernest_sampler} object.
}
\description{
Initialize a \link{ernest_sampler} object to perform nested sampling with a given
log-likelihood function, prior distribution, and likelihood-restricted prior
specification.
}
\examples{
prior <- create_uniform_prior(n_dim = 2, lower = -1, upper = 1)
ll_fn <- function(x) -sum(x^2)
sampler <- nested_sampling(ll_fn, prior, n_points = 100)
sampler
}
