% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_run_new.R
\name{new_ernest_run}
\alias{new_ernest_run}
\title{Construct a ernest run object}
\usage{
new_ernest_run(sampler, control, result)
}
\value{
A validated ernest_run object
}
\description{
Construct a ernest run object
}
