% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErnestSampler.R
\name{new_ernest_sampler}
\alias{new_ernest_sampler}
\title{Construct an ernest sampler}
\usage{
new_ernest_sampler(
  log_lik,
  prior_transform,
  n_dim,
  n_points,
  first_update,
  between_update,
  verbose,
  wrk = NULL,
  ...,
  class = character()
)
}
\description{
Construct an ernest sampler
}
