% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_transform.R
\name{new_prior_transform}
\alias{new_prior_transform}
\title{Constructor}
\usage{
new_prior_transform(
  fn = NULL,
  dim = NULL,
  names = NULL,
  distribution = NULL,
  lb = NULL,
  ub = NULL
)
}
\description{
Constructor
}
