% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{new_sampler}
\alias{new_sampler}
\title{Make a likelihood-restricted prior sampler}
\usage{
new_sampler(
  log_lik = NULL,
  prior_transform = NULL,
  num_dim = 0,
  name = "Sampler Superclass",
  description = "Powers LRPS within Ernest",
  ...,
  subclass = character()
)
}
\arguments{
\item{log_lik}{A function that takes in one vector of parameters and returns
a scalar log-likelihood value.}

\item{prior_transform}{An object of class "prior_transform" used to transform
values in the unit hypercube values to the original parameter space.}

\item{num_dim}{An integer number of dimensions in the parameter space.}

\item{name}{A name for the given sampler.}

\item{description}{A short description of the sampler.}

\item{...}{Name-value pairs for additional elements of samplers that
subclass this sampler.}

\item{subclass}{The subclasses of this sampler.}
}
\value{
A \code{sampler} object, used to propose new particles under a likelihood
constraint.
}
\description{
This is the base class for creating a likelihood-based prior sampler. All
of ernest's samplers inherit from this superclass.
}
