% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps-class.R
\name{new_uniform_lrps}
\alias{new_uniform_lrps}
\title{Uniform LRPS Sampler (Internal)}
\usage{
new_uniform_lrps(
  log_lik_fn = NULL,
  prior = NULL,
  unit_lik_fn = NULL,
  n_dim = NULL,
  max_loop = 1000000L,
  cache = NULL
)
}
\description{
#' Internal R6 subclass for uniform sampling in the unit cube under a likelihood
#' constraint.
#'
#' @keywords internal
#' @noRd
}
