% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{nobs.ernest_run}
\alias{nobs.ernest_run}
\title{Return the number of posterior particles within an ernest sampling run}
\usage{
\method{nobs}{ernest_run}(run)
}
\arguments{
\item{run}{An \code{ernest_run} object}
}
\value{
A single integer, describing the number of rows in the posterior sample
}
\description{
Return the number of posterior particles within an ernest sampling run
}
