% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ernest_estimate}
\alias{plot.ernest_estimate}
\title{Plot diagnostics for a nested sampling run}
\usage{
\method{plot}{ernest_estimate}(x, which = c("evidence", "weight", "likelihood"), ...)
}
\arguments{
\item{x}{An \link{ernest_estimate} or \link{ernest_run} object.}

\item{which}{An optional character vector, containing any of \code{evidence},
\code{weight}, or \code{likelihood}. Selects the diagnostic plot(s) to produce from \code{x}
(See Details).}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
Invisibly returns \code{x}. A \code{ggplot2::ggplot()} object is printed as a
side effect.
}
\description{
Shows the normalised likelihood, importance weights, and evidence as
functions of log-volume.
}
\details{
Produces plots that visualize the relationship between the prior volume and
quantities estimated during the run. Use \code{which} to select the quantities
you would like to visualize:
\itemize{
\item \code{which = "evidence"} Shows the log-volume regions within a run that
contributed the most to the model evidence estimates (and its error);
\item \code{which = "weight"} Shows the range of samples within the run that feature
points with the most posterior mass.
\item \code{which = "likelihood"} Shows how smoothly the sampler generated new points
which likelihood-restricted prior sampling (see \link{ernest_lrps});
}

If \code{x} is an \code{ernest_run}, these estimates are derived from the run;
if \code{x} is an \code{ernest_estimate} (or \code{ndraws != 0}), these values are
simulated.
}
\note{
Using \code{which} with multiple options requires the \code{patchwork} package to be
installed.
}
\examples{
# Plot integration results from a run.
data(example_run)
plot(example_run)

# Simulate results before plotting.
plot(example_run, ndraws = 100)
}
\seealso{
\itemize{
\item \code{\link[=calculate]{calculate()}} for generating \code{ernest_estimate} objects.
\item \code{\link[=visualize]{visualize()}} for plotting the posterior distributions generated by a run.
}
}
