% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ernest_sampler}
\alias{plot.ernest_sampler}
\title{Plot an \code{ernest_sampler} object}
\usage{
\method{plot}{ernest_sampler}(x, exponentiate = TRUE, true_log_z = NULL, ...)
}
\arguments{
\item{x}{An \code{ernest_sampler} object.}

\item{exponentiate}{Whether to transform log values before plotting.
Note that volume estimates are always plotted in negative log units.}

\item{true_log_z}{The analytic evidence value on a log scale.}

\item{...}{Must be left empty.}
}
\value{
A plot of the run's progress, which is made up of three stacked
plots:
\itemize{
\item Normalized likelihood values over log volumes.
\item importance weights over log volumes.
\item Model evidence over log volumes, with an error envelope showing 1 and 2
SDs from the estimate.
}
}
\description{
Use \code{ggplot} to create a plot of the evidence, importance weights, and
normalized likelihood values over the estimated volumes from a nested
sampling run.
}
