% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ErnestSampler}
\alias{plot.ErnestSampler}
\title{Plot an \code{ErnestSampler} object}
\usage{
\method{plot}{ErnestSampler}(x, true_log_z = NULL, ...)
}
\arguments{
\item{x}{An \code{ErnestSampler} object.}

\item{true_log_z}{The analytic evidence value on a log scale. Currently does
nothing.}

\item{...}{Must be left empty.}
}
\description{
Use \code{ggplot} to create a plot of the evidence, importance weights, and normalized
likelihood values over the estimated volumes from a nested sampling run.
}
