% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_transform.R
\name{prior_transform}
\alias{prior_transform}
\alias{prior_transform.distribution}
\alias{prior_transform.function}
\title{Specify a prior transformation for nested sampling}
\usage{
prior_transform(x, ...)

\method{prior_transform}{distribution}(..., .names = NULL)

\method{prior_transform}{`function`}(fn, num_dim, .names, ...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named probability distributions
describing the marginal priors for each parameter.}

\item{.names}{A character vector of parameter names. Must be provided if
\code{x} is a function}

\item{fn}{A function that takes a vector of parameters in the unit hypercube and
returns a vector of parameters in the prior space.}

\item{num_dim}{The number of dimensions of the prior space.}
}
\value{
A \code{prior_transform} object.
}
\description{
Specify a prior transformation for nested sampling
}
