% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps-class.R, R/multi_ellipsoid.R,
%   R/rwmh_cube.R, R/slice_rectangle.R, R/unif_cube.R, R/unif_ellipsoid.R
\name{propose}
\alias{propose}
\alias{propose.multi_ellipsoid}
\alias{propose.rwmh_cube}
\alias{propose.slice_rectangle}
\alias{propose.unif_cube}
\alias{propose.unif_ellipsoid}
\title{Generate a new point using LRPS}
\usage{
propose(x, original = NULL, criterion = -Inf)

\method{propose}{multi_ellipsoid}(x, original = NULL, criterion = -Inf)

\method{propose}{rwmh_cube}(x, original = NULL, criterion = -Inf)

\method{propose}{slice_rectangle}(x, original = NULL, criterion = -Inf)

\method{propose}{unif_cube}(x, original = NULL, criterion = -Inf)

\method{propose}{unif_ellipsoid}(x, original = NULL, criterion = -Inf)
}
\arguments{
\item{x}{[\link{ernest_lrps}]\cr A likelihood-restricted prior sampler.}

\item{original}{\verb{[double(x$nvar)]}\cr A parameter vector which can be used
to start the proposal process. Optional; if \code{NULL} proposals are generated
from sampling the unconstrained \link[=unif_cube]{unit cube}.}

\item{criterion}{\verb{[double(1)]}\cr A log-likelihood value that restricts
the region of prior space to sample from.}
}
\value{
\verb{[list]}, containing at least these named elements:
\itemize{
\item \code{unit}: \verb{[double(x$nvar)]} The replacement point, expressed in the
unit-cube.
\item \code{log_lik}: \verb{[double(1)]} The log-likelihood value at \code{unit}.
\item \code{neval}: \verb{[integer(1)]} The number of evaluations of \code{unit_log_fn} needed
to generate \code{unit}. Do not confuse this with the number of samples evaluated
during the call to \code{propose}, as \code{unit_log_fn} may be capable of evaluating
multiple parameters with a single call.
}
}
\description{
Use an LRPS method to replace points within the live set during a nested
sampling run. This method must be implemented on all classes that inherit
from \code{ernest_lrps}.
}
\keyword{internal}
