% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{propose_live}
\alias{propose_live}
\title{Propose a point by sampling from the bounded prior distribution.}
\usage{
propose_live(sampler, original, min_lik)
}
\arguments{
\item{sampler}{The sampler object, inheriting from \code{ernest_sampler}}

\item{original}{The original position of the live point.}

\item{min_lik}{The log-likelihood criterion for the new particle.}
}
\value{
A particle, which is a list with four components: "unit", "parameter",
"log_lik", and "num_calls". Returns NULL if the log-likelihood cannot be
overcome after a million tries.
}
\description{
Propose a point by sampling from the bounded prior distribution.
}
