% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{propose_uniform}
\alias{propose_uniform}
\title{Propose a new live point by randomly sampling within the unit hypercube.}
\usage{
propose_uniform(sampler, min_lik)
}
\arguments{
\item{sampler}{The sampler object, inheriting from \code{ernest_sampler}}

\item{min_lik}{The log-likelihood criterion for the new particle.}
}
\value{
A particle, which is a list with four components: "unit", "parameter",
"log_lik", and "num_calls". Returns NULL if the log-likelihood cannot be
overcome after a million tries.
}
\description{
Propose a new live point by randomly sampling within the unit hypercube.
}
