% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_priors.R
\name{recycle_params}
\alias{recycle_params}
\title{Recycle prior parameters to the right length}
\usage{
recycle_params(names, lower, upper, ..., .call = caller_env())
}
\description{
Recycle prior parameters to the right length
}
