% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{refresh_sampler}
\alias{refresh_sampler}
\title{Refresh a sampler}
\usage{
refresh_sampler(sampler)
}
\arguments{
\item{sampler}{A sampler.}
}
\value{
The \code{sampler} after a call to the corresponding constructor.
}
\description{
This runs the sampler through the method-specific \verb{new_*_sampler()} function
to ensure that all of the elements of the sampler are still valid.
}
