% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_run_example.R
\name{run_example}
\alias{run_example}
\title{Example Nested Sampling Run with Ernest}
\usage{
run_example()
}
\value{
An \code{ernest_run} object containing the results of the nested
sampling run.
}
\description{
This dataset contains the results of a nested sampling run performed on a
three-dimensional Gaussian likelihood with independent uniform priors on
each parameter.
}
\details{
The likelihood used to generate the points is \eqn{MVN(0, \Sigma)}, with
each variance in \eqn{\Sigma} set to 1 and each covariance set to 0.95.
The prior for each parameter is uniform on the interval \verb{[-10, 10\]}.

[-10, 10]: R:-10,\%2010\%5C
}
\examples{
\dontrun{
# Load the example run
run_example()
}
}
\seealso{
\code{\link[=nested_sampling]{nested_sampling()}} \linkS4class{ernest_run}
}
