% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_cube.R
\name{rw_cube}
\alias{rw_cube}
\title{Likelihood-Prior Restricted Sampling with A Random Walk within the Unit Cube}
\usage{
rw_cube(steps = 20, epsilon = 0.1)
}
\arguments{
\item{steps}{The number of steps to take in the random walk}

\item{epsilon}{The initial step size of the random walk. This is refined throughout
the run to target a 0.5 acceptance rate.}
}
\value{
An object of class \code{rw_cube}, which inherits from \code{ernest_sampler}
}
\description{
Generate particles by randomly walking away from an existing live point.
The walk reflects off of the bounds of the unit cube.
}
\details{
This sampler is based off the original approach adopted by John Skilling.
A metropolis-hastings random walk is conducted from a given live point,
where each step is accepted based whether on its evaluated likelihood is better
compared against the likelihood criterion.
}
