% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_run}
\alias{sample_run}
\title{Return the importance-weighed posterior sample from a nested sampling run}
\usage{
sample_run(run, size = nobs(run))
}
\arguments{
\item{run}{An \code{ernest_run} object}

\item{size}{The number of rows to returns, defaults to the full sample}

\item{...}{Not yet}
}
\value{
A tibble with the importance-weighed posterior sample
}
\description{
Return the importance-weighed posterior sample from a nested sampling run
}
