% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{sampler_info}
\alias{sampler_info}
\title{Short lil line helping to print info about the samplers}
\format{
An object of class \code{character} of length 2.
}
\usage{
sampler_info
}
\description{
Short lil line helping to print info about the samplers
}
\keyword{datasets}
