% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_transform.R
\name{set_prior_transform}
\alias{set_prior_transform}
\alias{set_prior_transform.distribution}
\alias{set_prior_transform.function}
\alias{set_prior_transform.prior_transform}
\alias{set_prior_transform.list}
\title{Specify a prior transformation for nested sampling}
\usage{
set_prior_transform(x, ...)

\method{set_prior_transform}{distribution}(x, names, ...)

\method{set_prior_transform}{`function`}(x, names, lb = -Inf, ub = Inf, ...)

\method{set_prior_transform}{prior_transform}(...)

\method{set_prior_transform}{list}(x)
}
\arguments{
\item{x}{Either a distribution3 distribution, a function, or a prior_transform object.}

\item{...}{If 'x' is a prior transform object, additional prior_transform
objects to combine together. Else, these dots should be empty.}

\item{names}{A vector of character strings, naming each transformed variable.}

\item{lb, ub}{If 'x' is a function, the lower and upper bound for the
transformed variable, respectively.}
}
\value{
A prior_transform object.
}
\description{
Specify a prior transformation for nested sampling
}
