% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\title{Generate samples with slice sampling}
\usage{
slice(enlarge = 1)
}
\arguments{
\item{enlarge}{Optional double, greater than or equal to 1. Factor by which
to inflate the hyperrectangle's volume before sampling (see Details).}
}
\value{
An object of class \code{c("slice", "ernest_lrps")} that can be
used with \code{\link[=ernest_sampler]{ernest_sampler()}} to specify the sampling behaviour.
}
\description{
Create new live points by evolving a current live point through
slice sampling within a bounding hyperrectangle, shrinking the rectangle
when proposals are rejected.
}
\details{
The slice LRPS generates proposals by uniformly sampling within a bounding
hyperrectangle that contains regions of the parameter space satisfying the
likelihood criterion. Sampling begins by selecting a known live point
\eqn{\theta} that satisfies the criterion. Each iteration proposes
a new point within this rectangle via uniform sampling and compares it
against the criterion; if rejected, a new hyperrectangle is drawn such that
the proposed point is on its boundary and \eqn{\theta} is in its interior.
This continues until either a valid proposal is found or the rectangle has
shrunk to the point where no further clamping operations can be performed.

By default, the hyperrectangle spans the extreme values of the current
set of live points in each dimension. This may risk excluding valid regions
of the parameter space, particularly where the posterior is multimodal or
highly non-Gaussian. To mitigate this, set \code{enlarge > 1}, which inflates the
hyperrectagle's volume by the specified factor before sampling. Setting
\code{enlarge} to \code{NA} disables this behaviour, instead slicing from the unit
hypercube at each iteration.
}
\examples{
# Basic usage with default parameters
lrps <- slice()

# More patient sampler for complex posteriors
patient_lrps <- slice(enlarge = 1.25)

}
\references{
Neal, R. M. (2000). Slice Sampling (Version 1). arXiv.
\url{https://doi.org/10.48550/ARXIV.PHYSICS/0009028}
}
\seealso{
Other ernest_lrps: 
\code{\link{adaptive_rwmh}()},
\code{\link{multi_ellipsoid}()},
\code{\link{rwmh_cube}()},
\code{\link{unif_cube}()},
\code{\link{unif_ellipsoid}()}
}
\concept{ernest_lrps}
