% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_run_metrics.R
\name{summarise_run_metrics}
\alias{summarise_run_metrics}
\title{Get a \code{tibble} of a nested sampling run}
\usage{
summarise_run_metrics(x, exponentiate = TRUE, normalise = FALSE)
}
\arguments{
\item{x}{An object of class \code{ernest_run}}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
likelihood, importance weight, and evidence values.}

\item{normalise}{Logical indicating whether or not to normalise the
likelihood values.}
}
\value{
A \code{tibble} with columns
}
\description{
Get a \code{tibble} of a nested sampling run
}
