% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{summary.ernest_estimate}
\alias{summary.ernest_estimate}
\title{Summarise an evidence estimate from nested sampling}
\usage{
\method{summary}{ernest_estimate}(object, ...)
}
\arguments{
\item{object}{An \code{ernest_estimate} object.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A list with summary statistics for the evidence estimate:
\itemize{
\item \code{n_iter}: Number of iterations.
\item \code{n_points}: Number of live points.
\item \code{log_volume}: Mean log volume at the final iteration.
\item \code{log_evidence}: Mean log evidence at the final iteration.
\item \code{log_evidence_err}: Standard error of the log evidence.
** If \code{ndraws = 0}, this is computed analytically.
** If \code{ndraws} is 1 or greater, it is estimated as the standard deviation
of \code{log_evidence} across the draws.
\item \code{run}: A \link[posterior:draws_rvars]{posterior::draws_rvars} of all draws and variables.
}
}
\description{
Provides a summary of an \code{ernest_estimate} object.
}
\seealso{
\code{\link[=calculate]{calculate()}}
}
