% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ernest_sampler.R
\name{summary.ernest_sampler}
\alias{summary.ernest_sampler}
\title{Summarize an \code{ernest_sampler} object}
\usage{
\method{summary}{ernest_sampler}(object, ...)
}
\arguments{
\item{object}{An \code{ernest_sampler} object.}

\item{...}{Must be empty.}
}
\value{
A list of class \code{summary.ernest_sampler} containing:
\itemize{
\item \code{n_points}: The number of live points used in the sampling process.
\item \code{n_iterations}: The total number of iterations performed during the run.
\item \code{n_calls}: The total number of calls made to the likelihood function.
\item \code{sampler}: A list containing details about the sampler configuration:
\itemize{
\item \code{lrps}: The likelihood-restricted prior sampler (LRPS) used.
\item \code{first_update}: The first update interval.
\item \code{update_interval}: The subsequent update interval.
}
\item \code{samples}: A list containing:
\itemize{
\item \code{original}: A matrix of sampled points in the original parameter space.
\item \code{unit}: A matrix of sampled points in the unit hypercube space.
}
\item \code{log_likelihood}: A vector of log-likelihood values for the sampled points.
\item \code{log_volume}: A vector of log-volume values for the sampled points.
\item \code{log_importance_weight}: A vector of log-importance weights for the
sampled points.
\item \code{log_evidence}: The estimated log-evidence of the model.
\item \code{log_evidence.sd}: The standard deviation of the log-evidence estimate.
\item \code{information}: The estimated information content of the model.
}
}
\description{
This function provides a detailed summary of a nested sampling run, including
information about the sampling process, the sampler configuration, and the
results of the run.
}
