% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_cube.R
\name{unit_cube}
\alias{unit_cube}
\title{Unit Cube Likelihood-Prior Sampling}
\usage{
unit_cube(max_attempts = 1e+06)
}
\value{
An object of class \code{unit_cube}, which inherits from \code{ernest_sampler}
}
\description{
Generate particles by naively sampling from a unit hypercube. This is a
simple but inefficient method of sampling from the prior distribution.
}
\details{
This sampler can be useful in a few edge cases where the number of live
points is small compared to the number of dimensions, or if the user wants to
run tests to check the sampling behaviour.
}
