% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{update_bounds}
\alias{update_bounds}
\title{Update the bounds of a given sampler with the current distribution
of live points}
\usage{
update_bounds(sampler, live)
}
\arguments{
\item{sampler}{The current live points}

\item{live}{The current live points}
}
\value{
A sampler, with its bounding behaviour updated.
}
\description{
Update the bounds of a given sampler with the current distribution
of live points
}
