% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{update_sampler}
\alias{update_sampler}
\title{Update a sampler object}
\usage{
update_sampler(sampler, ...)
}
\arguments{
\item{sampler}{A sampler.}

\item{...}{Name-value pairs of \emph{existing} elements in \code{sampler} that should
be updated.}
}
\description{
Updates the parameters within a sampler object. Unlike using \code{$},
this function checks that the updated name already exists in the object,
and runs \code{refresh_sampler()} afterwards to check validity.
}
