% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErnestSampler.R
\name{validate_ernest_sampler}
\alias{validate_ernest_sampler}
\title{Validate an Ernest Sampler Object}
\usage{
validate_ernest_sampler(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
x, invisibly.
}
\description{
Validate an Ernest Sampler Object
}
