% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{variables.ernest_prior}
\alias{variables.ernest_prior}
\alias{variables.ernest_sampler}
\alias{nvariables.ernest_prior}
\alias{nvariables.ernest_sampler}
\title{Get Variable Names from an \code{ernest_prior} object}
\usage{
\method{variables}{ernest_prior}(x, ...)

\method{variables}{ernest_sampler}(x, ...)

\method{nvariables}{ernest_prior}(x, ...)

\method{nvariables}{ernest_sampler}(x, ...)
}
\arguments{
\item{x}{An \code{ernest_prior} or \code{ernest_sampler} object.}

\item{...}{Must be left empty.}
}
\value{
\code{variables()} returns a vector of all variable names. \code{nvariables()}
returns the number of variables.
}
\description{
Get variable names from \code{ernest} objects.
}
