% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{variables.ernest_sampler}
\alias{variables.ernest_sampler}
\alias{variables<-.ernest_sampler}
\alias{set_variables.ernest_sampler}
\alias{nvariables.ernest_sampler}
\title{View, set, and count the variables in an \code{ernest_sampler} object}
\usage{
\method{variables}{ernest_sampler}(x, ...)

\method{variables}{ernest_sampler}(x, ...) <- value

\method{set_variables}{ernest_sampler}(x, variables, ...)

\method{nvariables}{ernest_sampler}(x, ...)
}
\arguments{
\item{x}{An \code{ernest_sampler} object.}

\item{...}{Must be left empty.}

\item{variables, value}{Either a character vector or empty tibble of variable names,
or a scalar value describing the dimensions of the parameter space (in which
case, variable names will be denoted with \code{X...i}).}
}
\value{
For \code{\link[=variables]{variables()}}, a character vector of variable names. For
\code{\link[=variables<-]{variables<-()}} and \code{\link[=set_variables]{set_variables()}}, the modified \code{ernest_sampler} object.
For \code{\link[=nvariables]{nvariables()}}, a scalar integer.
}
\description{
Read and modify the variable names within an \code{ernest_sampler} object, or
count the variables (i.e., the dimensionality of the nested sampling problem.
}
\note{
The \code{ernest_sampler} stores an empty prototype tibble to keep track
of the variable names and the dimensionality of the sampled prior space.
Overwriting the number of parameters within an existing sampler is not supported.
}
