% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{as_draws.ernest_sampler}
\alias{as_draws.ernest_sampler}
\alias{as_draws_list.ernest_sampler}
\alias{as_draws_matrix.ernest_sampler}
\title{Transform \code{ernest_sampler} to \code{posterior::draws} Objects}
\usage{
\method{as_draws}{ernest_sampler}(x, scale = c("original", "unit"), ...)

\method{as_draws_list}{ernest_sampler}(x, scale = c("original", "unit"), ...)

\method{as_draws_matrix}{ernest_sampler}(x, scale = c("original", "unit"), ...)
}
\arguments{
\item{x}{An \code{ernest_sampler} object.}

\item{scale}{A character string, either \code{"original"} (default) or \code{"unit"}.
If \code{"unit"}, the returned \code{draws} object contains points in the unit
hypercube. If \code{"original"}, the points are in the original parameter space.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{posterior::draws} object.
}
\description{
Convert an \code{ernest_sampler} object to a format compatible with the
posterior package.
}
\seealso{
\code{\link[posterior:draws]{posterior::as_draws()}}, \code{\link[posterior:draws_list]{posterior::as_draws_list()}},
\code{\link[posterior:draws_matrix]{posterior::as_draws_matrix()}}
}
