% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{as_draws.ErnestSampler}
\alias{as_draws.ErnestSampler}
\alias{as_draws_list.ErnestSampler}
\title{Transform \code{ErnestSampler} to \code{posterior::draws} objects}
\usage{
\method{as_draws}{ErnestSampler}(x, scale = c("original", "unit"), inc_live = TRUE, ...)

\method{as_draws_list}{ErnestSampler}(x, scale = c("original", "unit"), inc_live = TRUE, ...)
}
\arguments{
\item{x}{An \code{ErnestSampler} object, or another \R object for which
the methods are defined.}

\item{scale}{One of \code{original} or \code{unit}. If \code{unit}, the returned
\code{draws} object will contain points expressed in the unit hypercube, which
was originally used by ernest for generated sampling. If \code{original} (the
default), the returned \code{draws} object will contain points in the original
parameter space defined by the prior transformation.}

\item{inc_live}{Should the live points be included? Defaults to \code{TRUE}.}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\description{
Transform an \code{ErnestSampler} object to a format supported by the
posterior package.
}
