% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{as_draws.ernest_run}
\alias{as_draws.ernest_run}
\alias{as_draws_matrix.ernest_run}
\alias{as_draws_rvars.ernest_run}
\title{Convert Nested Samples to \code{draws} Objects}
\usage{
\method{as_draws}{ernest_run}(x, ..., unit_cube = FALSE, radial = FALSE)

\method{as_draws_matrix}{ernest_run}(x, ..., unit_cube = FALSE, radial = FALSE)

\method{as_draws_rvars}{ernest_run}(x, ...)
}
\arguments{
\item{x}{An \code{ernest_draws} object.}

\item{...}{These dots are for future extensions and must be empty.}

\item{unit_cube}{Whether to return the points in their unit cube coordinates.
If \code{FALSE} (default), the points are returned in the original prior space.}

\item{radial}{Whether to return an additional column \code{.radial}, containing
the radial coordinate (i.e., the squared sum of squares) for each sampled
point.}
}
\value{
A \code{draws_matrix} object containing the points and their associated
log weights.
}
\description{
Transform the points stored in an ``ernest_draws\verb{object into a format supported by the}posterior` package.
}
