% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws.R
\name{as_draws.ernest_run}
\alias{as_draws.ernest_run}
\alias{as_draws_matrix.ernest_run}
\alias{as_draws_rvars.ernest_run}
\title{Transform nested sampling runs to \code{draws} objects}
\usage{
\method{as_draws}{ernest_run}(x, units = c("original", "unit_cube"), radial = FALSE, ...)

\method{as_draws_matrix}{ernest_run}(x, units = c("original", "unit_cube"), radial = FALSE, ...)

\method{as_draws_rvars}{ernest_run}(x, units = c("original", "unit_cube"), radial = FALSE, ...)
}
\arguments{
\item{x}{\code{\link[=ernest_run]{[ernest_run]}}\cr Results from a nested sampling run.}

\item{units}{\verb{[character(1)]}\cr The scale of the sampled points:
\itemize{
\item \code{"original"}: Points are on the scale of the prior space.
\item \code{"unit_cube"}: Points are on the (0, 1) unit hypercube scale.
}}

\item{radial}{\verb{[logical(1)]}\cr If \code{TRUE}, returns an additional column
\code{.radial} containing the radial coordinate (i.e., the Euclidean norm) for
each sampled point.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
\code{\link[posterior:draws_matrix]{posterior::draws_matrix()}} or \code{\link[posterior:draws_rvars]{posterior::draws_rvars()}}\cr
A  object containing the posterior samples from the nested sampling run,
with a hidden \code{.weights} column containing the importance weights for each
sample.
}
\description{
Access the posterior sample and weights from a nested sampling run as an
object supported by the \link[posterior:posterior-package]{posterior} package.
}
\note{
To produce a weighted posterior sample, use
\code{\link[posterior:resample_draws]{posterior::resample_draws()}} to reweigh an object from \code{as_draws} using its
importance weights.
}
\examples{
library(posterior)
data(example_run)

# View importance weights
dm <- as_draws(example_run)
weights(dm) |> head()

# Summarise points after resampling
dm |>
  resample_draws() |>
  summarize_draws()

# View the radial coordinate in unit space over the run
dm_rad <- as_draws_rvars(
  example_run,
  units = "unit_cube",
  radial = TRUE
)
plot(x = draws_of(dm_rad$.radial))
}
\seealso{
\itemize{
\item \code{\link[posterior:draws]{posterior::as_draws()}} for details on the \code{draws} object.
}
}
