% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws.R
\name{as_draws_rvars.ernest_run}
\alias{as_draws_rvars.ernest_run}
\title{Transform to draws objects}
\usage{
\method{as_draws_rvars}{ernest_run}(x, resample_draws = TRUE, ...)
}
\arguments{
\item{x}{The earnest run object}
}
\value{
A \code{posterior::draws_matrix} object, which has classes
\code{posterior::draws} and \code{matrix}
}
\description{
Transform to draws objects
}
