% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate.ernest_run}
\alias{calculate.ernest_run}
\alias{ernest_estimate}
\title{Estimate Evidence using a Nested Sampling Run}
\usage{
\method{calculate}{ernest_run}(x, ndraws = 1000L, ...)
}
\arguments{
\item{x}{[\link{ernest_run}]\cr Results from a nested sampling run.}

\item{ndraws}{\verb{[integer(1)]}\cr The number of log-volume sequences to
simulate. If equal to zero, log-volume simulation is skipped and error in
the log-evidence estimates is approximated with analytical error
estimates.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
An \code{ernest_estimate} object, which inherits from \code{tbl_df}, \code{tbl},
and \code{data.frame}.

The iterative estimates from the nested sampling run. Contains the following
columns:
\itemize{
\item \code{log_lik}: \verb{[[double()]]} The log-likelihood of the model.
\item \code{log_volume}: \verb{[[posterior::rvar()]]} The log-volume of the prior space.
\item \code{log_weight}: \verb{[[posterior::rvar()]]} The log weights of the points in the
live set.
\item \code{log_evidence}: \verb{[[posterior::rvar()]]} The log-evidence of the model.
}

If \code{ndraws > 0}, \code{log_volume}, \code{log_weight}, and \code{log_evidence} each contain
\code{ndraws} simulated draws per iteration.

If \code{ndraws = 0}, \code{log_volume} and \code{log_weight} contain a single
deterministic draw per iteration, and \code{log_evidence} contains draws
from a normal approximation based on analytical variance estimates (see
the package vignetttes for more information). The  number of draws is
controlled with getOption("posterior.rvar_ndraws"), with a default of 1000.
}
\description{
Computes evidence and related quantities from a nested
sampling run, optionally by simulating the volumes of each nested
likelihood shell.
}
\examples{
# Load an example run
data(example_run)

# View results and analytical evidence errors.
calculate(example_run, ndraws = 0)

# Simulate 100 log-volume shrinkage sequences across the run.
calculate(example_run, ndraws = 100)

}
\references{
Higson, E., Handley, W., Hobson, M., & Lasenby, A. (2019).
Nestcheck: Diagnostic Tests for Nested Sampling Calculations. Monthly Notices
of the Royal Astronomical Society, 483(2), 2044–2056.
\doi{10.1093/mnras/sty3090}
}
