% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate.ernest_run}
\alias{calculate.ernest_run}
\title{Estimate Evidence using a Nested Sampling Run}
\usage{
\method{calculate}{ernest_run}(x, ..., ndraws = FALSE)
}
\arguments{
\item{x}{An \code{ernest_run} object.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ndraws}{A positive integer or a boolean. If \code{FALSE}, the log volumes
will be derived from their expected values. If not \code{FALSE}, the log volumes
will be simulated using \code{ndraws} samples from each volume's
joint distribution.}
}
\value{
A tibble, containing \code{run$n_iter + run$n_points} rows and the
following columns:
\itemize{
\item \code{log_lik}: The log-likelihood of the model.
\item \code{log_volume}: The log volume of the prior space.
\item \code{log_weight}: The log weights of the live points.
\item \code{log_evidence}: The log evidence of the model.
\item \code{log_evidence.err}: The standard error of the log evidence (only available
when \code{ndraws = 0}).
}

The tibble has the additional class \code{ernest_estimates}, which has its own
\link[=plot.ernest_estimates]{plot} method.

Each column is returned as an \code{\link[posterior:rvar]{posterior::rvar()}} vector.
}
\description{
Estimate Evidence using a Nested Sampling Run
}
\examples{
# Load an example run
data(ernest_run_example)

# View results as a tibble with `ndraws = FALSE` (the default).
calculate(ernest_run_example)

# Generate 100 simulated log volume values for each iteration.
calculate(ernest_run_example, ndraws = 100)

}
