% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate.ernest_run}
\alias{calculate.ernest_run}
\title{Estimate Evidence using a Nested Sampling Run}
\usage{
\method{calculate}{ernest_run}(x, ..., ndraws = FALSE)
}
\arguments{
\item{x}{An \code{ernest_run} object.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ndraws}{A positive integer or a boolean. If \code{FALSE}, the log volumes
will be derived from their expected values. If not \code{FALSE}, the log volumes
will be simulated using \code{ndraws} samples from each volume's
joint distribution.}
}
\value{
A tibble, containing the following columns:
\itemize{
\item \code{log_lik}: The log-likelihood of the model.
\item \code{log_volume}: The log volume of the prior space.
\item \code{log_weight}: The log weights of the live points.
\item \code{log_evidence}: The log evidence of the model.
}

These are all returned as \code{posterior::rvar()} vectors, which allows one vector
to store \code{ndraws} samples.

Additionally, in the case of \code{ndraws = 0}, the following column is also provided:
\itemize{
\item \code{log_evidence.err}: The standard error of the log evidence.
}
}
\description{
Estimate Evidence using a Nested Sampling Run
}
