% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate.ernest_run}
\alias{calculate.ernest_run}
\alias{ernest_estimate}
\title{Estimate Evidence using a Nested Sampling Run}
\usage{
\method{calculate}{ernest_run}(x, ndraws = 1000L, ...)
}
\arguments{
\item{x}{[\link{ernest_run}]\cr Results from a nested sampling run.}

\item{ndraws}{\verb{[integer(1)]}\cr The number of log-volume sequences to
simulate.
If equal to zero, no simulations will be made, and a one draw vector of
log-volumes are produced from the estimates contained in \code{x}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
[\code{\link[tibble:tibble]{tibble::tibble()}}] with class \code{ernest_estimate}.

The iterative estimates from the nested sampling run. Contains the following
columns:
\itemize{
\item \code{log_lik}: [\link{rvar}] The log-likelihood of the model.
\item \code{log_volume}: [\link{rvar}] The log-volume of the prior space.
\item \code{log_weight}: [\link{rvar}] The log weights of the points in the live set.
\item \code{log_evidence}: [\link{rvar}] The log-evidence of the model.
}

If \code{ndraws = 0}, an additional column is included:
\itemize{
\item \code{log_evidence_err}: [\link{rvar}] The standard error of the log-evidence.
}
}
\description{
Computes evidence and related quantities from a nested
sampling run, optionally by simulating the volumes of each nested
likelihood shell.
}
\examples{
# Load an example run
data(example_run)

# View results as a tibble with `ndraws = 0`.
calculate(example_run, ndraws = 0)

# Generate 100 simulated log-volume values for each iteration.
calculate(example_run, ndraws = 100)

}
\references{
Higson, E., Handley, W., Hobson, M., & Lasenby, A. (2019).
Nestcheck: Diagnostic Tests for Nested Sampling Calculations. Monthly Notices
of the Royal Astronomical Society, 483(2), 2044–2056.
\doi{10.1093/mnras/sty3090}
}
