% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate.ernest_run}
\alias{calculate.ernest_run}
\title{Estimate Evidence using a Nested Sampling Run}
\usage{
\method{calculate}{ernest_run}(x, ..., ndraws = FALSE)
}
\arguments{
\item{x}{An \code{ernest_run} object.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ndraws}{A positive integer or a boolean. If \code{FALSE}, the log volumes
will be derived from their expected values. If not \code{FALSE}, the log volumes
will be simulated using \code{ndraws} samples from each volume's
joint distribution.}

\item{approx_live}{If \code{ndraws} is not \code{FALSE}, whether to approximate the
distribution of log volumes of the live points using their associated
marginals.}
}
\value{
A tibble.
}
\description{
Estimate Evidence using a Nested Sampling Run
}
