% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate.ernest_run}
\alias{calculate.ernest_run}
\alias{ernest_estimate}
\title{Estimate Evidence using a Nested Sampling Run}
\usage{
\method{calculate}{ernest_run}(x, ndraws = 1000L, ...)
}
\arguments{
\item{x}{An \code{ernest_run} object.}

\item{ndraws}{A positive integer. The number of log-volume
sequences to simulate. If equal to zero, no simulations will be made, and a
one draw vector of log-volumes are produced from the estimates contained in
\code{x}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}, containing \code{n_iter + n_points} rows
and several columns:
\itemize{
\item \code{log_lik}: The log-likelihood of the model.
\item \code{log_volume}: The log-volume of the prior space.
\item \code{log_weight}: The log weights of the live points.
\item \code{log_evidence}: The log-evidence of the model.
\item \code{log_evidence_err}: The standard error of the log-evidence (only available
when \code{ndraws = 0}).
}

The tibble has the additional class \code{ernest_estimate}, which has its own
\link[=plot.ernest_estimate]{plot} method.

Each column is returned as an \code{\link[posterior:rvar]{posterior::rvar()}} vector.
}
\description{
Computes evidence and related quantities from a nested
sampling run, optionally by simulating the volumes of each nested
likelihood shell.
}
\examples{
# Load an example run
data(example_run)

# View results as a tibble with `ndraws = 0`.
calculate(example_run, ndraws = 0)

# Generate 100 simulated log-volume values for each iteration.
calculate(example_run, ndraws = 100)

}
\references{
Higson, E., Handley, W., Hobson, M., & Lasenby, A. (2019).
Nestcheck: Diagnostic Tests for Nested Sampling Calculations. Monthly Notices
of the Royal Astronomical Society, 483(2), 2044–2056.
\doi{10.1093/mnras/sty3090}
}
