% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate.ernest_sampler}
\alias{calculate.ernest_sampler}
\title{Calculate statistics from an \code{ernest_sampler} object.}
\usage{
\method{calculate}{ernest_sampler}(x, ...)
}
\arguments{
\item{x}{An \code{ernest_sampler} object.}

\item{...}{Ignored.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{log_vol}: The estimated log volume of the prior.
\item \code{log_lik}: The likelihood of the live points at that iteration.
\item \code{log_evidence}: The evidence estimate.
\item \code{log_evidence_err}: The error in the evidence estimate.
}
}
\description{
Calculate statistics from an \code{ernest_sampler} object.
}
