% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile.ernest_sampler}
\alias{compile.ernest_sampler}
\title{Configure an ErnestSampler object}
\usage{
\method{compile}{ernest_sampler}(object, refresh = FALSE, ...)
}
\arguments{
\item{object}{An \code{ernest_sampler} object.}

\item{refresh}{If \code{TRUE}, and if \code{object} already contains a list of live points,
the existing live points will be overwritten.}

\item{...}{Ignored.}
}
\value{
A copy of \code{object}, invisibly.
}
\description{
Clean and prepare an \code{ernest_sampler} so that it can be used to generate nested samples.
}
