% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{compile.ernest_sampler}
\alias{compile.ernest_sampler}
\title{Prepare an \code{ernest_sampler} object for nested sampling}
\usage{
\method{compile}{ernest_sampler}(object, clear = FALSE, ...)
}
\arguments{
\item{object}{An \code{ernest_sampler} object.}

\item{clear}{A logical value indicating whether to reload \code{object} before
creating new live points. If \code{TRUE}, former results from the sampler are
removed from the object. If \code{FALSE}, the sampler will not drop prior results,
and continue using live points from the last sampling iteration performed.}

\item{...}{Must be empty.}
}
\value{
The modified \code{ernest_sampler} object, invisibly.
}
\description{
Prepare an \code{ernest_sampler} object for nested sampling
}
\details{
This method wraps a call to \code{ernest_sampler$compile()}.
}
