% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_integral}
\alias{compute_integral}
\title{=====
Compute an integral}
\usage{
compute_integral(log_lik, log_vol)
}
\description{
@param log_lik Vector of log likelihoods
@param log_vol Vector of log volumes
}
\details{
@return The integral
}
