% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_likelihood-class.R
\name{create_likelihood}
\alias{create_likelihood}
\alias{create_likelihood.function}
\alias{create_likelihood.glm}
\title{Specify a (Log) Likelihood Function}
\usage{
create_likelihood(object)

\method{create_likelihood}{`function`}(object, ...)

\method{create_likelihood}{glm}(object, ...)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{x}{A log-likelihood function, or an object inheriting from \code{glm},
such as one from a call to \code{\link[=glm]{glm()}}.}
}
\value{
A function of class \code{create_likelihood}, which calculates the
log-likelihood of a model at a given parameter vector. The function
is guaranteed to return a finite value, \code{-Inf}, or an error.
}
\description{
Create a standardized log-likelihood function for use by ernest.
}
