% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-class.R
\name{create_likelihood}
\alias{create_likelihood}
\alias{ernest_likelihood}
\title{Prepare a likelihood function for nested sampling}
\usage{
create_likelihood(fn, on_nonfinite = c("warn", "quiet", "abort"))
}
\arguments{
\item{fn}{A function that takes a vector of parameters and returns a
scalar log-likelihood value (either a finite double or \code{-Inf}).}

\item{on_nonfinite}{A case-sensitive string. Action to perform when \code{fn}
returns a value that is non-finite and not \code{-Inf} (i.e., \code{NaN}, \code{NA}, \code{Inf}):
\itemize{
\item \code{"warn"}: Issue a warning and return \code{-Inf}.
\item \code{"quiet"}: Silently return \code{-Inf}.
\item \code{"abort"}: Stop execution and signal an error.
}}
}
\value{
A function with class \code{ernest_likelihood}. When provided a parameter vector
this function will always return either a scalar finite double, the value
\code{-Inf}, or an error message.
}
\description{
Creates a modified version of a log-likelihood function that always returns
either a finite value or \code{-Inf} for each vector of parameters provided.
}
\details{
Model likelihoods should be provided as a log-density function. The first
argument of \code{fn} should be a vector of free parameters.

If the model likelihood is conditional on some data, then use this step to
incorporate this data into your nested sampling run. We recommended using
an (anonymous function)\code{\link[rlang:as_function]{rlang::as_function()}} to do this (see Examples).

It is expected that the log-likelihood function returns a scalar finite
double or \code{-Inf} for each parameter vector. Non-finite values other than
\code{-Inf}, such as \code{NaN}, \code{Inf}, or \code{NA} (i.e. missing values) are handled
with the behavior of \code{on_nonfinite}.
}
\examples{
data(epilepsy, package = "brms")

# Pooled Model of Seizure Count via. Treatment and Baseline Seizure Count:
# glm(Count ~ zBase * Trt, family = poisson(), data = epilepsy)
#
# Building a Likelihood Function
mod <- stats::glm(count ~ zBase * Trt, family = poisson(), data = epilepsy)
mle <- stats::logLik(mod)

new_x <- model.matrix(
  object = mod$formula,
  data = mod$model,
  terms = mod$terms,
  contrasts.arg = attr(model.matrix(mod), "contrasts")
)
new_y <- model.response(mod$model)

linkinv <- stats::poisson()$linkinv

poisson_glm_ll <- function(x) {
  eta <- new_x \%*\% x
  mu <- linkinv(eta)
  sum(stats::dpois(new_y, lambda = mu, log = TRUE))
}
#' Custom function returns the expected log-lik. at the MLE.
all.equal(
  as.double(mle),
  poisson_glm_ll(mod$coefficients),
  check.attributes = FALSE
)

ll <- create_likelihood(poisson_glm_ll)
ll

# Controlling Missing Data Behaviour
new_x_na <- new_x
new_x_na[50:60, "zBase"] <- NA
na_poisson_glm_ll <- function(x) {
  eta <- new_x_na \%*\% x
  mu <- linkinv(eta)
  sum(stats::dpois(new_y, lambda = mu, log = TRUE))
}

# By default, ernest warns users and replaces NAs with `-Inf`
ll <- create_likelihood(na_poisson_glm_ll)
ll(mod$coefficients)

# Use `on_nonfinite` to specify how NAs should be treated
ll <- create_likelihood(na_poisson_glm_ll, on_nonfinite = "abort")
try(ll(mod$coefficients))

ll <- create_likelihood(na_poisson_glm_ll, on_nonfinite = "quiet")
ll(mod$coefficients)
}
