% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_likelihood-class.R
\name{create_likelihood}
\alias{create_likelihood}
\alias{create_likelihood.function}
\title{Specify a (Log) Likelihood Function}
\usage{
create_likelihood(object)

\method{create_likelihood}{`function`}(object, ...)
}
\arguments{
\item{object}{An object of class function or \code{ernest_likelihood}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A function of class \code{create_likelihood}, which calculates the
log-likelihood of a model at a given parameter vector. The function
is guaranteed to return a finite value, \code{-Inf}, or an error.
}
\description{
Create a standardized log-likelihood function for use by ernest.
}
