% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_priors.R
\name{create_normal_prior}
\alias{create_normal_prior}
\title{Specify a prior with normally-distributed marginals.}
\usage{
create_normal_prior(
  n_dim,
  mean = 0,
  sd = 1,
  varnames = NULL,
  lower = -Inf,
  upper = Inf
)
}
\arguments{
\item{n_dim}{(positive integer) The dimensionality of the prior distribution.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{varnames}{(optional character vector) A character vector of names for
the variables in the prior distribution.}

\item{lower, upper}{(optional numeric vector) The values to retain from a
truncated distribution.}
}
\value{
A \code{normal_prior}, which is a subclass of \code{ernest_prior} with
an efficient implementation of the unit hypercube transformation.
}
\description{
A specialization of \code{\link[=create_prior]{create_prior()}}, where the parameter space is
described by independently distributed normal variables that are possibly
truncated.
}
\examples{
prior <- create_normal_prior(3)
prior$fn(c(0.25, 0.5, 0.75))
}
\references{
For the truncation routine:
Nadarajah, S., & Kotz, S. (2006).
R Programs for Truncated Distributions. Journal of Statistical Software,
Code Snippets, 16(2), 1–8. \url{https://doi.org/10.18637/jss.v016.c02}
}
\seealso{
\code{\link[=create_prior]{create_prior()}} for a richer explanation of the \code{ernest_prior}
object.

Other special priors: 
\code{\link{create_cauchy_prior}()},
\code{\link{create_t_prior}()},
\code{\link{create_uniform_prior}()}
}
\concept{special priors}
