% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_prior-class.R
\name{create_prior}
\alias{create_prior}
\title{Specify a Prior Distribution for Nested Sampling}
\usage{
create_prior(fn, n_dim, varnames = NULL, lower = NULL, upper = NULL)
}
\arguments{
\item{fn}{A function that takes a vector of probabilities \code{p} and returns
a vector of quantiles.}

\item{n_dim}{The number of dimensions for the prior distribution.}

\item{varnames}{Optional names for the variables in the prior distribution.}

\item{lower}{Optional lower bounds for the prior distribution.}

\item{upper}{Optional upper bounds for the prior distribution.}
}
\value{
An object of type \code{ernest_prior} that represents the prior
distribution, including:
\itemize{
\item \code{fn}: The function that computes the quantiles.
\item \code{n_dim}: The number of dimensions for the prior distribution.
\item \code{varnames}: Names for the variables in the prior distribution.
\item \code{lower}: Lower bounds for the prior distribution.
\item \code{upper}: Upper bounds for the prior distribution.
}
}
\description{
Specify a Prior Distribution for Nested Sampling
}
