% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_prior-class.R
\name{create_normal_prior}
\alias{create_normal_prior}
\alias{create_t_prior}
\alias{create_cauchy_prior}
\alias{create_uniform_prior}
\title{Special Prior Distributions}
\usage{
create_normal_prior(
  n_dim,
  mean,
  sd,
  varnames = NULL,
  lower = -Inf,
  upper = Inf
)

create_t_prior(
  n_dim,
  df,
  mu = 0,
  sigma = 1,
  ncp = NULL,
  varnames = NULL,
  lower = -Inf,
  upper = Inf
)

create_cauchy_prior(
  n_dim,
  location = 0,
  scale = 1,
  varnames = NULL,
  lower = -Inf,
  upper = Inf
)

create_uniform_prior(lower = 0, upper = 1, n_dim, varnames = NULL)
}
\arguments{
\item{n_dim}{The number of dimensions for the prior distribution.}

\item{mean}{Vector of means.}

\item{sd}{Vector of standard deviations.}

\item{varnames}{Optional names for the variables in the prior distribution.}

\item{lower}{Optional lower bounds for the prior distribution.}

\item{upper}{Optional upper bounds for the prior distribution.}

\item{df, mu, sigma}{Vector of parameters for the Student-T distribution.}

\item{ncp}{Optional vector of non-centrality parameters.}

\item{location, scale}{Vector of parameters for the Cauchy distribution.}
}
\value{
A subclass of \code{ernest_prior}.
}
\description{
These functions create specific types of prior distributions for use in
ernest.
}
