% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_priors.R
\name{create_uniform_prior}
\alias{create_uniform_prior}
\title{Specify a prior with uniformly distributed marginals}
\usage{
create_uniform_prior(
  lower = 0,
  upper = 1,
  names = "Uniform",
  .n_dim = NULL,
  .name_repair = c("unique", "universal", "check_unique")
)
}
\arguments{
\item{lower, upper}{Numeric vector of bounds for the uniform distribution.}

\item{names}{An optional character vector. Names for the variables in the
prior distribution.}

\item{.n_dim}{An optional positive integer. The number of dimensions of the
prior distribution.}

\item{.name_repair}{An optional, case-sensitive string. How to repair
\code{names}. Options are \code{"unique"} (default), \code{"universal"}, or
\code{"check_unique"}. See \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} for details.}
}
\value{
A \code{uniform_prior}, a subclass of \code{ernest_prior} with an efficient
implementation of the unit hypercube transformation.
}
\description{
A specialisation of \code{\link[=create_prior]{create_prior()}} where the parameter space is
described by independent uniform marginals.
}
\note{
See \link[vctrs:theory-faq-recycling]{vctrs::vector_recycling_rules} for additional information on
how parameters are recycled to a common length.
}
\examples{
prior <- create_uniform_prior(lower = c(3, -2), upper = c(5, 4))
prior$fn(c(0.33, 0.67))
}
\seealso{
Other special_priors: 
\code{\link{create_normal_prior}()}
}
\concept{special_priors}
