% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_prior.R
\name{default_prior}
\alias{default_prior}
\alias{default_prior.glm}
\title{Construct a default prior distribution for a given model}
\usage{
default_prior(object, ...)

\method{default_prior}{glm}(object, ...)
}
\arguments{
\item{object}{A fitted model object from \code{\link[=glm]{glm()}}.}

\item{...}{}
}
\value{
A vector with class \code{distributional}, describing each prior placed
on the model coefficients, and named accordingly.
}
\description{
Construct a default prior distribution for a given model
}
