% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest-package.R
\docType{package}
\name{ernest-package}
\alias{ernest}
\alias{ernest-package}
\title{ernest: A Toolkit for Nested Sampling}
\description{
Bayesian evidence estimation and posterior inference with the nested sampling algorithm, along with S3 methods for simulating uncertainty and creating visualisations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/ernest}
  \item \url{https://docs.ropensci.org/ernest/}
  \item Report bugs at \url{https://github.com/ropensci/ernest/issues}
}

}
\author{
\strong{Maintainer}: Kyle Dewsnap \email{kyle.dewsnap@ubc.ca} (\href{https://orcid.org/0000-0003-2132-8083}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kyle Dewsnap \email{kyle.dewsnap@ubc.ca} (\href{https://orcid.org/0000-0003-2132-8083}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item TJ Mahr [reviewer]
  \item Robert Kubinec [reviewer]
}

}
\keyword{internal}
