% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest-package.R
\docType{package}
\name{ernest-package}
\alias{ernest}
\alias{ernest-package}
\title{ernest: The R Nested Sampling Toolkit}
\description{
A complete implementation of the nested sampling algorithm for R, allowing users to estimate the evidence and posterior of a statistical model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kylesnap.github.io/ernest/}
  \item Report bugs at \url{https://github.com/kylesnap/ernest/issues/}
}

}
\author{
\strong{Maintainer}: Kyle Dewsnap \email{kyle.dewsnap@ubc.ca} (\href{https://orcid.org/0000-0003-2132-8083}{ORCID})

}
\keyword{internal}
