% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_prior-class.R
\name{ernest_prior}
\alias{ernest_prior}
\alias{ernest_prior.function}
\alias{ernest_prior.distribution}
\title{Specify a prior space for nested sampling}
\usage{
ernest_prior(x, ...)

\method{ernest_prior}{`function`}(
  x,
  names,
  repair = c("unique", "universal", "check_unique", "unique_quiet", "universal_quiet"),
  ...
)

\method{ernest_prior}{distribution}(
  x,
  names = vec_names(x),
  repair = c("unique", "universal", "check_unique", "unique_quiet", "universal_quiet"),
  call = args$call,
  ...
)
}
\arguments{
\item{x}{A vector of class \code{distribution}, describing each variable in the
parameter space.}

\item{...}{Arguments forwarded to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.}

\item{names}{A character vector of length equal to \code{x} or \code{NULL}.}

\item{repair}{Either a string or a function. See \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} for
details on name repair.}
}
\value{
An \code{ernest_prior} object, which contains the following elements:
}
\description{
Specify a prior space for nested sampling
}
