% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrd-class.R
\name{ernest_rcrd}
\alias{ernest_rcrd}
\title{Storing nested sampling results}
\usage{
ernest_rcrd(
  unit = matrix(double(0)),
  log_lik = double(0),
  id = character(0),
  nlive = integer(0),
  neval = integer(0),
  birth_lik = double(0)
)
}
\arguments{
\item{unit}{\verb{[matrix()]}\cr Coordinates in the unit hypercube, with one row
per sample and one column per variable.}

\item{log_lik}{\verb{[double()]}\cr Log-likelihood values for each sample.}

\item{id}{\verb{[character()]}\cr Unique identifier for each point within a
nested sampler.}

\item{nlive}{\verb{[integer()]}\cr The number of live points in the run at the
time each sample was generated.}

\item{neval}{\verb{[integer()]}\cr The number of likelihood evaluations performed
to generate each sample.}

\item{birth_lik}{\verb{[double()]}\cr The log-likelihood threshold at which
each sample was generated, i.e. the log-likelihood of the worst live point
at the time of replacement.}
}
\value{
An \code{ernest_rcrd} object, which is a vctrs record class designed to
store the history of samples generated during a nested sampling run.
}
\description{
\code{ernest_rcrd} is a [\link[vctrs:new_rcrd]{vctrs::new_rcrd}] class that stores metadata for all
samples drawn from the nested sampling algorithm, tracking the history of
replacements in the live set and each point's contribution to evidence
estimation.
}
\details{
This object is designed to be used internally within the \code{ernest_run} class
to track the history of samples generated during a nested sampling run.
Generally, users will not interact with \code{ernest_rcrd} objects directly, and
instead will call methods on \code{ernest_run} objects which, internally,
manipulate the \code{ernest_rcrd} object.
}
\references{
Fowlie, A., Handley, W., Su, L., Nested Sampling with Plateaus,
Monthly Notices of the Royal Astronomical Society, 503(1), 1199–1205,
\doi{10.1093/mnras/stab590}
}
\keyword{internal}
