% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_sampler-class.R
\name{ernest_sampler-class}
\alias{ernest_sampler-class}
\alias{ernest_sampler}
\title{The Ernest Nested Sampler Object}
\description{
An R6 class that contains a nested sampling run.

This object is normally created by calling \code{\link[=nested_sampling]{nested_sampling()}}, and
interacted with by calling S3 methods like \code{\link[=generate]{generate()}} and \code{\link[=calculate]{calculate()}}.
}
\seealso{
\code{\link[=compile.ernest_sampler]{compile.ernest_sampler()}}

\code{\link[=generate.ernest_sampler]{generate.ernest_sampler()}}

\code{\link[=calculate.ernest_sampler]{calculate.ernest_sampler()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n_iterations}}{The total number of sampling iterations.}

\item{\code{n_calls}}{The total calls made to the likelihood function,
or \code{0L} if no calls have been made yet.}

\item{\code{prior}}{The \code{ernest_prior} object associated with the sampler.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_iterations}}{The total number of sampling iterations.}

\item{\code{n_calls}}{The total calls made to the likelihood function,
or \code{0L} if no calls have been made yet.}

\item{\code{prior}}{The \code{ernest_prior} object associated with the sampler.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ernest_sampler-new}{\code{ernest_sampler$new()}}
\item \href{#method-ernest_sampler-clear}{\code{ernest_sampler$clear()}}
\item \href{#method-ernest_sampler-compile}{\code{ernest_sampler$compile()}}
\item \href{#method-ernest_sampler-generate}{\code{ernest_sampler$generate()}}
\item \href{#method-ernest_sampler-get_live_points}{\code{ernest_sampler$get_live_points()}}
\item \href{#method-ernest_sampler-get_dead_points}{\code{ernest_sampler$get_dead_points()}}
\item \href{#method-ernest_sampler-calculate}{\code{ernest_sampler$calculate()}}
\item \href{#method-ernest_sampler-summary}{\code{ernest_sampler$summary()}}
\item \href{#method-ernest_sampler-print}{\code{ernest_sampler$print()}}
\item \href{#method-ernest_sampler-clone}{\code{ernest_sampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{ernest_sampler}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$new(
  log_lik_fn,
  prior,
  sampler,
  n_points,
  first_update,
  update_interval
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-clear"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-clear}{}}}
\subsection{Method \code{clear()}}{
Clear the previous runs from the sampler, including the sampler's live points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$clear()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-compile"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-compile}{}}}
\subsection{Method \code{compile()}}{
Generate a sample of live points from the prior, and check the live
points for finite values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$compile(clear = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-generate"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-generate}{}}}
\subsection{Method \code{generate()}}{
Generate samples from nested sampling until a given criterion is met.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$generate(max_iterations = Inf, max_calls = Inf, min_logz = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_iterations}}{The maximum number of iterations to perform. If set to
Inf, this stopping criterion is ignored.}

\item{\code{max_calls}}{The maximum number of calls to the likelihood function.
If set to Inf, this stopping criterion is ignored.}

\item{\code{min_logz}}{The minimum log-evidence value to achieve. Must be a number
strictly larger than zero.}

\item{\code{refresh}}{Whether to clear existing points from the sampler, starting
a run from scratch.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-get_live_points"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-get_live_points}{}}}
\subsection{Method \code{get_live_points()}}{
Access the sampler's live points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$get_live_points(units = c("original", "unit"), reorder = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{units}}{A string, either \code{"original"} or \code{"unit"}.
If \code{"original"}, the original points are returned.
If \code{"unit"}, the unit points are returned.}

\item{\code{reorder}}{Whether to reorder the points so they appear in increasing
log-likelihood values.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble} containing the live points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-get_dead_points"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-get_dead_points}{}}}
\subsection{Method \code{get_dead_points()}}{
Access the discarded (or \code{dead}) points list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$get_dead_points(units = c("original", "unit"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{units}}{A string, either \code{"original"} or \code{"unit"}.
If \code{"original"}, the original points are returned.
If \code{"unit"}, the unit points are returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble} containing the dead points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-calculate"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-calculate}{}}}
\subsection{Method \code{calculate()}}{
Calculate the evidence integral of the nested sampling run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$calculate(include_live = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_live}}{Whether to include the live points in the calculation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble} with columns reporting the results of the run.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-summary}{}}}
\subsection{Method \code{summary()}}{
Calculate the evidence integral of the nested sampling run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-print"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-print}{}}}
\subsection{Method \code{print()}}{
Print a brief summary of the sampler to the string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments forwarded to \code{\link[=format]{format()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
