% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_sampler-class.R
\name{ernest_sampler-class}
\alias{ernest_sampler-class}
\alias{ernest_sampler}
\title{The Ernest Nested Sampler Object}
\description{
An R6 class that contains a nested sampling run.

This object is normally created by calling \code{\link[=nested_sampling]{nested_sampling()}}, and
interacted with by calling S3 methods like \code{\link[=generate]{generate()}} and \code{\link[=calculate]{calculate()}}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{niterations}}{The total number of sampling iterations.}

\item{\code{ncalls}}{The total calls made to the likelihood function.}

\item{\code{variables}}{The variables associated with the prior.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ernest_sampler-new}{\code{ernest_sampler$new()}}
\item \href{#method-ernest_sampler-clear}{\code{ernest_sampler$clear()}}
\item \href{#method-ernest_sampler-compile}{\code{ernest_sampler$compile()}}
\item \href{#method-ernest_sampler-generate}{\code{ernest_sampler$generate()}}
\item \href{#method-ernest_sampler-get_live_points}{\code{ernest_sampler$get_live_points()}}
\item \href{#method-ernest_sampler-get_dead_points}{\code{ernest_sampler$get_dead_points()}}
\item \href{#method-ernest_sampler-calculate}{\code{ernest_sampler$calculate()}}
\item \href{#method-ernest_sampler-summary}{\code{ernest_sampler$summary()}}
\item \href{#method-ernest_sampler-print}{\code{ernest_sampler$print()}}
\item \href{#method-ernest_sampler-clone}{\code{ernest_sampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{ernest_sampler}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$new(
  log_lik_fn,
  prior,
  sampler,
  n_points,
  first_update,
  update_interval
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_lik_fn}}{A function representing the log-likelihood.}

\item{\code{prior}}{An \code{ernest_prior} object defining the prior distribution.}

\item{\code{sampler}}{A likelihood-restricted prior sampler (LRPS).}

\item{\code{n_points}}{The number of live points.}

\item{\code{first_update}}{The first update interval.}

\item{\code{update_interval}}{The subsequent update interval.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-clear"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-clear}{}}}
\subsection{Method \code{clear()}}{
Clears the previous runs from the sampler, including the sampler's live points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$clear()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-compile"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-compile}{}}}
\subsection{Method \code{compile()}}{
Generates a sample of live points from the prior and validates them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$compile(clear = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clear}}{A logical value indicating whether to clear existing points.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-generate"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-generate}{}}}
\subsection{Method \code{generate()}}{
Performs nested sampling until a stopping criterion is met.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$generate(max_iterations = Inf, max_calls = Inf, min_logz = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_iterations}}{The maximum number of iterations to perform.}

\item{\code{max_calls}}{The maximum number of calls to the likelihood function.}

\item{\code{min_logz}}{The minimum log-evidence value to achieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-get_live_points"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-get_live_points}{}}}
\subsection{Method \code{get_live_points()}}{
Accesses the sampler's live points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$get_live_points(units = c("original", "unit"), reorder = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{units}}{A string, either \code{"original"} or \code{"unit"}.}

\item{\code{reorder}}{Whether to reorder the points by increasing log-likelihood.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble} containing the live points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-get_dead_points"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-get_dead_points}{}}}
\subsection{Method \code{get_dead_points()}}{
Accesses the sampler's discarded (dead) points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$get_dead_points(units = c("original", "unit"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{units}}{A string, either \code{"original"} or \code{"unit"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble} containing the dead points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-calculate"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-calculate}{}}}
\subsection{Method \code{calculate()}}{
Calculates the evidence integral of the nested sampling run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$calculate(include_live = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_live}}{Whether to include live points in the calculation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble} with columns reporting the results of the run.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-summary}{}}}
\subsection{Method \code{summary()}}{
Summarize the nested sampling run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
See \code{\link[=summary.ernest_sampler]{summary.ernest_sampler()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-print"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-print}{}}}
\subsection{Method \code{print()}}{
Prints a brief summary of the sampler.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments, forwarded to \code{\link[=prettyNum]{prettyNum()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
