% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_sampler-class.R
\name{ernest_sampler}
\alias{ernest_sampler}
\title{The Ernest Sampler Object}
\value{
A copy of \code{self}, invisibly.

A single-row tibble reporting the state of the sample.
}
\description{
The Ernest Sampler Object

The Ernest Sampler Object
}
\details{
Objects of class \code{ernest_sampler} are responsible for managing the objects
involved in nested sampling.
\itemize{
\item Create a nested sampler with (compile)
\item Generate nested samples with (generate)
\item Calculate an estimated evidence integral with (calculate)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{verbose}}{Whether to produce output during a run.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_iterations}}{The total number of sampling iterations.}

\item{\code{n_calls}}{The total calls made to the likelihood function,
or \code{0L} if no calls have been made yet.}

\item{\code{variables}}{The names of each variable in the prior space.}

\item{\code{live_points}}{Access the live points list.}

\item{\code{dead_points}}{Access the dead points list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ernest_sampler-new}{\code{ernest_sampler$new()}}
\item \href{#method-ernest_sampler-compile}{\code{ernest_sampler$compile()}}
\item \href{#method-ernest_sampler-generate}{\code{ernest_sampler$generate()}}
\item \href{#method-ernest_sampler-calculate}{\code{ernest_sampler$calculate()}}
\item \href{#method-ernest_sampler-glance}{\code{ernest_sampler$glance()}}
\item \href{#method-ernest_sampler-format}{\code{ernest_sampler$format()}}
\item \href{#method-ernest_sampler-print}{\code{ernest_sampler$print()}}
\item \href{#method-ernest_sampler-clone}{\code{ernest_sampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-new}{}}}
\subsection{Method \code{new()}}{
Initialize an ernest sampler with an lrps, number of points, and
verbosity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$new(lrps, ptype, n_points, verbose = getOption("verbose"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lrps}}{The likelihood-restricted prior sampler to use}

\item{\code{ptype}}{Either a single integer, a vector of variable names, or a
zero-row \code{\link[tibble:tibble]{tibble::tibble()}} describing the dimensions of the prior space.}

\item{\code{n_points}}{The number of live points to use}

\item{\code{verbose}}{Whether to produce verbose output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-compile"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-compile}{}}}
\subsection{Method \code{compile()}}{
Create a starting set of live points
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$compile(refresh = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{refresh}}{If \code{TRUE}, and if the \code{ernest_sampler} already contains
a list of live points, the existing live points will be overwritten.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-generate"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-generate}{}}}
\subsection{Method \code{generate()}}{
Generate nested samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$generate(
  max_iterations = Inf,
  max_calls = Inf,
  min_logz = 0.05,
  refresh = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_iterations}}{The maximum number of iterations to run.}

\item{\code{max_calls}}{The maximum number of calls to the likelihood function.}

\item{\code{min_logz}}{The minimum log-evidence to reach before stopping.}

\item{\code{refresh}}{If \code{TRUE}, and if the \code{ernest_sampler} already contains
a list of live points, the existing live points will be overwritten.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-calculate"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-calculate}{}}}
\subsection{Method \code{calculate()}}{
Calculate the evidence integral.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$calculate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-glance"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-glance}{}}}
\subsection{Method \code{glance()}}{
Glance at an ernest_sampler object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$glance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-format"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-format}{}}}
\subsection{Method \code{format()}}{
Format an ernest_sampler object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$format(digits = getOption("digits"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{The number of digits to display}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-print"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-print}{}}}
\subsection{Method \code{print()}}{
Print an ernest_sampler object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments forwarded to \code{\link[=format]{format()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
