% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_sampler.R
\name{ernest_sampler}
\alias{ernest_sampler}
\title{Prepare a new nested sampling run}
\usage{
ernest_sampler(
  log_lik,
  prior,
  sampler = rwmh_cube(),
  nlive = 500,
  first_update = NULL,
  update_interval = NULL,
  seed = NA
)
}
\arguments{
\item{log_lik}{\verb{[function]} or [\link{ernest_likelihood}]\cr
A function which computes the log-likelihood of a given model. If a
function, it is wrapped with \code{\link[=create_likelihood]{create_likelihood()}}.}

\item{prior}{[\link{ernest_prior}]\cr Describes the prior space
within which to generate samples.}

\item{sampler}{[\link{ernest_lrps}]\cr Specifies the
likelihood-restricted prior sampling method used to replace points
within the live set.}

\item{nlive}{\verb{[integer(1)]}\cr The number of points to generate
within the live set. Strictly positive.}

\item{first_update}{\verb{[integer(1)]}\cr The number of likelihood
calls to make with the default \link[=unif_cube]{uniform LRPS} method before
swapping to the technique described by \code{sampler}. Optional; if left \code{NULL}
this is set to \code{nlive * 2.5}.}

\item{update_interval}{\verb{[integer(1)]}\cr The number of likelihood
calls between updates to the \code{sampler} object. Optional; if left \code{NULL}
this is set to \code{nlive * 1.5}.}

\item{seed}{\verb{[integer(1)]}\cr Sets the random seed controlling the
random number generator for nested sampling runs. Optional; if left \code{NA}
the \link{.Random.seed} set within R is preserved and restored after a run.}
}
\value{
\verb{[ernest_sampler]}\cr
A named list, containing a specification for a nested sampling run. Contains
the arguments passed to this function as well as an environment \code{run_env},
which is used to store the live set during sampling.
}
\description{
Initializes an \code{ernest_sampler} object containing the components required to
perform nested sampling. This object can then be used to build sequences of
nested samples with \code{\link[=generate]{generate()}}.
}
\details{
The \code{ernest_sampler} object is tested with \code{\link[=compile]{compile()}} before it is
returned. This helps to catch errors with the likelihood and prior
specifications. If this compilation step fails, review your \code{log_lik_fn} and
\code{prior} objects for their compliance.
}
\section{Verbosity}{

Messages from ernest can be silenced with the global options
\code{rlib_message_verbosity} and \code{rlib_warning_verbosity}. These options take the
values:
\itemize{
\item "default": Verbose unless the \code{.frequency} argument is supplied.
\item "verbose": Always verbose.
\item "quiet": Always quiet.
}

When set to quiet, messages are not displayed and the condition is not
signaled. See \code{\link[rlang:abort]{rlang::abort()}} for more information.
}

\examples{
prior <- create_uniform_prior(lower = c(-1, -1), upper = 1)
ll_fn <- function(x) -sum(x^2)
sampler <- ernest_sampler(ll_fn, prior, nlive = 100)
sampler

# Use a unit-cube LRPS (not recommended in practice)
unit_sampler <- ernest_sampler(
  ll_fn,
  prior,
  nlive = 100,
  sampler = unif_cube()
)
unit_sampler
}
