% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_sampler-class.R
\name{ernest_sampler}
\alias{ernest_sampler}
\title{Ernest Nested Sampler}
\value{
An \code{ernest_sampler} object.

\code{object}, invisibly.

A \code{tibble} with columns reporting the results of the run.
}
\description{
An R6 class that contains a nested sampling run.

This object is normally created by calling \code{\link[=nested_sampling]{nested_sampling()}}, and
interacted with by calling S3 methods like \code{\link[=generate]{generate()}}, \code{\link[=calculate]{calculate()}},
and \code{\link[=glance]{glance()}}.
}
\seealso{
\code{\link[=compile.ernest_sampler]{compile.ernest_sampler()}}

\code{\link[=generate.ernest_sampler]{generate.ernest_sampler()}}

\code{\link[=calculate.ernest_sampler]{calculate.ernest_sampler()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{verbose}}{Should a progress bar be displayed during sampling?}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_iterations}}{The total number of sampling iterations.}

\item{\code{n_calls}}{The total calls made to the likelihood function,
or \code{0L} if no calls have been made yet.}

\item{\code{variables}}{The names of each variable in the prior space.}

\item{\code{live_points}}{Access the live points list.}

\item{\code{dead_points}}{Access the dead points list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ernest_sampler-new}{\code{ernest_sampler$new()}}
\item \href{#method-ernest_sampler-compile}{\code{ernest_sampler$compile()}}
\item \href{#method-ernest_sampler-generate}{\code{ernest_sampler$generate()}}
\item \href{#method-ernest_sampler-calculate}{\code{ernest_sampler$calculate()}}
\item \href{#method-ernest_sampler-glance}{\code{ernest_sampler$glance()}}
\item \href{#method-ernest_sampler-format}{\code{ernest_sampler$format()}}
\item \href{#method-ernest_sampler-print}{\code{ernest_sampler$print()}}
\item \href{#method-ernest_sampler-clone}{\code{ernest_sampler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{ernest_sampler}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$new(
  lrps,
  ptype,
  n_points = 500L,
  verbose = getOption("verbose")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lrps}}{An \code{ernest_lrps} object describing the model and prior space.}

\item{\code{ptype}}{The parameters involved in nested sampling. Possible values
are:
\itemize{
\item A single integer, describing the dimensions of the prior space.
\item A character vector, naming each dimension in the prior space.
\item A zero-row \code{\link[tibble:tibble]{tibble::tibble()}}, where each column names a dimension of
the prior space.
}}

\item{\code{n_points}}{The number of live points to use during nested sampling.
Higher values allow for more accurate estimates of the evidence integral
at the cost of increased computational time.}

\item{\code{verbose}}{Whether to display a progress bar during a run.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-compile"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-compile}{}}}
\subsection{Method \code{compile()}}{
Prepare the \code{ernest_sampler} for generating nested samples by validating
the set of live points within the object, creating live points if none
exist yet.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$compile(refresh = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{refresh}}{Whether to clear existing points from \code{object} and generating new ones.
If \code{TRUE}, the function will clear both the live points and dead points gathered
from previous runs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-generate"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-generate}{}}}
\subsection{Method \code{generate()}}{
Generate samples from nested sampling until a given criterion is met.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$generate(
  max_iterations = Inf,
  max_calls = Inf,
  min_logz = 0.05,
  refresh = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_iterations}}{The maximum number of iterations to perform. If set to
Inf, this stopping criterion is ignored.}

\item{\code{max_calls}}{The maximum number of calls to the likelihood function.
If set to Inf, this stopping criterion is ignored.}

\item{\code{min_logz}}{The minimum log-evidence value to achieve. Must be a number
strictly larger than zero.}

\item{\code{refresh}}{Whether to clear existing points from the sampler, starting
a run from scratch.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{x}, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-calculate"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-calculate}{}}}
\subsection{Method \code{calculate()}}{
Calculate the marginal likelihood of a given model and return the estimates
in a tidy \code{\link[=tibble]{tibble()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$calculate(
  add_points = c("none", "unit", "parameter", "both"),
  add_progress = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add_points}}{A string, either \code{"none"}, \code{"unit"}, \code{"parameter"}, or \code{"both"}.
If \code{"none"}, no additional columns are added. If \code{"unit"}or \code{"parameter"},
the parameter values associated with each point are added, in their respective
units. If \code{"both"}, both the unit and parameter values are added.}

\item{\code{add_progress}}{Adds columns for the number of calls to the likelihood
function between each iteration.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-glance"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-glance}{}}}
\subsection{Method \code{glance()}}{
Glance at an ernest_sampler object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$glance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-format"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-format}{}}}
\subsection{Method \code{format()}}{
Format an ernest_sampler object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$format(digits = getOption("digits"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{The number of digits to display}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-print"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-print}{}}}
\subsection{Method \code{print()}}{
Print an ernest_sampler object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments forwarded to \code{\link[=format]{format()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ernest_sampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ernest_sampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ernest_sampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
