% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate.ernest_sampler}
\alias{generate.ernest_sampler}
\title{Generate nested samples from an \code{ernest_sampler} object.}
\usage{
\method{generate}{ernest_sampler}(
  x,
  max_iterations = Inf,
  max_calls = Inf,
  min_logz = 0.05,
  refresh = FALSE,
  ...
)
}
\arguments{
\item{x}{An \code{ernest_sampler} object.}

\item{max_iterations}{The maximum number of iterations needed to run the
sampler. Must be a number larger than zero.}

\item{max_calls}{Maximum number of calls to the likelihood function. Must
be a number larger than zero.}

\item{min_logz}{The threshold for the remaining prior volume to the total
evidence. Must represent a number larger or equal to zero.}

\item{refresh}{If \code{TRUE}, and if the \code{ernest_sampler} already contains
a list of live points, the existing live points will be overwritten.}

\item{...}{Must be left blank.}
}
\value{
\code{x}, invisibly.
}
\description{
Generate nested samples from an \code{ernest_sampler} object.
}
