% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate.ernest_sampler}
\alias{generate.ernest_sampler}
\title{Generate Samples from Nested Sampling}
\usage{
\method{generate}{ernest_sampler}(
  x,
  max_iterations = Inf,
  max_calls = Inf,
  min_logz = 0.05,
  refresh = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{ernest_sampler}.}

\item{max_iterations}{The maximum number of iterations to perform. If set to
Inf, this stopping criterion is ignored.}

\item{max_calls}{The maximum number of calls to the likelihood function.
If set to Inf, this stopping criterion is ignored.}

\item{min_logz}{The minimum log-evidence value to achieve. Must be a number
strictly larger than zero.}

\item{refresh}{Whether to clear existing points from the sampler, starting
a run from scratch.}

\item{...}{Must be empty.}
}
\value{
\code{x}, invisibly.
}
\description{
Use this function to perform nested sampling until a given criterion is met.
Calling \code{\link[=generate]{generate()}} implicitly calls \code{\link[=compile]{compile()}} to ensure the sampler
object is valid before running samples.
}
