% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generate.ernest_sampler}
\alias{generate.ernest_sampler}
\title{Generate samples from nested sampling}
\usage{
\method{generate}{ernest_sampler}(x, max_iterations = Inf, max_calls = Inf, min_logz = 0.05, ...)
}
\arguments{
\item{x}{An object of class \code{ernest_sampler}.}

\item{max_iterations}{The maximum number of iterations to perform. If set to
\code{Inf}, this stopping criterion is ignored.}

\item{max_calls}{The maximum number of calls to the likelihood function.
If set to \code{Inf}, this stopping criterion is ignored.}

\item{min_logz}{The minimum log-evidence value to achieve. Must be a number
equal to or larger than zero.}

\item{...}{Must be empty.}
}
\value{
The modified \code{ernest_sampler} object, invisibly.
}
\description{
Generate samples from nested sampling
}
\details{
This method wraps a call to \code{ernest_sampler$generate()}.
}
