% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generate.ernest_sampler}
\alias{generate.ernest_sampler}
\title{Generate Samples from Nested Sampling}
\usage{
\method{generate}{ernest_sampler}(x, max_iterations = Inf, max_calls = Inf, min_logz = 0.05, ...)
}
\arguments{
\item{x}{An object of class \code{ernest_sampler}.}

\item{max_iterations}{The maximum number of iterations to perform. If set to
Inf, this stopping criterion is ignored.}

\item{max_calls}{The maximum number of calls to the likelihood function.
If set to Inf, this stopping criterion is ignored.}

\item{min_logz}{The minimum log-evidence value to achieve. Be a number equal or
larger than zero, or \code{NULL}, in which case it is transformed to zero.}

\item{...}{Must be empty.}

\item{refresh}{Whether to clear existing points from the sampler, starting
a run from scratch.}
}
\value{
\code{x}, invisibly.
}
\description{
Generate Samples from Nested Sampling
}
\details{
This method wraps a call to \code{\link[=ernest_sampler$generate]{ernest_sampler$generate()}}.
}
