% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.ernest_run}
\alias{merge.ernest_run}
\title{Merge \code{ernest_run} objects together.}
\usage{
\method{merge}{ernest_run}(x, y, suffix = c(".x", ".y"), ...)
}
\arguments{
\item{x}{[\link{ernest_run}]\cr An object containing a nested sampling run.}

\item{y}{[\link{ernest_run}]\cr Another nested sampling run to merge with \code{x}.}

\item{suffix}{\verb{[[character(2)]]} Suffixes to append to the IDs of \code{x} and \code{y}
if there are any duplicate IDs.}
}
\value{
[\link{ernest_run}] An object containing the merged nested sampling run.
}
\description{
Combines the samples from a set of nested sampling runs with differing
numbers of live points into a single run.
}
