% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.ernest_run}
\alias{merge.ernest_run}
\title{Merge \code{ernest_run} objects together.}
\usage{
\method{merge}{ernest_run}(x, y = NULL, ..., keep = c("first", "all"))
}
\arguments{
\item{x}{[\link{ernest_run}]\cr An object containing a nested sampling run.}

\item{y}{[\link{ernest_run}]\cr Another nested sampling run to merge with \code{x}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Additional \code{ernest_run}
objects to merge with \code{x}. Ignored if \code{y} is used.}

\item{keep}{A single string specifying how to determine the live set of the
merged run. See details.}
}
\description{
Merge \code{ernest_run} objects together.
}
