% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_balls.R
\name{mini_balls}
\alias{mini_balls}
\title{Generate samples from a p-norm ball}
\usage{
mini_balls(method = c("euclidean", "maximum"), bootstrap = NULL)
}
\arguments{
\item{method}{Sets the distance measure to be used. This must be one of
\code{"euclidean"} or \code{"maximum"}.}

\item{bootstrap}{Sets the number of bootstrap replications to use when estimating
the largest empty p-ball. If \code{NULL}, jacknife estimation is used instead.}
}
\value{
A list with class \code{c("mini_ball", "ernest_lrps")}. Use with
\code{\link[=ernest_sampler]{ernest_sampler()}} to specify nested sampling behaviour.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Propose new live points from a multidimensional p-norm ball centered
on a randomly selected live point. The radius of the balls are set
such that each ball encompasses at least two live points (i.e., the
largest empty p-ball).
}
\details{
The p-norm naturally defines a distance that makes the vector space
a metric space. For two points \code{x} and \code{y}, their distance in \eqn{L^p}
space is given by \deqn{||x-y||_p = \sigma{(x_i - y_i)^p}^{1/p}}.

The p-norm ball is the set of all vectors such that the distance between
itself and the ball's centre \code{c} is less than the radius \code{r}. The distance
\code{r} is updated throughout a run to ensure that at least one other live point
is contained within the ball.
}
\section{Status}{

This LRPS is experimental and has not been extensively validated across
different nested sampling problems. You are encouraged to use it, but please
exercise caution interpretting results and report any issues or unexpected
behaviour.
}

\examples{
data(example_run)
euclid_balls <- mini_balls(method = "euclidean")

suprenum_balls <- mini_balls(method = "maximum")

ernest_sampler(
  example_run$log_lik_fn,
  example_run$prior,
  sampler = euclid_balls
)
}
\references{
Buchner, J. (2014). A Statistical Test for Nested Sampling Algorithms.
Statistics and Computing, 26(1–2), 383–392.
\doi{10.1007/s11222-014-9512-y}

Buchner, J. (2019). Collaborative Nested Sampling: Big Data versus Complex
Physical Models. Publications of the Astronomical Society of the Pacific,
131(1004), 108005. \doi{10.1088/1538-3873/aae7fc}
}
\seealso{
Other ernest_lrps: 
\code{\link{multi_ellipsoid}()},
\code{\link{no_underrun}()},
\code{\link{rwmh_cube}()},
\code{\link{slice_rectangle}()},
\code{\link{unif_cube}()},
\code{\link{unif_ellipsoid}()}
}
\concept{ernest_lrps}
\keyword{internal}
