% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ellipsoid.R
\name{multi_ellipsoid}
\alias{multi_ellipsoid}
\title{Generate samples from multiple spanning ellipsoids}
\usage{
multi_ellipsoid(enlarge = 1.25, min_reduction = 0.7, allow_contact = TRUE)
}
\arguments{
\item{enlarge}{Double, greater than or equal to 1. Factor by which to inflate
the bounding ellipsoid's volume before sampling (see Details).}

\item{min_reduction}{Double between 0 and 1. The minimum reduction in total
volume required for an ellipsoid to be split in two. Lower values lead to
more aggressive splitting.}

\item{allow_contact}{Logical. Whether to allow ellipsoids to overlap.}
}
\value{
A list with class \code{c("multi_ellipsoid", "ernest_lrps")}. Use with
\code{\link[=ernest_sampler]{ernest_sampler()}} to specify nested sampling behaviour.
}
\description{
Partitions the prior space into a set of ellipsoids whose union bounds
the set of live points. Samples are created by randomly
selecting an ellipsoid (weighted by their respective volumes), then using it
to generate a random point as in \link{unif_ellipsoid}. Effective for multimodal
posteriors where a single ellipsoid would be inefficient.
}
\details{
Nested likelihood contours for multimodal distributions are poorly
represented by a single ellipsoid. This method fits multiple ellipsoids to
better capture disconnected or elongated regions.

Ellipsoids are generated using the following procedure:
\enumerate{
\item A single ellipsoid is fit to the set of live points, with volume \eqn{V}.
\item The live points are clustered into two groups using k-means clustering.
\item Ellipsoids are fit to each cluster.
\item The split ellipsoids are accepted if all of the following conditions are
met:
\itemize{
\item Both ellipsoids are non-degenerate
\item The combined volume of the split ellipsoids is less than
\eqn{min_{red.} * V}
\item (If \code{allow_contact} is \code{FALSE}) the ellipsoids do not intersect.
}
\item Steps 2–4 are repeated recursively on each new ellipsoid until no further
splits are accepted, updating \eqn{V} to the volume of the currently split
ellipsoid.
}
}
\section{Ellipsoids}{

Ellipsoids are stored in the \code{cache} environment of the LRPS object.
Ellipsoids are defined by their centre \eqn{c} and shape matrix \eqn{A}.
The set of points \eqn{x} contained within the ellipsoid is given by
\deqn{ x \in {\bf{R}}^n | (x-c) A (x-c)' \leq 1 }

The volume of the ellipsoid is
\eqn{V = \mathrm{Vol}(S_n) \sqrt{\det(A^{-1})}}, where
\eqn{\mathrm{Vol}(S_n)} is the volume of the unit hypersphere.

For sampling, we store the matrix \eqn{A^{-1/2}}, the inverse of the
positive-semidefinite square root of \eqn{A}. The ellipsoid can equivalently
be defined as the set of points \deqn{x = A^{-1/2} y + c,}
where \eqn{y} are points from the unit hypersphere.

For more on ellipsoids and their operations, see
\href{http://tcg.mae.cornell.edu/pubs/Pope_FDA_08.pdf}{Algorithms for Ellipsoids}
by S.B. Pope, Cornell University Report FDA 08-01 (2008).
}

\examples{
data(example_run)
lrps <- multi_ellipsoid(enlarge = 1.25, min_reduction = 0.5)

ernest_sampler(example_run$log_lik_fn, example_run$prior, sampler = lrps)
}
\references{
Feroz, F., Hobson, M. P., Bridges, M. (2009) MULTINEST: An Efficient and
Robust Bayesian Inference Tool for Cosmology and Particle Physics. Monthly
Notices of the Royal Astronomical Society. 398(4), 1601–1614.
\doi{10.1111/j.1365-2966.2009.14548.x}

For implementation, see:
https://github.com/kbarbary/nestle/blob/master/runtests.py
}
\seealso{
Other ernest_lrps: 
\code{\link{mini_balls}()},
\code{\link{no_underrun}()},
\code{\link{rwmh_cube}()},
\code{\link{slice_rectangle}()},
\code{\link{unif_cube}()},
\code{\link{unif_ellipsoid}()}
}
\concept{ernest_lrps}
