% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_sampling.R
\name{nested_sampling}
\alias{nested_sampling}
\title{Prepare a nested sampler to evaluate Bayesian evidence}
\usage{
nested_sampling(
  log_lik,
  prior,
  sampler = rwmh_cube(),
  n_points = 500,
  first_update = NULL,
  update_interval = NULL,
  on_warning = c("abort", "warn"),
  ...
)
}
\arguments{
\item{log_lik}{(function) A function that takes a vector of parameters and
returns the log-likelihood of the parameters. The function should return a
finite value or \code{-Inf} for invalid parameters (see \code{\link[=create_likelihood]{create_likelihood()}}).}

\item{prior}{(ernest_prior) An object with class \code{ernest_prior}, created by
\code{\link[=create_prior]{create_prior()}} or its specializations.}

\item{sampler}{(ernest_sampling) An \link{ernest_sampling} object, declaring which
likelihood-restricted prior sampler (LRPS) to use.}

\item{n_points}{(positive integer) The number of live points to use in the
nested sampling run.}

\item{first_update}{(optional positive integer) The number of likelihood
calls before adopting LRPS behaviour. If \code{NULL}, this is set to
\code{n_points * 2.5}.}

\item{update_interval}{(optional positive integer) The number of likelihood
calls between updates to the LRPS behaviour. If \code{NULL}, this is set to
\code{n_points * 1.5}.}

\item{on_warning}{(case-sensitive string) Action to perform when the sanity
test throws a warning message (see Details).
\itemize{
\item \code{"abort"}: Throw an error.
\item \code{"warn"}: Issue a warning and initialize the sampler.
}}

\item{...}{
  Arguments passed on to \code{\link[=create_likelihood.function]{create_likelihood.function}}
  \describe{
    \item{\code{error_action}}{(case-sensitive string) Action to perform once \code{fn}
throws an error.
\itemize{
\item \code{"abort"}: Stop execution and signal an error.
\item \code{"warn"}: Issue a warning and replace output with \code{-Inf}.
}}
    \item{\code{nonfinite_action}}{(case-sensitive string) Action to perform when \code{fn}
passes a value that is not a finite double nor \code{-Inf}.
\itemize{
\item \code{"warn"}: Issue a warning and replace values with \code{-Inf}.
\item \code{"pass"}: Silently replace values with \code{-Inf}.
\item \code{"abort"}: Stop execution and signal an error.
}}
    \item{\code{auto_batch}}{(logical) Whether to prepare \code{fn} so that it may be called
with a matrix of parameter values. If \code{FALSE}, its assumed that \code{fn}
can already produce a vector of likelihood values for a matrix with rows of
parameter vectors.}
  }}
}
\value{
An \link{ernest_sampler} object, prepared for nested sampling.
}
\description{
Initialize a \link{ernest_sampler} object to perform nested sampling with a given
log-likelihood function, prior distribution, and likelihood-restricted prior
specification.
}
\details{
This function prepares an \code{ernest_sampler} object for performing nested
sampling, given a log-likelihood function, prior, and sampler specification.
It checks that the provided log-likelihood and prior are compatible by
performing a basic sanity check by evaluating the log-likelihood at the
center of the prior space. This check ensures that the log-likelihood
returns a finite value or \code{-Inf} for typical parameter values.

The sanity check may fail even if \code{log_lik} is well-specified, for example
if the prior's support does not include the test point
(the center of the prior hypercube), or if the log-likelihood is only
defined for a subset of the parameter space. In such cases, silencing the
warning from \verb{nested_sampling`` may require you to adjust the prior or adjust the }on_warning` argument.
}
\examples{
prior <- create_uniform_prior(n_dim = 2, lower = -1, upper = 1)
ll_fn <- function(x) -sum(x^2)
sampler <- nested_sampling(ll_fn, prior, n_points = 100)
sampler
}
