% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_sampling.R
\name{nested_sampling}
\alias{nested_sampling}
\title{Build a Nested Sampler}
\usage{
nested_sampling(
  log_lik,
  prior,
  sampler = rwmh_cube(),
  n_points = 500,
  first_update = 2.5,
  update_interval = 1.5
)
}
\arguments{
\item{log_lik}{The log-likelihood function. This should take in a vector of
parameters and return a scalar numeric value that is either finite or \code{-Inf}.}

\item{prior}{An \code{ernest_prior} object, created by \code{\link[=create_prior]{create_prior()}} or
its specializations.}

\item{sampler}{An \link{ernest_sampling} object, declaring which likelihood-restricted
prior sampler to use for the nested sampling run.}

\item{n_points}{The number of live points to use in the nested sampling run.}

\item{first_update}{The number of calls to the likelihood function
before the first update of the sampler. If left as a double, this will be
multiplied by the number of live points before being coerced to an
integer.}

\item{update_interval}{The number of calls to the likelihood function
between subsequent updates of the sampler. If left as a double, this will
be multiplied by the number of live points before being coerced to an
integer.}
}
\value{
An \code{ernest_sampler} object.
}
\description{
Constructs an instance of an \code{ernest_sampler} containing the necessary
information for performing nested sampling.
}
