% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_sampling.R
\name{nested_sampling}
\alias{nested_sampling}
\title{Prepare a nested sampler to evaluate Bayesian evidence}
\usage{
nested_sampling(
  log_lik,
  prior,
  sampler = rwmh_cube(),
  n_points = 500,
  first_update = NULL,
  update_interval = NULL,
  ...
)
}
\arguments{
\item{log_lik}{Function, defining the log-likelihood of the model given
a vector of parameters. Passed to \code{\link[=create_likelihood]{create_likelihood()}}.}

\item{prior}{An \link{ernest_prior} object, created by \code{\link[=create_prior]{create_prior()}} or
its specializations.}

\item{sampler}{An \link{ernest_sampling} object, declaring which
likelihood-restricted prior sampler (LRPS) to use. One of \code{\link[=rwmh_cube]{rwmh_cube()}}
(recommended) or \code{\link[=unif_cube]{unif_cube()}}.}

\item{n_points}{Integer. The number of live points to use in the nested
sampling run.}

\item{first_update}{Integer or \code{NULL}. The number of likelihood calls before
the first update of the LRPS. If \code{NULL}, this is set to \code{n_points * 2.5}.}

\item{update_interval}{Integer or \code{NULL}. The number of likelihood calls
between subsequent updates of the LRPS. If \code{NULL}, this is set to
\code{n_points * 1.5}.}

\item{...}{Additional arguments passed to \code{\link[=create_likelihood]{create_likelihood()}}.}
}
\value{
An \link{ernest_sampler} object, prepared for nested sampling.
}
\description{
Initialize a \link{ernest_sampler} object to perform nested sampling with a given
log-likelihood function, prior distribution, and likelihood-restricted prior
specification.
}
\examples{
prior <- create_uniform_prior(n_dim = 2, lower = -1, upper = 1)
ll_fn <- function(x) -sum(x^2)
sampler <- nested_sampling(ll_fn, prior, n_points = 100)
sampler
}
