% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_sampling.R
\name{nested_sampling}
\alias{nested_sampling}
\alias{nested_sampling.function}
\title{Perform Nested Sampling}
\usage{
nested_sampling(x, ...)

\method{nested_sampling}{`function`}(
  x,
  prior_transform,
  sampler = unit_cube(),
  control = list(),
  ...
)
}
\arguments{
\item{x}{A function returning the log likelihood of a model given some parameters.}

\item{...}{Extra parameters that are passed to the log likelihood function.}

\item{prior_transform}{The prior distribution for the parameters, provided as
a list of \code{prior_transform} objects.}

\item{sampler}{The likelihood-restricted prior sampler to use.}

\item{control}{A list of control parameters for the nested sampling algorithm,
whose defaults are specified in \code{nested_sampling.control()}.}
}
\value{
Lots of stuff about the run.
}
\description{
Perform Nested Sampling
}
