% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_sampling.R
\name{nested_sampling}
\alias{nested_sampling}
\alias{nested_sampling.function}
\title{Perform Nested Sampling}
\usage{
nested_sampling(...)

\method{nested_sampling}{`function`}(fn, prior_transform, sampler = unit_cube(), ...)
}
\arguments{
\item{...}{A list of control parameters for the nested sampling algorithm,
whose defaults are specified in \code{nested_sampling.control()}.}

\item{fn}{A function returning the log likelihood of a model given some parameters.}

\item{prior_transform}{The prior distribution for the parameters, provided as
a list of \code{prior_transform} objects.}

\item{sampler}{The likelihood-restricted prior sampler to use.}
}
\value{
Lots of stuff about the run.
}
\description{
Perform Nested Sampling
}
