% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_sampling.R
\name{nested_sampling}
\alias{nested_sampling}
\alias{nested_sampling.function}
\alias{nested_sampling.glm}
\title{Build a Nested Sampler}
\usage{
nested_sampling(x, ...)

\method{nested_sampling}{`function`}(
  x,
  prior,
  names = NULL,
  sampler = rwmh_cube(),
  n_points = 500,
  first_update = 2.5,
  update_interval = 1.5,
  ...
)

\method{nested_sampling}{glm}(
  x,
  prior = default_prior(x),
  names = NULL,
  sampler = rwmh_cube(),
  n_points = 500,
  first_update = 2.5,
  update_interval = 1.5,
  ...
)
}
\description{
The top-level function constructs an instance of an \code{ernest_sampling},
containing the necessary information for nested sampling.
}
