% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_sampling.R
\name{nested_sampling}
\alias{nested_sampling}
\title{Prepare a nested sampler to evaluate Bayesian evidence}
\usage{
nested_sampling(
  log_lik,
  prior,
  sampler = rwmh_cube(),
  n_points = 500,
  first_update = NULL,
  update_interval = NULL,
  ...
)
}
\arguments{
\item{log_lik}{(function) A function or \code{ernest_likelihood} that takes a
vector of parameters and returns the log-likelihood. The function should
return a finite value or \code{-Inf} for invalid parameters
(see \code{\link[=create_likelihood]{create_likelihood()}}).}

\item{prior}{(ernest_prior) An object with class \code{ernest_prior}, created by
\code{\link[=create_prior]{create_prior()}} or its specializations.}

\item{sampler}{(ernest_sampling) An \link{ernest_sampling} object, declaring which
likelihood-restricted prior sampler (LRPS) to use.}

\item{n_points}{(positive integer) The number of live points to use in the
nested sampling run.}

\item{first_update}{(optional positive integer) The number of likelihood
calls before adopting LRPS behaviour. If \code{NULL}, this is set to
\code{n_points * 2.5}.}

\item{update_interval}{(optional positive integer) The number of likelihood
calls between updates to the LRPS behaviour. If \code{NULL}, this is set to
\code{n_points * 1.5}.}

\item{...}{
  Arguments passed on to \code{\link[=create_likelihood.function]{create_likelihood.function}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
An \link{ernest_sampler} object, prepared for nested sampling.
}
\description{
Initialize a \link{ernest_sampler} object to perform nested sampling with a given
log-likelihood function, prior distribution, and likelihood-restricted prior
specification.
}
\details{
This function prepares an \code{ernest_sampler} object for performing nested
sampling, given a log-likelihood function, prior, and sampler specification.
It checks that the provided log-likelihood and prior are compatible by
performing a basic sanity check by evaluating the log-likelihood at the
center of the prior space. This check ensures that the log-likelihood
returns a finite value or \code{-Inf} for typical parameter values.
}
\examples{
prior <- create_uniform_prior(n_dim = 2, lower = -1, upper = 1)
ll_fn <- function(x) -sum(x^2)
sampler <- nested_sampling(ll_fn, prior, n_points = 100)
sampler
}
