% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps-class.R
\name{new_ernest_lrps}
\alias{new_ernest_lrps}
\alias{ernest_lrps}
\title{Create a new likelihood-restricted prior sampler (LRPS)}
\usage{
new_ernest_lrps(
  unit_log_fn = NULL,
  nvar = NULL,
  max_loop = getOption("ernest.max_loop", 1000000L),
  cache = NULL,
  ...,
  .class = NULL,
  .call = caller_env()
)
}
\arguments{
\item{unit_log_fn}{\verb{[function]}\cr Takes a matrix of points in the unit
cube and returns a numeric vector of log-likelihood values. Optional, and
updated when called by \code{ernest_sampler()}.}

\item{nvar}{\verb{[integer(1)]}\cr Number of dimensions within the prior space.
Optional, and updated when called by \code{ernest_sampler()}.}

\item{max_loop}{\verb{[integer(1)]}\cr Maximum number of attempts to generate
points. Inferred from the \code{ernest.max_loop} option.}

\item{cache}{\verb{[environment]} Environment for caching information required for
the specific LRPS method. Created if left \code{NULL}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name-value pairs for
additional elements for \verb{ernest_lrps'} subclasses}

\item{.class}{\verb{[character()]}\cr Names for the subclass of \code{ernest_lrps}.}

\item{.call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
\verb{[ernest_lrps]}, a named list containing the arguments provided.
}
\description{
Constructs an LRPS, accepted by \code{\link[=ernest_sampler]{ernest_sampler()}} and used to create new
points during a nested sampling run.
}
\details{
Nested sampling relies on generating a series of points in the prior space
with increasing log-likelihood values. This is accomplished using a
likelihood-restricted prior sampler (LRPS), which generates independent and
identically distributed points from the prior, subject to a hard likelihood
constraint.

To create your own LRPS, subclass \code{new_ernest_lrps} and provide S3 methods
for \code{\link[=propose]{propose()}} and \code{\link[=update_lrps]{update_lrps()}} for your subclass.
}
\keyword{internal}
