% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps-class.R
\name{new_ernest_lrps}
\alias{new_ernest_lrps}
\alias{ernest_lrps}
\title{Create a new likelihood-restricted prior sampler (LRPS)}
\usage{
new_ernest_lrps(
  unit_log_fn = NULL,
  n_dim = NULL,
  max_loop = getOption("ernest.max_loop", 1000000L),
  cache = NULL,
  ...,
  .class = NULL,
  .call = caller_env()
)
}
\arguments{
\item{unit_log_fn, n_dim}{Provided when \code{\link[=ernest_sampler]{ernest_sampler()}} is called with a
given \code{ernest_lrps}:
\itemize{
\item \code{unit_log_fn} (function, optional): Takes a matrix of points in the unit
cube and returns a numeric vector of log-likelihood values.
\item \code{n_dim} (integer, optional): Number of dimensions of the prior space.
}}

\item{max_loop}{Positive integer. Maximum number of attempts to generate
points via region-based sampling. Usually hidden from users, but can be set
via the \code{ernest.max_loop} option.}

\item{cache}{(environment, optional) Environment for caching values. If
\code{NULL}, a new environment is created.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name-value pairs for
additional elements for subclasses of this LRPS.}

\item{.class}{(character vector, optional) Subclasses of this LRPS.}

\item{.call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An LRPS specification: a list containing the input arguments, with a
class specific to the LRPS type.
}
\description{
Nested sampling relies on generating a series of points in the prior space
with increasing log-likelihood values. This is accomplished using a
likelihood-restricted prior sampler (LRPS), which generates independent and
identically distributed points from the prior, subject to a hard likelihood
constraint.

To create your own LRPS, subclass \code{new_ernest_lrps} and provide S3 methods
for \code{\link[=propose]{propose()}} and \code{\link[=update_lrps]{update_lrps()}} for your subclass.
}
\keyword{internal}
