% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps-class.R
\name{new_ernest_lrps}
\alias{new_ernest_lrps}
\alias{ernest_lrps}
\title{Create a new likelihood-restricted prior sampler (LRPS)}
\usage{
new_ernest_lrps(
  unit_log_fn = NULL,
  n_dim = NULL,
  max_loop = getOption("ernest.max_loop", 1000000L),
  cache = NULL,
  ...,
  .class = NULL,
  .call = caller_env()
)
}
\arguments{
\item{unit_log_fn, ndim}{These values are provided when \code{\link[=ernest_sampler]{ernest_sampler()}} is
called with a given \code{ernest_lrps}:
\itemize{
\item \code{unit_log_fn} (function, optional) A function that takes a matrix of
points in the unit cube and returns a numeric vector of log-likelihood
values.
\item \code{n_dim} (integer, optional) The number of dimensions of the prior space.
}}

\item{max_loop}{(strictly positive integer) The maximum number of attempts
to generate points through region-based sampling methods. Normally hidden
from users, but can be set through the option \code{ernest.max_loop}.}

\item{cache}{(environment, optional) An environment to use for caching
values. If \code{NULL}, a new environment is created.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name-value pairs for
additional elements of samplers that subclass this LRPS.}

\item{.class}{(character vector, optional) The subclasses of this LRPS.}

\item{.call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An LRPS specification, which is a list containing the inputs used
as arguments to the function, along with a class specific to the type
of LRPS being created.
}
\description{
Nested sampling relies on generating a series of points in the prior space
with increasing log-likelihood values. This is accomplished through using an
likelihood-restricted prior sampler (LRPS), which provides a technique to
generate identically and independently distributed points from the prior
space, subject to a hard likelihood constraint.

If you want to create your own LRPS, you may do so by subclassing
\code{new_ernest_lrps}. You will also need to provide S3 methods for \code{\link[=propose]{propose()}}
and \code{\link[=update_lrps]{update_lrps()}} for your subclass.
}
\keyword{internal}
