% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{parallel_likelihood}
\alias{parallel_likelihood}
\alias{parallel_prior}
\title{Parallelization in ernest}
\usage{
parallel_likelihood(
  scalar_fn,
  vectorized_fn,
  ...,
  .on_nonfinite = c("warn", "quiet", "abort")
)

parallel_prior(
  point_fn,
  vectorized_fn,
  ...,
  .names,
  .lower = -Inf,
  .upper = Inf,
  .repair = c("unique", "universal", "check_unique", "unique_quiet", "universal_quiet")
)
}
\arguments{
\item{scalar_fn, point_fn, vectorized_fn}{\verb{[function]}\cr Used identically to
\code{\link[=create_likelihood]{create_likelihood()}} or \code{\link[=create_prior]{create_prior()}}, but with the additional
requirement that the functions are 'fresh', meaning they should be declared
in the call to \code{parallel_likelihood()} or \code{parallel_prior()} rather than
stored in a variable.}

\item{...}{Named arguments that should be captured in the function's
environment and serialized for remote execution.}

\item{.on_nonfinite}{\verb{[character]}\cr How the sampler should handle values
returned by \code{fn} or \code{matrix_fn} that are not finite and not equal to \code{-Inf}.
Must be one of:
\itemize{
\item \code{"warn"}: Issue a warning and return \code{-Inf}.
\item \code{"quiet"}: Silently return \code{-Inf}.
\item \code{"abort"}: Stop execution and signal an error.
}}

\item{.names}{\verb{[character()]}\cr Unique names for each variable in the prior
distribution.}

\item{.lower, .upper}{\verb{[double()]}\cr Expected lower and upper bounds for the
parameter vectors after transformation.}

\item{.repair}{\verb{[character(1)]}\cr Name repair strategy for \code{names}. One of
\code{"check_unique"}, \code{"unique"}, \code{"universal"}, \code{"unique_quiet"}, or
\code{"universal_quiet"}. See \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} for details.}
}
\value{
For \code{parallel_likelihood()}, a \link[carrier:crate]{carrier::crate} with the additional class
\link{ernest_likelihood}.

For \code{parallel_prior()}, an \link{ernest_prior} whose transformation function is
a \link[carrier:crate]{carrier::crate}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Nested sampling runs can be performed in parallel using \CRANpkg{mirai}.

To parallelize a run:
\itemize{
\item Create a portable likelihood with \code{parallel_likelihood()}.
\item Create a portable prior with \code{parallel_prior()} or use one of ernest's
\link[=special_priors]{special priors}.
\item Set worker daemons with \code{mirai::daemons()} and ensure workers can load
ernest.
\item Call \code{\link[=generate-ernest]{generate()}} with \code{parallel} != \code{FALSE}.
}
}
\section{Parallelized nested sampling}{

Splitting an initial live set into subruns lets you use many live points
without running a single very large job. Run several smaller nested sampling
jobs concurrently and merge their records to obtain the same statistical
benefits as a single large run.

When \code{\link[=generate-ernest]{generate()}} is called with \code{parallel != FALSE}, the
resulting run contains a \code{.parallel} element with per-worker summaries (see
\code{\link[=glance.ernest_run]{glance.ernest_run()}}). To protect runs from over-parallelization (which
contributes to bias in likelihood-restricted prior sampling), the number of
live points per worker is checked against the number of variables in the
parameter space. If any worker has fewer than
\code{getOption("ernest.min_nlive_nvar", 10L)} live points per variable, a warning
is issued and the number of workers is automatically reduced.

Additionally, ernest will warn the user if the parallelization resulted in
subruns with suspiciously low effective sample size; this is evidence that
paralleilzation introduced dependence between the subruns, which may bias
estimation. This warning is configured with
\code{getOption("ernest.min_ess", 400L)}.
}

\section{Daemons}{

How parallelization occurs is determined by \code{\link[mirai:daemons]{mirai::daemons()}}. Daemons must
be set prior to calling \code{generate()} with \code{parallel != FALSE}, otherwise an
error will be thrown.

It is usual to set daemons once per session. You can leave them running on
your local machine as they consume almost no resources whilst waiting to
receive tasks. The following sets up 6 daemons locally:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mirai::daemons(6)
}\if{html}{\out{</div>}}
}

\references{
Documentation adapted from the \CRANpkg{mirai},
\CRANpkg{carrier}, and \CRANpkg{purrr} packages.
}
\seealso{
\code{\link[mirai:daemons]{mirai::daemons()}} \code{\link[carrier:crate]{carrier::crate()}} \link{generate-ernest}
}
