% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ernest_estimates}
\alias{plot.ernest_estimates}
\alias{plot.ernest_run}
\title{Plot the Integration Progress From a Nested Sampling Run}
\usage{
\method{plot}{ernest_estimates}(x, ...)

\method{plot}{ernest_run}(x, ndraws = 0, ...)
}
\arguments{
\item{x}{An object of class 'ernest_run' or 'ernest_estimates'.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ndraws}{For \code{plot.ernest_run}, the number of draws to use for
estimating the evidence integral. If 0, the original results are used.}
}
\description{
Using an \code{ernest_estimates} or an \code{ernest_run} object, \code{plot} creates
a faceted \code{ggplot()} showing iterative integration results over estimates
of the log prior volume. These plots show the likelihood values (normalized),
the posterior weights, and the model evidence. The latter quantity is
accompanied with an error ribbon capturing uncertainty.
}
\details{
Calling \code{plot.ernest_run(run, ndraws = n)} is identical to calling
\code{plot.ernest_estimates(calculate(run, ndraws = n))}.
}
