% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ernest_estimates}
\alias{plot.ernest_estimates}
\alias{plot.ernest_run}
\title{Plot diagnostics for a nested sampling run}
\usage{
\method{plot}{ernest_estimates}(x, ...)

\method{plot}{ernest_run}(x, ..., ndraws = 0)
}
\arguments{
\item{x}{Either an \code{ernest_estimates} (from \code{\link[=calculate]{calculate()}}) or an
\code{ernest_run} object.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ndraws}{Number of simulated log. volume values to generate before plotting.
If \code{ndraws = 0}, plot the log. volume estimates generated during the run.}
}
\value{
\code{x}, invisibly. A \code{ggplot2::ggplot()} object is printed as a
side effect.

If \code{x} is an \code{ernest_estimates} object, or if \code{ndraws} is greater than one,
the plot show \code{ndraw} simulated log. volume values at each iteration. Else,
the plot will use the log. volume estimates that were recorded during the
run.
}
\description{
Show the normalized likelihood, importance weights, and evidence as functions
of log. volume.
}
\note{
Plotting simulated results requires the \code{ggdist} package.
}
\examples{
# Plot integration results from a run.
data(ernest_run_example)
plot(ernest_run_example)

# Simulate results before plotting.
library(ggdist)
data(ernest_run_example)
plot(ernest_run_example, ndraws = 50)
# Simulate results from a run, then plot simulated results.
library(ggdist)
data(ernest_run_example)

sim <- calculate(ernest_run_example, ndraws = 50)
plot(sim)
}
\seealso{
\code{\link[=calculate]{calculate()}} for generating \code{ernest_estimates} objects; also, \code{\link[=visualize]{visualize()}} to
plot the posterior distributions of the parameters from a run.
}
