% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ernest_estimate}
\alias{plot.ernest_estimate}
\alias{plot.ernest_run}
\title{Plot diagnostics for a nested sampling run}
\usage{
\method{plot}{ernest_estimate}(x, ...)

\method{plot}{ernest_run}(x, ..., ndraws = 0)
}
\arguments{
\item{x}{An \link{ernest_estimate} or \link{ernest_run} object.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ndraws}{A positive integer. The number of log-volume
sequences to simulate. If equal to zero, no simulations will be made, and a
one draw vector of log-volumes are produced from the estimates contained in
\code{x}.}
}
\value{
Invisibly returns \code{x}. A \code{ggplot2::ggplot()} object is printed as a
side effect.

The plot is faceted into three panels. The horizontal axis shows log-volume:
If \code{x} is an \code{ernest_run}, these estimates are derived from the run;
if \code{x} is an \code{ernest_estimate} (or \code{ndraws != 0}), these values are
simulated.

The three \code{y} axes are:
\itemize{
\item \strong{Evidence:} Estimate with an error ribbon drawn from either
the estimated standard error (if \code{ernest_run}) or from the median credible
interval (MCI) (if \code{ernest_estimate});
\item \strong{Normalised Likelihood:} The likelihood value of the criteria used
to draw new points from the likelihood-restricted prior sampler, normalised
by the maximum likelihood generated during the run;
\item \strong{Posterior Weight:} The density of the posterior weights attributed to
regions of volume within the prior. For \code{ernest_estimate} objects,
an error ribbon is drawn with the MCI of this estimate.
}
}
\description{
Shows the normalised likelihood, importance weights, and evidence as
functions of log-volume.
}
\note{
For \code{ernest_estimate}, \code{ndraws} should be large enough to plot the
}
\examples{
# Plot integration results from a run.
data(example_run)
plot(example_run)

# Simulate results before plotting.
plot(example_run, ndraws = 100)
# Simulate results from a run, then plot simulated results.
sim <- calculate(example_run, ndraws = 100)
plot(sim)
}
\seealso{
\itemize{
\item \code{\link[=calculate]{calculate()}} for generating \code{ernest_estimate} objects.
\item \code{\link[=visualize]{visualize()}} for plotting the posterior distributions generated by a run.
}
}
