% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ernest_estimate}
\alias{plot.ernest_estimate}
\alias{plot.ernest_run}
\title{Plot diagnostics for a nested sampling run}
\usage{
\method{plot}{ernest_estimate}(x, ...)

\method{plot}{ernest_run}(x, ..., ndraws = 0)
}
\arguments{
\item{x}{(ernest_estimate or ernest_run) An object either of class
\code{ernest_estimate} (containing uncertainty simulations) or of class
\code{ernest_run} (containing a nested sampling run).}

\item{...}{These dots are for future extensions and must be empty.}

\item{ndraws}{(positive integer or zero, optional) The number of log volume
sequences to simulate. If equal to zero, no simulations will be made, and a
one draw vector of log volumes are produced from the estimates contained in
\code{x}. If \code{NULL}, \code{getOption("posterior.rvar_ndraws")} is used (default 4000).}
}
\value{
\code{x}, invisibly. A \code{ggplot2::ggplot()} object is printed as a
side effect.

The plot is faceted in three frames. The horizontal axis shows values of log.
volume: If \code{x} is an \code{ernest_run} these estimates are derived from the run,
if \code{x} is an \code{ernest_estimate} (or \code{ndraws != 0}), these values are
simulated.

The three \code{y} axes are as follows:
\itemize{
\item \strong{Evidence:} Estimate with a corresponding error ribbon drawn from either
the estimated standard error (if \code{ernest_run}) or drawn from the high density
credible interval (HDCI) (if \code{ernest_sampler});
\item \strong{Normalized Likelihood:} The likelihood value of the criteria used
to draw new points from the likelihood-restricted prior sampler, normalized
by the maximum likelihood generated during the run, and;
\item \strong{Posterior Weight}: The density of the posterior weights attributed to
regions of volume within the prior. If using an \code{ernest_estimate} object,
an error ribbon is drawn with the HDCI of this estimate.
}
}
\description{
Show the normalized likelihood, importance weights, and evidence as functions
of log. volume.
}
\note{
For \code{ernest_estimate}, \code{ndraws} must be sufficiently large to calculate
HDI. If HDI calculation fails, the user will be warned and an \code{ernest_run}
plot will be generated instead.
}
\examples{
# Plot integration results from a run.
data(example_run)
plot(example_run)

# Simulate results before plotting.
plot(example_run, ndraws = 50)
# Simulate results from a run, then plot simulated results.
sim <- calculate(example_run, ndraws = 50)
plot(sim)
}
\seealso{
\code{\link[=calculate]{calculate()}} for generating \code{ernest_estimate} objects; also,
\code{\link[=visualize]{visualize()}} to plot the posterior distributions of the parameters from a
run.
}
