% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ernest_estimate}
\alias{plot.ernest_estimate}
\alias{plot.ernest_run}
\title{Plot diagnostics for a nested sampling run}
\usage{
\method{plot}{ernest_estimate}(x, which = c("evidence", "weight", "likelihood"), ...)

\method{plot}{ernest_run}(x, which = c("evidence", "weight", "likelihood"), ndraws = NULL, ...)
}
\arguments{
\item{x}{[\link{ernest_run}] or [\link{ernest_estimate}]\cr An object containing
results from nested sampling.}

\item{which}{\verb{[character()]}\cr Choose which plots to display. Must be one
or more of \code{"evidence"}, \code{"weight"}, and \code{"likelihood"}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ndraws}{\verb{[integer(1)]}\cr The number of log-volume sequences to
simulate.
If equal to zero, no simulations will be made, and a one draw vector of
log-volumes are produced from the estimates contained in \code{x}.}
}
\value{
\verb{[invisible(x)]}. A \code{ggplot2::ggplot()} object is printed as a
side effect.
}
\description{
Visualizes key diagnostics from a nested sampling run, including the
normalized likelihood, posterior weights, and evidence as functions
of log-prior volume.
}
\details{
Interpreting these plots can help diagnose issues such as poor or
insufficient prior sampling and model misspecification. Use \code{which} to select
the plots to display:
\itemize{
\item \code{which = "evidence"}: Plots the estimated marginal
likelihood (evidence) as a function of log-prior volume, with uncertainty
intervals. Peaks in this plot indicate regions of prior volume that
contribute most to the evidence estimate.
\item \code{which = "weight"}: Shows the distribution of
posterior mass across log-prior volume. This plot highlights which regions
of the prior volume contain the most posterior probability, helping to
identify where the sampler concentrated its effort.
\item \code{which = "likelihood"}: Displays the normalized
likelihood as a function of log-prior volume. Smoothness in this plot
indicates effective likelihood-restricted prior sampling, while
irregularities may suggest sampling difficulties or, in some cases,
misspecified likelihood functions.
}

If \code{x} is an \code{ernest_run}, the plots are based on the actual run data. Error
ribbons are drawn around the \code{evidence} plot from analytic estimates of
uncertainty (see {summary.ernest_run}).

If \code{x} is an \code{ernest_estimate} (or if \code{ndraws} is specified), the plots
are based on simulated values from the log-volume. The highest density
continuous intervals (HDCIs) are computed using \code{\link[ggdist:point_interval]{ggdist::median_hdci()}} for
both the \code{evidence} and \code{weight} plots.
}
\note{
Plotting multiple diagnostics with \code{which} requires the patchwork package.
Plotting \code{ernest_estimate} objects requires the ggdist package.
}
\examples{
# Plot results from a run.
data(example_run)
plot(example_run)

# Simulate results before plotting.
set.seed(42)
plot(example_run, ndraws = 100)

# Equivalently, simulate results, then plot:
set.seed(123)
est <- calculate(example_run, ndraws = 100)
plot(est)
}
\seealso{
\itemize{
\item \code{\link[=calculate]{calculate()}} for generating \code{ernest_estimate} objects.
\item \code{\link[=visualize]{visualize()}} for plotting the posterior distributions generated by a run.
}
}
