% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ernest_run}
\alias{print.ernest_run}
\title{Print a summary of the results}
\usage{
\method{print}{ernest_run}(run, digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{run}{An \code{ernest} run}

\item{digits}{The number of digits to output}
}
\value{
The run object, invisibly.
}
\description{
Print a summary of the results
}
