% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps-class.R, R/rwmh_cube.R, R/unif_cube.R
\name{propose}
\alias{propose}
\alias{propose.rwmh_cube}
\alias{propose.unif_cube}
\title{Generate a new point using LRPS}
\usage{
propose(x, original = NULL, criteria = NULL, ...)

\method{propose}{rwmh_cube}(x, original = NULL, criteria = NULL, ...)

\method{propose}{unif_cube}(x, original = NULL, criteria = NULL, ...)
}
\arguments{
\item{x}{(ernest_lrps) An \code{ernest_lrps} object.}

\item{original}{(double vector or matrix, optional) Points within the prior
space that are used to start the proposal process. If \code{NULL}, a new point is
generated by sampling from the unconstrained uniform unit cube as in
\code{\link[=unif_cube]{unif_cube()}}.}

\item{criteria}{(double vector) A vector of log-likelihood values
that the proposed points must satisfy.}

\item{...}{Additional arguments passed to subclass methods.}
}
\value{
Calls to this generic should always return a list with the following
elements:
\itemize{
\item \code{unit}: A matrix of proposed points in the prior space.
\item \code{log_lik}: A numeric vector of log-likelihood values corresponding to the
proposed points.
\item \code{n_call}: The number of calls made to the \code{unit_log_fn} during the proposal
process.
}
}
\description{
A developer facing function, only used if you are creating your own
\link{ernest_lrps} subclass.

When specifying your subclass, you must subclass this
this method to specify how your sampler generates new points that satisfy
the likelihood constraint.
}
\keyword{internal}
