% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps-class.R, R/mini_balls.R,
%   R/multi_ellipsoid.R, R/no_underrun.R, R/rwmh_cube.R, R/slice_rectangle.R,
%   R/unif_cube.R, R/unif_ellipsoid.R
\name{propose}
\alias{propose}
\alias{propose.mini_balls}
\alias{propose.multi_ellipsoid}
\alias{propose.no_underrun}
\alias{propose.rwmh_cube}
\alias{propose.slice_rectangle}
\alias{propose.unif_cube}
\alias{propose.unif_ellipsoid}
\title{Generate a new point using LRPS}
\usage{
propose(x, original = NULL, criterion = -Inf)

\method{propose}{mini_balls}(x, original = NULL, criterion = -Inf)

\method{propose}{multi_ellipsoid}(x, original = NULL, criterion = -Inf)

\method{propose}{no_underrun}(x, original = NULL, criterion = -Inf)

\method{propose}{rwmh_cube}(x, original = NULL, criterion = -Inf)

\method{propose}{slice_rectangle}(x, original = NULL, criterion = -Inf)

\method{propose}{unif_cube}(x, original = NULL, criterion = -Inf)

\method{propose}{unif_ellipsoid}(x, original = NULL, criterion = -Inf)
}
\arguments{
\item{x}{An \code{ernest_lrps} object.}

\item{original}{Optional double vector. Points in the prior space used to
start the proposal process. If \code{NULL}, a new point is generated by
sampling from the unconstrained unit cube (see \code{\link[=unif_cube]{unif_cube()}}).}

\item{criterion}{Double scalar. A log-likelihood value that proposed points
must satisfy.}
}
\value{
A list with:
\itemize{
\item \code{unit}: Matrix of proposed points in the prior space.
\item \code{log_lik}: Numeric vector of log-likelihood values for the proposed
points.
\item \code{n_call}: Number of calls made to \code{unit_log_fn} during the proposal.
}
}
\description{
Developer-facing function, used when creating your own \link{ernest_lrps}
subclass.

When specifying your subclass, you must implement this method to define how
your sampler generates new points that satisfy the likelihood constraint.
}
\keyword{internal}
