% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{generate}
\alias{compile}
\alias{calculate}
\alias{glance}
\alias{as_draws}
\alias{as_draws_matrix}
\alias{as_draws_df}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{calculate}}, \code{\link[generics]{compile}}, \code{\link[generics]{generate}}, \code{\link[generics]{glance}}}

  \item{posterior}{\code{\link[posterior:draws]{as_draws}}, \code{\link[posterior:draws_df]{as_draws_df}}, \code{\link[posterior:draws_matrix]{as_draws_matrix}}}
}}

