% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_draws.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_draws_matrix}
\alias{as_draws_rvars}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{posterior}{\code{\link[posterior:draws_matrix]{as_draws_matrix}}, \code{\link[posterior:draws_rvars]{as_draws_rvars}}}
}}

