% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-generics.R, R/import-posterior.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{calculate}
\alias{compile}
\alias{generate}
\alias{visualize}
\alias{as_draws}
\alias{as_draws_matrix}
\alias{as_draws_rvars}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{calculate}}, \code{\link[generics]{compile}}, \code{\link[generics]{generate}}, \code{\link[generics]{visualize}}}

  \item{posterior}{\code{\link[posterior:draws]{as_draws}}, \code{\link[posterior:draws_matrix]{as_draws_matrix}}, \code{\link[posterior:draws_rvars]{as_draws_rvars}}}
}}

