% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_rectangle.R
\name{slice_rectangle}
\alias{slice_rectangle}
\title{Slice sampling within a bounding hyperrectangle}
\usage{
slice_rectangle(steps = 3L, adaptive = FALSE, max_steps = 100)
}
\arguments{
\item{steps}{\verb{[integer(1)]} Number of reslicing iterations per proposal.}

\item{adaptive}{\verb{[logical(1)]} If TRUE, adapt \code{steps} using the mean
Mahalanobis distance among live points.}

\item{max_steps}{\verb{[integer(1)]} Maximum \code{steps} allowed when
\code{adaptive = TRUE}.}
}
\value{
\verb{[slice_rectangle]} A list inheriting from \code{ernest_lrps}.
}
\description{
Propose new live points by evolving an existing live point using slice
sampling inside a bounding hyperrectangle. The rectangle is shrunk when a
proposal is rejected.
}
\details{
Sampling starts from a live point that meets the likelihood criterion. Each
iteration draws a uniform proposal inside the rectangle and accepts it if it
satisfies the criterion. Rejected proposals cause the rectangle to be
shrunk so the rejected point lies on the boundary and the original point
remains inside. The process repeats until a valid proposal is found or the
rectangle can no longer be reduced.

Use \code{steps > 1} to increase exploration. When \code{adaptive = TRUE}, \code{steps} is
increased or decreased depending on whether proposed moves are closer or
farther than the mean Mahalanobis distance of the live set.
}
\examples{
lrps <- slice_rectangle()
adaptive_lrps <- slice_rectangle(max_steps = 30, adaptive = TRUE)

}
\references{
Neal, R. M. (2000). Slice Sampling (Version 1). arXiv.
\doi{10.48550/ARXIV.PHYSICS/0009028}
Buchner, J. (2021). UltraNest - A Robust, General Purpose Bayesian Inference
Engine. Journal of Open Source Software, 6(60), 3001.
\doi{10.21105/joss.03001}
}
\seealso{
Other ernest_lrps:
\code{\link[=multi_ellipsoid]{multi_ellipsoid()}},
\code{\link[=rwmh_cube]{rwmh_cube()}},
\code{\link[=unif_cube]{unif_cube()}},
\code{\link[=unif_ellipsoid]{unif_ellipsoid()}}
}
\concept{ernest_lrps}
