% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_rectangle.R
\name{slice_rectangle}
\alias{slice_rectangle}
\title{Generate new points with slice sampling}
\usage{
slice_rectangle(enlarge = 1)
}
\arguments{
\item{enlarge}{\verb{[double(1)]}\cr Factor by which to inflate the
hyperrectangle's volume before sampling (see Details). Optional, and must be
greater or equal to 1 if provided; if left \code{NA}, sampling is initially
bounded by the unit hypercube at each iteration.}
}
\value{
\verb{[slice_rectangle]}, a named list that inherits from
[\link{ernest_lrps}].
}
\description{
Create new samples for the live set by evolving a current point in the set
through slice sampling within a bounding hyperrectangle, shrinking the
rectangle when proposals are rejected.
}
\details{
The slice LRPS generates proposals by uniformly sampling within a bounding
hyperrectangle that contains regions of the parameter space satisfying the
likelihood criterion. Sampling begins by selecting a known live point
\eqn{\theta} that satisfies the criterion. Each iteration proposes
a new point within this rectangle via uniform sampling and compares it
against the criterion; if rejected, a new hyperrectangle is drawn such that
the proposed point is on its boundary and \eqn{\theta} is in its interior.
This continues until either a valid proposal is found or the rectangle has
shrunk to the point where no further clamping operations can be performed.

By default, the hyperrectangle spans the extreme values of the current
live set in each dimension. This may risk excluding valid regions
of the parameter space, particularly where the posterior is multimodal or
highly non-Gaussian. To mitigate this, set \code{enlarge > 1}, which inflates the
hyperrectagle's volume by the specified factor before sampling. Setting
\code{enlarge} to \code{NA} disables this behaviour, instead slicing from the unit
hypercube at each iteration.
}
\examples{
# Basic usage with default parameters
lrps <- slice_rectangle()

# More patient sampler for complex posteriors
patient_lrps <- slice_rectangle(enlarge = 1.25)

}
\references{
Neal, R. M. (2000). Slice Sampling (Version 1). arXiv.
\doi{10.48550/ARXIV.PHYSICS/0009028}
}
\seealso{
Other ernest_lrps: 
\code{\link{multi_ellipsoid}()},
\code{\link{rwmh_cube}()},
\code{\link{unif_cube}()},
\code{\link{unif_ellipsoid}()}
}
\concept{ernest_lrps}
