% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-class.R
\name{summary.ernest_run}
\alias{summary.ernest_run}
\title{Summarize a nested sampling run}
\usage{
\method{summary}{ernest_run}(object, ...)
}
\arguments{
\item{object}{[\link{ernest_run}]\cr Results from a nested sampling run.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A named list, containing:
\itemize{
\item \code{nlive}: \verb{[integer(1)]} Number of points in the live set.
\item \code{niter}: \verb{[integer(1)]} Number of iterations performed.
\item \code{neval}: \verb{[integer(1)]} Number of times the likelihood function was
evaluated.
\item \code{ess}: \verb{[numeric(1)]} Effective sample size of the posterior samples,
calculated using Kish's formula for the posterior importance weights
\eqn{w_i}: \deqn{ESS = \frac{(\sum_i w_i)^2}{\sum_i w_i^2}}
\item \code{log_evidence}: \verb{[numeric(1)]} Log-evidence estimate.
\item \code{log_evidence_err}: \verb{[numeric(1)]} Standard error of log-evidence.
\item \code{information}: \verb{[numeric(1)]} Estimated Kullback-Leibler divergence between
the prior and posterior.
\item \code{reweighted_samples}: [\link[posterior:draws_matrix]{posterior::draws_matrix}] Posterior samples,
resampled by normalized weights.
\item \code{mle}: \verb{[list]} Maximum likelihood estimate extracted during the run,
stored in a list with the elements:
\itemize{
\item \code{log_lik}: \verb{[double(1)]} The maximum log-likelihood value.
\item \code{original}, \code{unit_cube}: \verb{[double(nvar)]} The parameter values at the
MLE, expressed in the original parameter space and within the unit cube.
}
\item \code{posterior}: [\link{data.frame}] with columns for the posterior mean, sd,
median, and the 15th and 85th percentiles for each parameter.
\item \code{seed}: The RNG seed used.
}
}
\description{
Returns a concise summary of an \code{ernest_run} object, including key
statistics and a description of the posterior distribution.
}
\examples{
data(example_run)
run_sm <- summary(example_run)
run_sm
run_sm$posterior
}
\seealso{
\code{\link[=generate.ernest_run]{generate.ernest_run()}} \code{\link[=as_draws.ernest_run]{as_draws.ernest_run()}}
\code{\link[=glance.ernest_run]{glance.ernest_run()}}
}
