% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-class.R
\name{summary.ernest_run}
\alias{summary.ernest_run}
\alias{log_evidence}
\title{Summarize a nested sampling run}
\usage{
\method{summary}{ernest_run}(object, ...)

log_evidence(object, ...)
}
\arguments{
\item{object}{[\link{ernest_run}]\cr Results from a nested sampling run.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A named list, containing:
\itemize{
\item \code{nlive}: \verb{[integer(1)]} Number of points in the live set.
\item \code{niter}: \verb{[integer(1)]} Number of iterations.
\item \code{neval}: \verb{[integer(1)]} Number of likelihood evaluations.
\item \code{log_evidence}: \verb{[numeric(1)]} Log-evidence estimate.
\item \code{log_evidence_err}: \verb{[numeric(1)]} Standard error of log-evidence.
\item \code{information}: \verb{[numeric(1)]} Estimated Kullback-Leibler divergence between
the prior and posterior.
\item \code{reweighted_samples}: [\link[posterior:draws_matrix]{posterior::draws_matrix}] Posterior samples,
resampled by normalized weights.
\item \code{mle}: \verb{[list]} Maximum likelihood estimate extracted during the run,
stored in a list with the elements:
\itemize{
\item \code{log_lik}: \verb{[double(1)]} The maximum log-likelihood value.
\item \code{original}, \code{unit_cube}: \verb{[double(nvar)]} The parameter values at the
MLE, expressed in the original parameter space and within the unit cube.
}
\item \code{posterior}: [\link{data.frame}] with columns for the posterior mean, sd,
median, and the 15th and 85th percentiles for each parameter.
\item \code{seed}: The RNG seed used.
}

For \code{log_evidence()}, a single numeric value with the log-evidence estimate.
}
\description{
Returns a concise summary of an \code{ernest_run} object, including key
statistics and a description of the posterior distribution.
}
\examples{
data(example_run)
run_sm <- summary(example_run)
log_evidence(example_run)
run_sm
run_sm$posterior
}
\seealso{
\itemize{
\item \code{\link[=generate]{generate()}} for details on the \code{ernest_run} object.
\item \code{\link[=as_draws]{as_draws()}} for details on how posterior samples are extracted.
}
}
