% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-class.R
\name{summary.ernest_run}
\alias{summary.ernest_run}
\title{Summarize a nested sampling run}
\usage{
\method{summary}{ernest_run}(object, ...)
}
\arguments{
\item{object}{An \code{ernest_run} object.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A list of class \code{summary.ernest_run} with:
\itemize{
\item \code{nlive}: Number of live points.
\item \code{niter}: Number of iterations.
\item \code{neval}: Number of likelihood evaluations.
\item \code{log_evidence}: Log-evidence estimate.
\item \code{log_evidence_err}: Standard error of log-evidence.
\item \code{information}: Estimated Kullback-Leibler divergence between the prior
and posterior.
\item \code{reweighted_samples}: Posterior samples resampled by normalized weights.
\item \code{mle}: Maximum likelihood estimate extracted during the run, stored in a
list with the elements:
\itemize{
\item \code{log_lik}: The maximum log-likelihood value.
\item \code{original}, \code{unit_cube}: The corresponding sample in the original
parameter space and the unit hypercube, respectively.
}
\item \code{posterior}: \code{tibble} with posterior mean, sd, median, 15th and
85th percentiles for each parameter.
\item \code{seed}: The RNG seed used.
}
}
\description{
Returns a concise summary of an \code{ernest_run} object, including key statistics and posterior summaries.
}
\examples{
data(example_run)
run_sm <- summary(example_run)
run_sm
run_sm$posterior
}
\seealso{
\itemize{
\item \code{\link[=generate]{generate()}} for details on the \code{ernest_run} object.
\item \code{\link[=as_draws]{as_draws()}} for details on how posterior samples are extracted.
}
}
