% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ernest_run.R
\name{summary.ernest_run}
\alias{summary.ernest_run}
\alias{print.summary.ernest_run}
\title{Summarising Nested Sampling Runs}
\usage{
\method{summary}{ernest_run}(object, exponentiate = TRUE, normalise = FALSE)

\method{print}{summary.ernest_run}(x, digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{object}{An object of class \code{ernest_run}}

\item{exponentiate}{Whether or not to exponentiate the fdsafnaslk}

\item{normalise}{Whether or not to normalise the likelihood values}
}
\value{
A list with the following components
}
\description{
Summarising Nested Sampling Runs
}
