% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-class.R
\name{summary.ernest_run}
\alias{summary.ernest_run}
\title{Summarise a nested sampling run}
\usage{
\method{summary}{ernest_run}(object, ...)
}
\arguments{
\item{object}{(ernest_run) An object of class \code{ernest_run}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
An object of class \code{summary.ernest_run}, a list with:
\itemize{
\item \code{n_iter}: Number of iterations (number of dead points).
\item \code{n_points}: Number of live points at the end of the run.
\item \code{log_volume}, \code{log_evidence}, \code{log_evidence_err}: The final estimates of
the quantities performed by the run that generated \code{object}.
\item \code{run}, A tibble with \code{n_iter + n_points} rows, containing the vectors
\code{call}, \code{log_lik}, \code{log_volume}, \code{log_weight}, \code{log_evidence},
\code{log_evidence_err}, and \code{information}.
}
}
\description{
Provides a summary of an \code{ernest_run} object.
}
\examples{
# Load an example run
data(example_run)

# Summarise the run and view a tibble of its results.
run_sm <- summary(example_run)
run_sm
run_sm$run
}
\seealso{
\linkS4class{ernest_run} for the \code{ernest_run} object. \code{\link[=as_draws]{as_draws()}} for
how to summarize the posterior distribution generated by nested sampling.
}
