% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-class.R
\name{summary.ernest_run}
\alias{summary.ernest_run}
\title{Summarise a nested sampling run}
\usage{
\method{summary}{ernest_run}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ernest_run}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A list of class \code{summary.ernest_run} with the following components:
\itemize{
\item \code{n_iter}: Integer. Number of iterations performed.
\item \code{n_points}: Integer. Number of live points used in the run.
\item \code{n_calls}: Integer. Total number of likelihood function calls.
\item \code{log_volume}: Double. Final estimated log-prior volume.
\item \code{log_evidence}: Double. Final log-evidence estimate.
\item \code{log_evidence_err}: Double. Standard deviation of the log-evidence
estimate.
\item \code{draws}: Posterior draws as returned by \code{\link[=as_draws]{as_draws()}}.
}

In addition, the \code{run} tibble stores the state of the run at each iteration
within these columns:
\itemize{
\item \code{call}: Cumulative number of likelihood calls.
\item \code{log_lik}: Log-likelihood for each sample.
\item \code{log_volume}: Estimated log-prior volume.
\item \code{log_weight}: Unnormalized log-weights (relative to evidence).
\item \code{log_evidence}: Cumulative log-evidence.
\item \code{log_evidence_err}: Standard deviation of log-evidence.
\item \code{information}: Estimated KL divergence at each iteration.
}
}
\description{
Provides a summary of an \code{ernest_run} object, including key statistics and a
tibble of results for each iteration.
}
\examples{
# Load an example run
data(example_run)

# Summarise the run and view a tibble of its results.
run_sm <- summary(example_run)
run_sm
run_sm$run
}
\seealso{
\itemize{
\item \code{\link[=generate]{generate()}} describes the \code{ernest_run} object.
\item \code{\link[=as_draws]{as_draws()}} provides more information on \code{draws} objects.
}
}
