% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ernest_run.R
\name{summary.ernest_run}
\alias{summary.ernest_run}
\title{Summarise a nested sampling run}
\usage{
\method{summary}{ernest_run}(object, ...)
}
\arguments{
\item{object}{An \code{ernest_run} object.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
An object of class \code{summary.ernest_run}, a list with:
\itemize{
\item \code{n_iter}: Number of iterations (number of dead points).
\item \code{n_points}: Number of live points at the end of the run.
\item \code{log_volume}: Final log-volume.
\item \code{log_evidence}: Final log-evidence estimate.
\item \code{log_evidence_err}: Final log-evidence error (standard deviation).
\item \code{run}, A tibble with columns:
\item 
\itemize{
\item \code{call}: The cumulative number of likelihood calls at each iteration.
}
\item 
\itemize{
\item \code{log_weight}: Log-posterior weights for each sample, now normalised
by the final log-evidence estimate.
}
\item 
\itemize{
\item \code{log_lik}, \code{log_volume}, \code{log_evidence}, \code{log_evidence_err} and
\code{information}: Run results as in \link{ernest_run}.
}
}
}
\description{
Provides a summary of an \code{ernest_run} object, including cumulative likelihood calls,
log-likelihood, log-volume, log-weights (relative to the final evidence), cumulative
log-evidence, evidence error, and information at each iteration.
}
