% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ernest_sampler}
\alias{summary.ernest_sampler}
\title{Summarise an \code{ernest_sampler} object}
\usage{
\method{summary}{ernest_sampler}(object, ...)
}
\arguments{
\item{object}{An \code{ernest_sampler} object.}

\item{...}{Must be empty.}
}
\value{
A list with class \code{summary.ernest_sampler} containing these named elements:
\itemize{
\item n_points: The number of live points used in the run; a scalar integer.
\item n_iter: The number of iterations; a scalar integer.
\item n_call: The total number of function calls; a scalar integer.
\item eff: The overall sampling efficiency; a scalar double.
\item log_weight: The posterior log-weight associated with each point; a vector of doubles.
\item log_lik: The log-likelihood associated with each point; a vector of doubles.
\item log_vol: The estimated log-volume associated with each point; a vector of doubles.
\item log_z: The cumulative evidence estimate at each iteration; a vector of doubles.
\item log_z_err: The estimated error (standard deviation) of \code{log_z}; a vector of doubles.
\item information: The estimated information at each iteration; a vector of doubles.
}
}
\description{
\code{\link[=summary]{summary()}} method for \code{ernest_sampler} objects, producing an \code{S3} object
describing a unit's run.
}
