% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif_cube.R
\name{unif_cube}
\alias{unif_cube}
\title{Generate new points from the unconstrained prior distribution}
\usage{
unif_cube()
}
\value{
\verb{[unif_cube]}, a named list that inherits from [\link{ernest_lrps}].
}
\description{
Use rejection sampling across the entire prior distribution to create new
samples. This is highly inefficient as an LRPS, but may be useful for
testing the behaviour of a nested sampling specification.
}
\section{Loop safety}{

An LRPS may fail to produce a point satisfying the likelihood constraint
(for example, due to rounding errors or poor parameterisation). To avoid
infinite loops during a nested sampling run, \code{propose()} will terminate a
run if it cannot find a likelihood-restricted point after a maximum number of
calls to the likelihood function. Adjust this limit with the
\code{ernest.max_loop} option (default: 1e6).
}

\examples{
data(example_run)
lrps <- unif_cube()

ernest_sampler(example_run$log_lik_fn, example_run$prior, sampler = lrps)
}
\references{
Speagle, J. S. (2020). Dynesty: A Dynamic Nested Sampling Package
for Estimating Bayesian Posteriors and Evidences. Monthly Notices of the
Royal Astronomical Society, 493, 3132–3158.
\doi{10.1093/mnras/staa278}
}
\seealso{
Other ernest_lrps:
\code{\link[=multi_ellipsoid]{multi_ellipsoid()}},
\code{\link[=rwmh_cube]{rwmh_cube()}},
\code{\link[=slice_rectangle]{slice_rectangle()}},
\code{\link[=unif_ellipsoid]{unif_ellipsoid()}}
}
\concept{ernest_lrps}
