% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif_cube.R
\name{unif_cube}
\alias{unif_cube}
\title{Likelihood-Prior Restricted Sampling within the Unit Cube}
\usage{
unif_cube(max_attempts = 1e+06)
}
\arguments{
\item{max_attempts}{The maximum number of attempts to find a valid point during
region-based sampling, including during uniform sampling.}
}
\value{
An object of class \code{unif_cube}, which inherits from \code{ernest_sampler}
}
\description{
Generate particles by naively sampling from a unit hypercube. This is a
simple but inefficient method of sampling from the prior distribution.
}
\details{
This sampler can be useful in a few edge cases where the number of live
points is small compared to the number of dimensions, or if the user wants to
run tests to check the sampling behaviour.
}
