% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif_cube.R
\name{unif_cube}
\alias{unif_cube}
\title{Uniformly sample from the unconstrained prior distribution}
\usage{
unif_cube()
}
\value{
A list, with class \code{c("unif_cube", "ernest_lrps")}. Can be used with
\link{nested_sampling} to specify the sampling behaviour of a nested sampling run.
}
\description{
Generate new live points by performing rejection sampling across the entire
prior distribution. This is horribly inefficient as an LRPS, but you might
find this helpful for testing the behaviour of a nested sampling
specification.
}
\references{
Speagle, J. S. (2020). Dynesty: A Dynamic Nested Sampling Package
for Estimating Bayesian Posteriors and Evidences. Monthly Notices of the
Royal Astronomical Society, 493, 3132–3158.
\url{https://doi.org/10.1093/mnras/staa278}
}
\concept{ernest_lrps}
