% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif_ellipsoid.R
\name{unif_ellipsoid}
\alias{unif_ellipsoid}
\title{Generate new points from the spanning ellipsoid}
\usage{
unif_ellipsoid(enlarge = 1)
}
\arguments{
\item{enlarge}{\verb{[double(1)]}\cr Factor by which to inflate the bounding
ellipsoid's volume before sampling (see Details). Must be at least 1.}
}
\value{
\verb{[unif_ellipsoid]}, a named list that inherits from [\link{ernest_lrps}].
}
\description{
Uses the bounding ellipsoid of the live set to define the region of prior
space that contains new points. Effective for unimodal and roughly-Gaussian
posteriors.
}
\details{
Nested likelihood contours rarely form perfect ellipses, so sampling
from the spanning ellipsoid without enlargement may exclude valid regions.
This can bias proposals towards the ellipsoid centre and overestimate
evidence. When working with multimodal likelihood surfaces, consider setting
\code{enlarge > 1} or using \code{\link[=multi_ellipsoid]{multi_ellipsoid()}}.

The covariance matrix of the points is used to estimate the ellipsoid's
shape. In exceptional cases (e.g., perfect collinearity), this matrix may be
singular. Should this occur, the covariance matrix is reconditioned by
adjusting its eigenvalues. Should this also fail, the algorithm falls back
to sampling from the circumscribed sphere bounding the unit hypercube.
}
\section{Ellipsoids}{

Ellipsoids are stored in the \code{cache} environment of the LRPS object.
Ellipsoids are defined by their centre \eqn{c} and shape matrix \eqn{A}.
The set of points \eqn{x} contained within the ellipsoid is given by
\deqn{ x \in {\bf{R}}^n | (x-c) A (x-c)' \leq 1 }

The volume of the ellipsoid is
\eqn{V = \mathrm{Vol}(S_n) \sqrt{\det(A^{-1})}}, where
\eqn{\mathrm{Vol}(S_n)} is the volume of the unit hypersphere.

For sampling, we store the matrix \eqn{A^{-1/2}}, the inverse of the
positive-semidefinite square root of \eqn{A}. The ellipsoid can equivalently
be defined as the set of points \deqn{x = A^{-1/2} y + c,}
where \eqn{y} are points from the unit hypersphere.

For more on ellipsoids and their operations, see
\href{http://tcg.mae.cornell.edu/pubs/Pope_FDA_08.pdf}{Algorithms for Ellipsoids}
by S.B. Pope, Cornell University Report FDA 08-01 (2008).
}

\examples{
data(example_run)
lrps <- unif_ellipsoid()

ernest_sampler(example_run$log_lik_fn, example_run$prior, sampler = lrps)
}
\references{
Feroz, F., Hobson, M. P., Bridges, M. (2009) MULTINEST: An Efficient and
Robust Bayesian Inference Tool for Cosmology and Particle Physics. Monthly
Notices of the Royal Astronomical Society. 398(4), 1601–1614.
\doi{10.1111/j.1365-2966.2009.14548.x}

Mukherjee, P., Parkinson, D., & Liddle, A. R. (2006). A Nested Sampling
Algorithm for Cosmological Model Selection. The Astrophysical Journal,
638(2), L51. \doi{10.1086/501068}
}
\seealso{
Other ernest_lrps:
\code{\link[=multi_ellipsoid]{multi_ellipsoid()}},
\code{\link[=rwmh_cube]{rwmh_cube()}},
\code{\link[=slice_rectangle]{slice_rectangle()}},
\code{\link[=unif_cube]{unif_cube()}}
}
\concept{ernest_lrps}
