% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif_ellipsoid.R
\name{unif_ellipsoid}
\alias{unif_ellipsoid}
\title{Uniformly sample from the ellipsoid hull of live points}
\usage{
unif_ellipsoid(scale = 1)
}
\arguments{
\item{scale}{A double, used to inflate the volume of the ellipsoid before
sampling from it.}
}
\value{
A list with class \code{c("unif_ellipsoid", "ernest_lrps")}. Can be used
with \code{\link[=ernest_sampler]{ernest_sampler()}} to specify the sampling behaviour of a nested
sampling run.
}
\description{
Generate new live points from the ellipsoid hull or spanning ellipsoid of
the current set of live points. Individual points are generated through
rejection sampling within the ellipsoid, while \code{\link[=update_lrps]{update_lrps()}} updates the
bounding ellipsoid using \link[cluster:ellipsoidhull]{cluster::ellipsoidhull}.
}
