% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif_ellipsoid.R
\name{unif_ellipsoid}
\alias{unif_ellipsoid}
\title{Generate samples from the spanning ellipsoid}
\usage{
unif_ellipsoid(enlarge = 1.25)
}
\arguments{
\item{enlarge}{Double, larger or equal to 1. Factor to inflate the bounding
ellipsoid's volume before sampling (see Details).}
}
\value{
A list with class \code{c("unif_ellipsoid", "ernest_lrps")}. Use with
\code{\link[=ernest_sampler]{ernest_sampler()}} to specify nested sampling behaviour.
}
\description{
Uses the bounding ellipsoid of live points to define the prior region for new
samples. Effective for unimodal, roughly Gaussian posteriors.
}
\details{
Nested likelihood contours rarely form perfect ellipses, so sampling
from the spanning ellipsoid without enlargement may exclude valid regions.
This can bias proposals toward the ellipsoid centre and overestimate
evidence. Setting \code{enlarge = 1} will produce a warning.

Ellipsoids are stored in the LRPS cache in quadric representation. That is,
an ellipsoid with center \eqn{c} and shape matrix \eqn{Q} contains points
\eqn{x \in R^n} that satisfy the inequality
\deqn{(x - c) Q^-1 (x - c)^T \leq 1}
The square root of the precision matrix \eqn{Q}, scaled by \code{enlarge}, is
also stored for efficient point generation.

The covariance matrix of the points is used estimate \eqn{Q}. In exceptional
cases (e.g. perfect collinearity), this matrix may be singular. Should this
occur, the covariance matrix is reconditioned by adjusting its eigenvalues.
Should this also fail, the algorithm falls back to sampling from the
circumscribed sphere bounding the unit hypercube.
}
\examples{
data(example_run)
lrps <- unif_ellipsoid(enlarge = 1.25)

ernest_sampler(example_run$log_lik_fn, example_run$prior, sampler = lrps)
}
\references{
Feroz, F., Hobson, M. P., Bridges, M. (2009) MULTINEST: An Efficient and
Robust Bayesian Inference Tool for Cosmology and Particle Physics. Monthly
Notices of the Royal Astronomical Society. 398(4), 1601–1614.
\url{https://doi.org/10.1111/j.1365-2966.2009.14548.x}

Mukherjee, P., Parkinson, D., & Liddle, A. R. (2006). A Nested Sampling
Algorithm for Cosmological Model Selection. The Astrophysical Journal,
638(2), L51. \url{https://doi.org/10.1086/501068}
}
\seealso{
Other ernest_lrps: 
\code{\link{adaptive_rwmh}()},
\code{\link{rwmh_cube}()},
\code{\link{unif_cube}()}
}
\concept{ernest_lrps}
