% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrps-class.R, R/multi_ellipsoid.R,
%   R/rwmh_cube.R, R/slice_rectangle.R, R/unif_ellipsoid.R
\name{update_lrps}
\alias{update_lrps}
\alias{update_lrps.multi_ellipsoid}
\alias{update_lrps.rwmh_cube}
\alias{update_lrps.slice_rectangle}
\alias{update_lrps.unif_ellipsoid}
\title{Update an LRPS}
\usage{
update_lrps(x, ...)

\method{update_lrps}{multi_ellipsoid}(x, unit = NULL, log_volume = NA, ...)

\method{update_lrps}{rwmh_cube}(x, unit = NULL, ...)

\method{update_lrps}{slice_rectangle}(x, unit = NULL, ...)

\method{update_lrps}{unif_ellipsoid}(x, unit = NULL, log_volume = NA, ...)
}
\arguments{
\item{x}{An \code{ernest_lrps} object.}

\item{unit}{A matrix of live points within the sampler. If NULL, no LRPS
updates based on the state of the live points will be made.}

\item{log_volume}{The current log-volume of the nested sampling run.}
}
\value{
An updated \code{ernest_lrps} object with the same class as \code{x}, possibly
with updated parameters.
}
\description{
During a nested sampling run, you may wish to update the internal parameters
of the LRPS based on sampler performance or other criterion. The frequency of
these updates is set by the \code{first_update} and \code{update_interval} arguments of
\code{\link[=ernest_sampler]{ernest_sampler()}}.

If you are creating your own \link{ernest_lrps} subclass, implement this method to
specify any special update behaviour. The default method reconstructs the
LRPS with current parameters and resets the likelihood call counter in the
cache.
}
\keyword{internal}
