% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{variables<-.ernest_prior}
\alias{variables<-.ernest_prior}
\alias{variables<-.ernest_sampler}
\title{Set Variable Names in an \code{ernest_prior} or \code{ernest_sampler} Object}
\usage{
\method{variables}{ernest_prior}(x, ...) <- value

\method{variables}{ernest_sampler}(x, ...) <- value
}
\arguments{
\item{x}{An \code{ernest_prior} or \code{ernest_sampler} object.}

\item{...}{Must be empty.}

\item{value}{A character vector of unique variable names.}
}
\value{
The modified object with updated variable names.
}
\description{
Assign variable names to \code{ernest} objects.
}
\seealso{
\code{\link[posterior:variables-set]{posterior::variables<-()}}
}
