% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.ernest_run}
\alias{visualize.ernest_run}
\title{Visualize posterior distributions or traces from a nested sampling run}
\usage{
\method{visualize}{ernest_run}(
  x,
  ...,
  .which = c("density", "trace"),
  .units = c("original", "unit_cube")
)
}
\arguments{
\item{x}{[\link{ernest_run}] The results of a nested sampling run.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more variables
to plot from the run. If omitted, all variables are plotted.}

\item{.which}{\verb{[charcacter(1)]} Character string specifying the type of plot
to produce. Options are \code{"density"} for the posterior density of each
parameter or \code{"trace"} for the trace of variables along log-volume.}

\item{.units}{\verb{[character(1)]}\cr The scale of the sampled points:
\itemize{
\item \code{"original"}: Points are on the scale of the prior space.
\item \code{"unit_cube"}: Points are on the (0, 1) unit hypercube scale.
}}
}
\value{
A \code{ggplot2::ggplot()} object.
}
\description{
Produces visualizations of the posterior distributions or the evolution of
variables along the log-prior volume from a nested sampling run.
}
\details{
The \code{visualize()} function is designed to quickly explore the results of a
nested sampling run through two types of plots:
\itemize{
\item \strong{Density plots} show the marginal posterior for each selected variable,
using \code{ggdist::stat_halfeye()} to visualize uncertainty and distribution
shape.
\item \strong{Trace plots} display the evolution of variables as a function of
log-volume, with points coloured by posterior weight. This can help
diagnose sampling behavior and identify regions of interest in the prior
volume.
}

Posterior weights are derived from the individual contributions of each
sampled point in the prior space to a run's log-evidence estimate. A point's
weight is a function of (a) the point's likelihood and (b) the estimated
amount of volume within that point's likelihood contour. See ernest's
vignettes for more information.
}
\note{
This function requires \CRANpkg{tidyselect}. If \code{which = "trace"} is
selected, \CRANpkg{ggdist} is also required.
}
\examples{
# Load example run
library(ggdist)
data(example_run)

# Plot posterior densities for all parameters
visualize(example_run, .which = "density")
}
\seealso{
\code{\link[=as_draws_rvars]{as_draws_rvars()}}

Other visualizations:
\code{\link[=plot.ernest_estimate]{plot.ernest_estimate()}}
}
\concept{visualizations}
