% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.ernest_run}
\alias{visualize.ernest_run}
\title{Plot the posterior distribution of an \code{ernest_run}}
\usage{
\method{visualize}{ernest_run}(x, ..., type = c("density", "trace"), vars = NULL, plot = TRUE)
}
\arguments{
\item{x}{An \link{ernest_run} object.}

\item{...}{
  Arguments passed on to \code{\link[=as_draws.ernest_run]{as_draws.ernest_run}}
  \describe{
    \item{\code{units}}{Case-sensitive string. The scale for the sampled points:
\itemize{
\item \code{"original"}: Points are on the scale of the prior space.
\item \code{"unit_cube"}: Points are on the (0, 1) unit hypercube scale.
}}
    \item{\code{radial}}{Logical. If \code{TRUE}, returns an additional column \code{.radial}
containing the radial coordinate (i.e., the Euclidean norm) for each
sampled point.}
  }}

\item{type}{Case-sensitive string. The type of plot to create:
\itemize{
\item \code{"density"}: Shows the posterior density of each parameter.
\item \code{"trace"}: Shows the distribution of points along estimates of the log
prior volume.
}}

\item{vars}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Variables to
plot from the run. If \code{NULL}, all variables are plotted.}

\item{plot}{Logical. If \code{TRUE}, returns a \code{ggplot} of the visualisation;
if \code{FALSE}, returns a \code{tibble} of the data used to create the plot.}
}
\value{
A \code{ggplot} object if \code{plot = TRUE}, otherwise a \code{tibble}.
}
\description{
Create a plot of the posterior distributions from a nested sampling run,
or trace the evolution of discarded live points along the log prior
volume.
}
\examples{
# Load example run
library(ggdist)
data(example_run)

# Plot posterior distributions of the parameters
visualize(example_run, type = "density")

# Plot the trace of the radial coordinate in unit scale
visualize(
  example_run,
  type = "trace",
  vars = ".radial",
  units = "unit_cube",
  radial = TRUE
)
}
\seealso{
\code{\link[=plot]{plot()}} for visualising evidence estimates from an
\code{ernest_run}.
}
