% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.ernest_run}
\alias{visualize.ernest_run}
\title{Plot the posterior distribution of an \code{ernest_run}}
\usage{
\method{visualize}{ernest_run}(x, ..., type = c("density", "trace"), vars = NULL, plot = TRUE)
}
\arguments{
\item{x}{An \link{ernest_run} object.}

\item{...}{
  Arguments passed on to \code{\link[=as_draws.ernest_run]{as_draws.ernest_run}}
  \describe{
    \item{\code{units}}{Case-sensitive string. The scale in which to return the
sampled points:
\itemize{
\item \code{"original"}: Points are expressed on the scale of the prior space.
\item \code{"unit_cube"}: Points are expressed on the scale of the (0-1)-unit
hypercube.
}}
    \item{\code{radial}}{Logical. Whether to return an additional column \code{.radial},
containing the radial coordinate (i.e., the squared sum of squares) for
each sampled point.}
  }}

\item{type}{A case-sensitive string. The type of plot to create.
\itemize{
\item \code{"density"}: Shows the posterior density of each distribution.
\item \code{"trace"}: Shows the distribution of points along estimates of the log
prior volume.
}}

\item{vars}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> The variables to
plot from the run. If left \code{NULL}, every variable is plotted.}

\item{plot}{Logical, whether to return a \code{ggplot} of the visualization, or
a \code{tibble} of the data used to create the plot.}
}
\value{
Either a \code{ggplot} object if \code{plot = TRUE}, or a \code{tibble}.
}
\description{
Create a plot of the posterior distributions from a nested sampling run,
or trace the evolution of the discarded live points along the log prior
volume.
}
\note{
This method requires the \code{ggdist} and \code{tidyselect} optional packages.
}
\examples{
# Load example run
library(ggdist)
data(example_run)

# Plot posterior distributions of the parameters
visualize(example_run, type = "density")

# Plot the trace of the radial coordinate in unit scale
visualize(
  example_run,
  type = "trace",
  vars = ".radial",
  units = "unit_cube",
  radial = TRUE
)
}
\seealso{
\code{\link[=plot]{plot()}} for visualizing the evidence estimates from an
\code{ernest_run}.
}
